### 2021-11-25 #3953 @Kevin-Mattheus-Moerman
@quantumsteve please remove the invalid (all zero) orcid id's from the paper. Instead leave out the orcid specification for these if they do not have an orcid id, see also below and:
https://joss.readthedocs.io/en/latest/submitting.html#example-paper-and-bibliography

```yaml
authors:
  - name: Adrian M. Price-Whelan^[co-first author] # note this makes a footnote saying 'co-first author'
    orcid: 0000-0003-0872-7098
    affiliation: "1, 2" # (Multiple affiliations must be quoted)
  - name: Author Without ORCID^[co-first author] # note this makes a footnote saying 'co-first author'
    affiliation: 2
```

After fixing the paper you can update the pdf draft here by calling `@whedon generate pdf from branch paper`
