### 2021-11-29 d0e8db7
Speed up unit tests (#408)
* setting OMP_NUM_THREADS=1 to reduce oversubscription

* Fix issues building with clang 13.0

Signed-off-by: Steven Hahn <hahnse@ornl.gov>

* Use openmp on tests without -mpi

Signed-off-by: Steven Hahn <hahnse@ornl.gov>
