+++ src/cblacs_grid.hpp
@@ -1,9 +1,11 @@
 #pragma once
 
+#include "mpi.h"
+
 class cblacs_grid
 {
 public:
-  cblacs_grid();
+  cblacs_grid(MPI_Comm localCommunicator);
   int get_context() const { return ictxt_; }
   int get_myrow() const { return myrow_; }
   int get_mycol() const { return mycol_; }
@@ -12,5 +14,5 @@ public:
   ~cblacs_grid();
 
 private:
-  int ictxt_, nprow_{1}, npcol_, myrow_, mycol_;
+  int bhandle_, ictxt_, nprow_{1}, npcol_, myrow_, mycol_;
 };
