+++ src/distribution.cpp
@@ -2,6 +2,8 @@
 #include "lib_dispatch.hpp"
 #include "tools.hpp"
 
+#include "cblacs_grid.hpp"
+
 #include <cmath>
 #include <csignal>
 #include <list>
@@ -1118,6 +1120,12 @@ void bcast(int *value, int size, int rank)
   (void)size;
   (void)rank;
 #endif
+}
+
+std::shared_ptr<cblacs_grid> get_grid()
+{
+  auto grid = std::make_shared<cblacs_grid>(distro_handle.get_global_comm());
+  return grid;
 }
 
 template void reduce_results(fk::vector<float> const &source,
