+++ src/time_advance.cpp
@@ -256,7 +256,8 @@ implicit_advance(PDE<P> const &pde,
 {
   expect(time >= 0);
 #ifdef ASGARD_USE_SCALAPACK
-  std::shared_ptr<cblacs_grid> sgrid = std::make_shared<cblacs_grid>();
+  std::shared_ptr<cblacs_grid> sgrid =
+      get_grid(); // std::make_shared<cblacs_grid>();
 #endif
   static fk::matrix<P, mem_type::owner, resource::host> A;
   static std::vector<int> ipiv;
