+++ src/distribution.cpp
@@ -2,8 +2,6 @@
 #include "lib_dispatch.hpp"
 #include "tools.hpp"
 
-#include "cblacs_grid.hpp"
-
 #include <cmath>
 #include <csignal>
 #include <list>
@@ -1122,11 +1120,13 @@ void bcast(int *value, int size, int rank)
 #endif
 }
 
+#ifdef ASGARD_USE_SCALAPACK
 std::shared_ptr<cblacs_grid> get_grid()
 {
   auto grid = std::make_shared<cblacs_grid>(distro_handle.get_global_comm());
   return grid;
 }
+#endif
 
 template void reduce_results(fk::vector<float> const &source,
                              fk::vector<float> &dest,
