+++ src/distribution.hpp
@@ -309,4 +309,6 @@ fk::vector<P> row_to_col_major(fk::vector<P> const &x, int size_r);
 
 void bcast(int *value, int size, int rank);
 
+#ifdef ASGARD_USE_SCALAPACK
 std::shared_ptr<cblacs_grid> get_grid();
+#endif
