### 2021-12-14 bfc092e
Remove commented out code
Signed-off-by: Steven Hahn <hahnse@ornl.gov>
