+++ src/boundary_conditions_tests.cpp
@@ -174,8 +174,7 @@ TEMPLATE_TEST_CASE("problem separability", "[boundary_condition]", double,
             unscaled_parts_0[0], unscaled_parts_0[1], *pde, start_element,
             stop_element, test_time);
 
-    TestType const tol_factor =
-        std::is_same<TestType, double>::value ? 1e-17 : 1e-8;
+    auto constexpr tol_factor = get_tolerance<TestType>(10);
     rmse_comparison(bc_advanced_0, bc_advanced_1, tol_factor);
   }
 
@@ -228,8 +227,8 @@ TEMPLATE_TEST_CASE("problem separability", "[boundary_condition]", double,
       fk::vector<TestType, mem_type::const_view> const bc_section(
           bc_init, index, index + bc_advanced.size() - 1);
 
-      TestType const tol_factor =
-          std::is_same<TestType, double>::value ? 1e-17 : 1e-4;
+      auto constexpr tol_factor = get_tolerance<TestType>(10);
+
       rmse_comparison(bc_section, bc_advanced, tol_factor);
 
       index += bc_advanced.size();
@@ -240,8 +239,7 @@ TEMPLATE_TEST_CASE("problem separability", "[boundary_condition]", double,
 TEMPLATE_TEST_CASE("compute_boundary_conditions", "[boundary_condition]",
                    double, float)
 {
-  TestType const tol_factor =
-      std::is_same<TestType, double>::value ? 1e-15 : 1e-6;
+  auto constexpr tol_factor = get_tolerance<TestType>(10);
 
   std::string const gold_filename_prefix =
       "../testing/generated-inputs/"
@@ -277,8 +275,7 @@ TEMPLATE_TEST_CASE("compute_boundary_conditions", "[boundary_condition]",
 TEMPLATE_TEST_CASE("boundary_conditions_vector", "[boundary_condition]", double,
                    float)
 {
-  TestType const tol_factor =
-      std::is_same<TestType, double>::value ? 1e-13 : 1e-4;
+  auto constexpr tol_factor = get_tolerance<TestType>(1000);
 
   SECTION("diffusion_1 level 2 degree 2")
   {
