+++ src/time_advance_tests.cpp
@@ -24,7 +24,7 @@ static auto constexpr num_steps = 5;
 
 template<typename P>
 void time_advance_test(parser const &parse, std::string const &filepath,
-                       P const tolerance_factor = 1e-6)
+                       P const tolerance_factor)
 {
   auto const num_ranks = get_num_ranks();
   if (num_ranks > 1 && parse.using_implicit() &&
@@ -109,8 +109,7 @@ TEMPLATE_TEST_CASE("time advance - diffusion 2", "[time_advance]", double,
     int const degree = 2;
     int const level  = 2;
 
-    TestType const tol_factor =
-        std::is_same<TestType, double>::value ? 1e-15 : 1e-5;
+    auto constexpr tol_factor = get_tolerance<TestType>(100);
 
     std::string const gold_base = "../testing/generated-inputs/time_advance/"
                                   "diffusion2_sg_l2_d2_t";
@@ -127,8 +126,7 @@ TEMPLATE_TEST_CASE("time advance - diffusion 2", "[time_advance]", double,
   {
     int const degree = 3;
     int const level  = 3;
-    TestType const tol_factor =
-        std::is_same<TestType, double>::value ? 1e-15 : 1e-5;
+    auto constexpr tol_factor = get_tolerance<TestType>(100);
 
     std::string const gold_base = "../testing/generated-inputs/time_advance/"
                                   "diffusion2_sg_l3_d3_t";
@@ -145,8 +143,7 @@ TEMPLATE_TEST_CASE("time advance - diffusion 2", "[time_advance]", double,
   {
     int const degree = 4;
     int const level  = 4;
-    TestType const tol_factor =
-        std::is_same<TestType, double>::value ? 1e-12 : 1e-1;
+    auto constexpr tol_factor = get_tolerance<TestType>(1000000);
     std::string const gold_base = "../testing/generated-inputs/time_advance/"
                                   "diffusion2_sg_l4_d4_t";
 
@@ -161,8 +158,7 @@ TEMPLATE_TEST_CASE("time advance - diffusion 2", "[time_advance]", double,
   SECTION("diffusion2, explicit/non-uniform level, sparse grid, degree 2")
   {
     int const degree = 2;
-    TestType const tol_factor =
-        std::is_same<TestType, double>::value ? 1e-15 : 1e-5;
+    auto constexpr tol_factor = get_tolerance<TestType>(100);
 
     fk::vector<int> const levels{4, 5};
     std::string const gold_base = "../testing/generated-inputs/time_advance/"
@@ -314,8 +310,7 @@ TEMPLATE_TEST_CASE("time advance - diffusion 1", "[time_advance]", double,
     int const level             = 3;
     std::string const gold_base = "../testing/generated-inputs/time_advance/"
                                   "diffusion1_sg_l3_d3_t";
-    TestType const tol_factor =
-        std::is_same<TestType, double>::value ? 1e-15 : 1e-5;
+    auto constexpr tol_factor = get_tolerance<TestType>(100);
 
     auto const full_grid = false;
     parser const parse(
@@ -331,8 +326,7 @@ TEMPLATE_TEST_CASE("time advance - diffusion 1", "[time_advance]", double,
     int const level             = 4;
     std::string const gold_base = "../testing/generated-inputs/time_advance/"
                                   "diffusion1_sg_l4_d4_t";
-    TestType const tol_factor =
-        std::is_same<TestType, double>::value ? 1e-11 : 1e-2;
[…]
