+++ src/pde/pde_fokkerplanck1_4p4.hpp
@@ -102,6 +102,12 @@ private:
     return 1.0;
   }
 
+  static fk::vector<P> moment_dV(fk::vector<P> const x, P const t = 0)
+  {
+    ignore(t);
+    return fk::vector<P>(std::vector<P>(x.size(), 1.0));
+  }
+
   // specify source functions...
 
   // N/A
@@ -128,7 +134,7 @@ private:
   static P g_func_t1_z(P const x, P const time)
   {
     ignore(time);
-    return -E * (1 - std::pow(x, 2));
+    return -E * sqrt(1 - std::pow(x, 2));
   }
   static P g_func_t2_z1(P const x, P const time)
   {
@@ -142,6 +148,12 @@ private:
     return 1.0;
   }
 
+  static P dV_z(P const x, P const time)
+  {
+    ignore(time);
+    return sqrt(1.0 - std::pow(x, 2));
+  }
+
   // define dimensions
   inline static dimension<P> const dim0_ =
       dimension<P>(-1.0,   // domain min
@@ -149,7 +161,8 @@ private:
                    2,      // levels
                    2,      // degree
                    f0_vec, // initial condition
-                   "x");   // name
+                   moment_dV,
+                   "x"); // name
 
   inline static std::vector<dimension<P>> const dimensions_ = {dim0_};
 
@@ -160,8 +173,11 @@ private:
   // -E d/dz((1-z^2) f)
 
   inline static partial_term<P> const partial_term_0 = partial_term<P>(
-      coefficient_type::grad, g_func_t1_z, flux_type::downwind,
-      boundary_condition::dirichlet, boundary_condition::dirichlet);
+      coefficient_type::div, g_func_t1_z, partial_term<P>::null_gfunc,
+      flux_type::downwind, boundary_condition::dirichlet,
+      boundary_condition::dirichlet, homogeneity::homogeneous,
+      homogeneity::homogeneous, {}, partial_term<P>::null_scalar_func, {},
+      partial_term<P>::null_scalar_func, dV_z);
 
   inline static term<P> const termE_z =
       term<P>(false,           // time-dependent
@@ -175,12 +191,20 @@ private:
   //
   // +C * d/dz( (1-z^2) df/dz )
   inline static partial_term<P> const partial_term_1 = partial_term<P>(
-      coefficient_type::grad, g_func_t2_z1, flux_type::downwind,
-      boundary_condition::dirichlet, boundary_condition::dirichlet);
+      coefficient_type::div, partial_term<P>::null_gfunc,
+      partial_term<P>::null_gfunc, flux_type::downwind,
+      boundary_condition::dirichlet, boundary_condition::dirichlet,
+      homogeneity::homogeneous, homogeneity::homogeneous, {},
+      partial_term<P>::null_scalar_func, {}, partial_term<P>::null_scalar_func,
+      dV_z);
 
   inline static partial_term<P> const partial_term_2 =
-      partial_term<P>(coefficient_type::grad, g_func_t2_z2, flux_type::upwind,
-                      boundary_condition::neumann, boundary_condition::neumann);
+      partial_term<P>(coefficient_type::grad, partial_term<P>::null_gfunc,
+                      partial_term<P>::null_gfunc, flux_type::upwind,
+                      boundary_condition::neumann, boundary_condition::neumann,
+                      homogeneity::homogeneous, homogeneity::homogeneous, {},
+                      partial_term<P>::null_scalar_func, {},
+                      partial_term<P>::null_scalar_func, dV_z);
 
   inline static term<P> const termC_z =
       term<P>(false,           // time-dependent
