+++ src/pde/pde_fokkerplanck1_pitch_E.hpp
@@ -100,6 +100,12 @@ private:
     return 1.0;
   }
 
+  static fk::vector<P> moment_dV(fk::vector<P> const x, P const t = 0)
+  {
+    ignore(t);
+    return fk::vector<P>(std::vector<P>(x.size(), 1.0));
+  }
+
   // specify source functions...
 
   // N/A
@@ -128,7 +134,7 @@ private:
   {
     // suppress compiler warnings
     ignore(time);
-    return -1 * (1 - std::pow(x, 2));
+    return -sqrt(1 - std::pow(x, 2));
   }
   static P g_func_2(P const x, P const time)
   {
@@ -138,6 +144,12 @@ private:
     return 1.0;
   }
 
+  static P dV_z(P const x, P const time)
+  {
+    ignore(time);
+    return sqrt(1.0 - std::pow(x, 2));
+  }
+
   // define dimensions
   inline static dimension<P> const dim0_ =
       dimension<P>(-1.0,                   // domain min
@@ -145,7 +157,8 @@ private:
                    2,                      // levels
                    2,                      // degree
                    initial_condition_dim0, // initial condition
-                   "x");                   // name
+                   moment_dV,
+                   "x"); // name
 
   inline static std::vector<dimension<P>> const dimensions_ = {dim0_};
 
@@ -157,9 +170,12 @@ private:
   // term2_z.G = @(z,p,t,dat) -1.*(1-z.^2); % G function for use in coeff_matrix
   // construction. term2_z.LF = -1; % Upwind term2_z.name = 'd_dz';
 
-  inline static partial_term<P> const partial_term_0 =
-      partial_term<P>(coefficient_type::grad, g_func_1, flux_type::downwind,
-                      boundary_condition::neumann, boundary_condition::neumann);
+  inline static partial_term<P> const partial_term_0 = partial_term<P>(
+      coefficient_type::div, g_func_1, partial_term<P>::null_gfunc,
+      flux_type::downwind, boundary_condition::neumann,
+      boundary_condition::neumann, homogeneity::homogeneous,
+      homogeneity::homogeneous, {}, partial_term<P>::null_scalar_func, {},
+      partial_term<P>::null_scalar_func, dV_z);
 
   inline static term<P> const term0_dim0_ =
       term<P>(false,           // time-dependent

[…]
