### 2022-01-07 0de8571
Modify pterm construction for new operator
