### 2022-01-07 ce63782
clang-format
Signed-off-by: Steven Hahn <hahnse@ornl.gov>
