+++ src/time_advance_tests.cpp
@@ -124,8 +124,8 @@ TEMPLATE_TEST_CASE("time advance - diffusion 2", "[time_advance]", double,
 
   SECTION("diffusion2, explicit, sparse grid, level 3, degree 3")
   {
-    int const degree = 3;
-    int const level  = 3;
+    int const degree          = 3;
+    int const level           = 3;
     auto constexpr tol_factor = get_tolerance<TestType>(100);
 
     std::string const gold_base = "../testing/generated-inputs/time_advance/"
@@ -141,9 +141,9 @@ TEMPLATE_TEST_CASE("time advance - diffusion 2", "[time_advance]", double,
 
   SECTION("diffusion2, explicit, sparse grid, level 4, degree 4")
   {
-    int const degree = 4;
-    int const level  = 4;
-    auto constexpr tol_factor = get_tolerance<TestType>(1000000);
+    int const degree            = 4;
+    int const level             = 4;
+    auto constexpr tol_factor   = get_tolerance<TestType>(1000000);
     std::string const gold_base = "../testing/generated-inputs/time_advance/"
                                   "diffusion2_sg_l4_d4_t";
 
@@ -157,7 +157,7 @@ TEMPLATE_TEST_CASE("time advance - diffusion 2", "[time_advance]", double,
 
   SECTION("diffusion2, explicit/non-uniform level, sparse grid, degree 2")
   {
-    int const degree = 2;
+    int const degree          = 2;
     auto constexpr tol_factor = get_tolerance<TestType>(100);
 
     fk::vector<int> const levels{4, 5};
@@ -345,7 +345,7 @@ TEMPLATE_TEST_CASE("time advance - continuity 1", "[time_advance]", float,
   TestType const cfl = 0.01;
 
   auto constexpr tol_factor = get_tolerance<TestType>(10);
- 
+
   auto const num_dims = 1;
 
   SECTION("continuity1, explicit, level 2, degree 2, sparse grid")
@@ -399,8 +399,8 @@ TEMPLATE_TEST_CASE("time advance - continuity 2", "[time_advance]", float,
 {
   std::string const pde_choice = "continuity_2";
   TestType const cfl           = 0.01;
-  auto constexpr tol_factor = get_tolerance<TestType>(10);
-  auto const num_dims = 2;
+  auto constexpr tol_factor    = get_tolerance<TestType>(10);
+  auto const num_dims          = 2;
 
   SECTION("continuity2, explicit, level 2, degree 2, sparse grid")
   {
@@ -468,8 +468,8 @@ TEMPLATE_TEST_CASE("time advance - continuity 3", "[time_advance]", float,
 {
   std::string const pde_choice = "continuity_3";
   TestType const cfl           = 0.01;
-  auto constexpr tol_factor = get_tolerance<TestType>(10);
-  auto const num_dims = 3;
+  auto constexpr tol_factor    = get_tolerance<TestType>(10);
+  auto const num_dims          = 3;
 
   SECTION("continuity3, explicit, level 2, degree 2, sparse grid")
   {
@@ -525,8 +525,8 @@ TEMPLATE_TEST_CASE("time advance - continuity 6", "[time_advance]", float,
 {
   std::string const pde_choice = "continuity_6";
   TestType const cfl           = 0.01;
-  auto constexpr tol_factor = get_tolerance<TestType>(10);
-  auto const num_dims = 6;
+  auto constexpr tol_factor    = get_tolerance<TestType>(10);
+  auto const num_dims          = 6;
 
   SECTION("continuity6, level 2, degree 3, sparse grid")
   {
[…]
