### 2022-01-11 679de7a
Initial update of coefficients with new operator
