### 2022-01-11 60f30b5
Add missing continuity3 updates
