### 2022-01-11 b7ba2ed
Decrease tolerance of failing test
Signed-off-by: Steven Hahn <hahnse@ornl.gov>
