### 2022-01-11 4806fc7
Remove commented out code
Signed-off-by: Steven Hahn <hahnse@ornl.gov>
