### 2022-01-12 8174ce0
Initialize cblacs with the mpi communicator (#411)
* Initialize cblacs with mpi communicator

Signed-off-by: Steven Hahn <hahnse@ornl.gov>

* Fix failing serial build

Signed-off-by: Steven Hahn <hahnse@ornl.gov>

* Remove commented out code

Signed-off-by: Steven Hahn <hahnse@ornl.gov>

* Fix scalapack build

Signed-off-by: Steven Hahn <hahnse@ornl.gov>

* Remove commented out code

Signed-off-by: Steven Hahn <hahnse@ornl.gov>
