+++ src/pde/pde_diffusion2.hpp
@@ -42,10 +42,12 @@ private:
     return fx;
   }
 
-  static fk::vector<P> moment_dV(fk::vector<P> const x, P const t = 0)
+  static P moment_dV(P const x, P const time)
   {
-    ignore(t);
-    return fk::vector<P>(std::vector<P>(x.size(), 1.0));
+    // suppress compiler warnings
+    ignore(x);
+    ignore(time);
+    return 1.0;
   }
 
   /* Define the dimension */
