+++ src/pde/pde_continuity3.hpp
@@ -93,10 +93,12 @@ private:
     return fx;
   }
 
-  static fk::vector<P> moment_dV(fk::vector<P> const x, P const t = 0)
+  static P moment_dV(P const x, P const time)
   {
-    ignore(t);
-    return fk::vector<P>(std::vector<P>(x.size(), 1.0));
+    // suppress compiler warnings
+    ignore(x);
+    ignore(time);
+    return 1.0;
   }
 
   static P exact_time(P const time) { return std::sin(2.0 * time); }
