+++ src/pde/pde_continuity1.hpp
@@ -63,10 +63,12 @@ private:
     return fx;
   }
 
-  static fk::vector<P> moment_dV_dim0(fk::vector<P> const x, P const t = 0)
+  static P moment_dV_dim0(P const x, P const time)
   {
-    ignore(t);
-    return fk::vector<P>(std::vector<P>(x.size(), 1.0));
+    // suppress compiler warnings
+    ignore(x);
+    ignore(time);
+    return 1.0;
   }
 
   static P exact_time(P const time) { return std::sin(time); }
