### 2022-01-19 f3195e0
Add mass coeff get and setters to pde and term
