### 2022-01-21 8baa1e6
Add patch to openblas. Use get_tolerance function to fix precision failure (#412)
* Add patch to openblas. Use get_tolerance function to fix precision failure.

Signed-off-by: Steven Hahn <hahnse@ornl.gov>

* clang-format

Signed-off-by: Steven Hahn <hahnse@ornl.gov>

* Decrease tolerance of failing test

Signed-off-by: Steven Hahn <hahnse@ornl.gov>
