### 2022-01-24 4eb7294
Update boundary conditions for mass terms
