+++ src/boundary_conditions.cpp
@@ -63,7 +63,7 @@ std::array<unscaled_bc_parts<P>, 2> boundary_conditions::make_unscaled_bc_parts(
 
           std::vector<fk::vector<P>> p_term_left_bcs = generate_partial_bcs(
               dimensions, dim_num, p_term.left_bc_funcs, transformer, t_init,
-              partial_terms, p_num, std::move(trace_bc));
+              terms_vec, partial_terms, p_num, std::move(trace_bc));
 
           fk::vector<P> combined =
               combine_dimensions(d.get_degree(), table, start_element,
@@ -80,7 +80,7 @@ std::array<unscaled_bc_parts<P>, 2> boundary_conditions::make_unscaled_bc_parts(
 
           std::vector<fk::vector<P>> p_term_right_bcs = generate_partial_bcs(
               dimensions, dim_num, p_term.right_bc_funcs, transformer, t_init,
-              partial_terms, p_num, std::move(trace_bc));
+              terms_vec, partial_terms, p_num, std::move(trace_bc));
 
           fk::vector<P> combined =
               combine_dimensions(d.get_degree(), table, start_element,
@@ -260,8 +260,9 @@ std::vector<fk::vector<P>> boundary_conditions::generate_partial_bcs(
     std::vector<dimension<P>> const &dimensions, int const d_index,
     std::vector<vector_func<P>> const &bc_funcs,
     basis::wavelet_transform<P, resource::host> const &transformer,
-    P const time, std::vector<partial_term<P>> const &partial_terms,
-    int const p_index, fk::vector<P> &&trace_bc)
+    P const time, std::vector<term<P>> const &terms,
+    std::vector<partial_term<P>> const &partial_terms, int const p_index,
+    fk::vector<P> &&trace_bc)
 {
   expect(d_index < static_cast<int>(dimensions.size()));
 
@@ -269,8 +270,38 @@ std::vector<fk::vector<P>> boundary_conditions::generate_partial_bcs(
 
   for (int dim_num = 0; dim_num < d_index; ++dim_num)
   {
-    partial_bc_vecs.emplace_back(forward_transform(
-        dimensions[dim_num], bc_funcs[dim_num], transformer, time));
+    auto const &f = bc_funcs[dim_num];
+    auto const &g = terms[dim_num].get_partial_terms()[p_index].g_func;
+    vector_func<P> const bc_func = [f, g](fk::vector<P> const x, P const t) {
+      // evaluate f(x,t) * g(x,t)
+      fk::vector<P> fx(f(x, t));
+      std::transform(fx.begin(), fx.end(), x.begin(), fx.begin(),
+                     [g, t](P f_elem, P const x_elem) -> P {
+                       return f_elem * g(x_elem, t);
+                     });
+      return fx;
+    };
+    // TODO: add dV to forward transform
+    partial_bc_vecs.emplace_back(
+        forward_transform(dimensions[dim_num], bc_func, transformer, time));
+
+    // Apply inverse mat
+    int n = partial_bc_vecs.back().size();
+    std::vector<int> ipiv(n);
+    fk::matrix<P, mem_type::const_view> const lhs_mass(
+        terms[dim_num].get_partial_terms()[p_index].get_lhs_mass(), 0, n - 1, 0,
+        n - 1);
+    // TODO: verify the following call is correct
+    fm::gesv(lhs_mass, partial_bc_vecs.back(), ipiv);
+
+    // Apply previous pterms
+    for (int p_num = 0; p_num < p_index; ++p_num)
+    {
+      fk::matrix<P, mem_type::const_view> const pterm_coeffs(
+          terms[dim_num].get_partial_terms()[p_num].get_coefficients(), 0,
[…]
