### 2022-01-25 833e1c8
Apply lhs mass matrix for pterm in bc
