### 2022-01-25 ff97710
Add interior penalty terms for diffusion 1D
