+++ src/transformations.hpp
@@ -47,7 +47,7 @@ combine_dimensions(int const, elements::table const &, int const, int const,
 
 template<typename P, typename F>
 fk::vector<P> forward_transform(
-    dimension<P> const &dim, F function,
+    dimension<P> const &dim, F function, g_func_type const dv_func,
     basis::wavelet_transform<P, resource::host> const &transformer,
     P const t = 0)
 {
@@ -105,6 +105,13 @@ fk::vector<P> forward_transform(
 
     // get the f(v) initial condition at the quadrature points.
     fk::vector<P> f_here = function(mapped_roots, t);
+
+    // apply dv to f(v)
+    std::transform(f_here.begin(), f_here.end(), mapped_roots.begin(),
+                   f_here.begin(), [dv_func, t](P f_elem, P const x_elem) -> P {
+                     return f_elem * dv_func(x_elem, t);
+                   });
+
     // ensuring function returns vector of appropriate size
     expect(f_here.size() == weights.size());
     std::transform(f_here.begin(), f_here.end(), weights.begin(),
@@ -154,10 +161,14 @@ inline fk::vector<P> transform_and_combine_dimensions(
   std::vector<fk::vector<P>> dimension_components;
   dimension_components.reserve(pde.num_dims);
 
+  auto const &dimensions = pde.get_dimensions();
+
   for (int i = 0; i < pde.num_dims; ++i)
   {
+    auto const &dim = dimensions[i];
     dimension_components.push_back(forward_transform<P>(
-        pde.get_dimensions()[i], v_functions[i], transformer, time));
+        dim, v_functions[i], dim.moment_dV, transformer, time));
+    // TODO: add mass_matrix \ fList, md_eval_function:21
   }
 
   return combine_dimensions(degree, table, start, stop, dimension_components,
