+++ src/transformations_tests.cpp
@@ -110,6 +110,10 @@ TEMPLATE_TEST_CASE("forward multi-wavelet transform", "[transformations]",
       ignore(t);
       return x * static_cast<TestType>(2.0);
     };
+    g_func_type const tenth_func = [](TestType x, TestType t) {
+      ignore(t);
+      return x * 0.1;
+    };
 
     auto const pde = make_PDE<TestType>(PDE_opts::continuity_1, levels, degree);
     auto const dim = pde->get_dimensions()[0];
@@ -126,7 +130,7 @@ TEMPLATE_TEST_CASE("forward multi-wavelet transform", "[transformations]",
         "_neg1_pos1_double.dat"));
 
     fk::vector<TestType> const test =
-        forward_transform<TestType>(dim, double_it, transformer);
+        forward_transform<TestType>(dim, double_it, tenth_func, transformer);
 
     rmse_comparison(gold, test, tol_factor);
   }
@@ -155,7 +159,7 @@ TEMPLATE_TEST_CASE("forward multi-wavelet transform", "[transformations]",
             std::to_string(degree) + "_" + std::to_string(levels) +
             "_neg2_pos2_doubleplus.dat"));
     fk::vector<TestType> const test =
-        forward_transform<TestType>(dim, double_plus, transformer);
+        forward_transform<TestType>(dim, double_plus, partial_term<TestType>::null_gfunc, transformer);
 
     rmse_comparison(gold, test, tol_factor);
   }
