### 2022-02-03 5b1ca18
Fix vector bounds on partial boundary conditions
