### 2022-02-08 7f916ed
Fix gemm result not being set in pterm coeff calc
