### 2022-02-09 ef6c5dc
Fix domain and issues with some pterms in fokkerplanck2d
