### 2022-02-10 d5b748c
Add function to compute dimension mass matrices and update coeff chaining
