### 2022-02-14 f6b901b
Remove time dependence from diffusion 2 pde
