+++ src/pde/pde_diffusion2.hpp
@@ -99,13 +99,13 @@ private:
       partial_term<P>::null_gfunc);
 
   inline static term<P> const term_0 =
-      term<P>(true,            // time-dependent
+      term<P>(false,           // time-dependent
               fk::vector<P>(), // additional data vector
               "",              // name
               {partial_term_0, partial_term_1});
 
   inline static term<P> const term_1 =
-      term<P>(true,            // time-dependent
+      term<P>(false,           // time-dependent
               fk::vector<P>(), // additional data vector
               "",              // name
               {partial_term_I_, partial_term_I_});
