+++ src/matlab_plot.cpp
@@ -172,6 +172,20 @@ void matlab_plot::set_var(std::string const &name,
   matlab_inst_->setVariable(name, var, type);
 }
 
+// pushes a matrix to a matlab workspace with the given name
+template<typename T>
+void matlab_plot::push(std::string const &name, fk::matrix<T> const &data,
+                       ml_wksp_type const type)
+{
+  // create a matlab array object
+  auto const &v = fk::vector<T>(data);
+  auto ml_array = create_array(
+      {static_cast<size_t>(data.ncols()), static_cast<size_t>(data.nrows())},
+      v);
+
+  set_var(name, ml_array, type);
+}
+
 template<typename T>
 fk::vector<T> matlab_plot::generate_nodes(int const degree, int const level,
                                           T const min, T const max) const
@@ -316,6 +330,149 @@ void matlab_plot::plot_fval(PDE<P> const &pde, elements::table const &table,
   call("plot_fval");
 }
 
+// copies data about a PDE object to a matlab workspace
+template<typename P>
+void matlab_plot::copy_pde(PDE<P> const &pde)
+{
+  expect(is_open());
+
+  std::vector<std::string> pde_names = {"solvePoisson",
+                                        "hasAnalyticSoln",
+                                        "coefficients",
+                                        "dims",
+                                        "terms",
+                                        "sources",
+                                        "dt"};
+
+  matlab::data::StructArray ml_pde = factory_.createStructArray({1}, pde_names);
+
+  ml_pde[0]["solvePoisson"] = factory_.createScalar<bool>(pde.do_poisson_solve);
+  ml_pde[0]["hasAnalyticSoln"] =
+      factory_.createScalar<bool>(pde.has_analytic_soln);
+
+  // create the dimensions cell
+  matlab::data::CellArray ml_dims =
+      factory_.createCellArray({1, static_cast<size_t>(pde.num_dims)});
+  for (auto i = 0; i < pde.num_dims; ++i)
+  {
+    ml_dims[0][i] = make_dimension(pde.get_dimensions()[i]);
+  }
+  ml_pde[0]["dims"] = ml_dims;
+
+  // create the terms cell
+  matlab::data::CellArray ml_terms = factory_.createCellArray(
+      {static_cast<size_t>(pde.num_dims), static_cast<size_t>(pde.num_terms)});
+  for (auto i = 0; i < pde.num_dims; ++i)
+  {
+    for (auto j = 0; j < pde.num_terms; ++j)
+    {
+      ml_terms[i][j] = make_term(pde.get_terms()[j][i]);
+    }
+  }
+  ml_pde[0]["terms"] = ml_terms;
+
+  ml_pde[0]["dt"] = factory_.createScalar<P>(pde.get_dt());
+
+  // push the pde struct to the matlab workspace
+  matlab_inst_->setVariable("pde", ml_pde);
+}
+
+template<typename P>
+matlab::data::StructArray matlab_plot::make_term(term<P> const &term)
+{
+  std::vector<std::string> names    = {"time_dependent", "name", "data",
+                                    "coefficients", "pterms"};
+  matlab::data::StructArray ml_term = factory_.createStructArray({1}, names);
+
+  ml_term[0]["time_dependent"] =
+      factory_.createScalar<bool>(term.time_dependent);
+  ml_term[0]["name"]    = factory_.createCharArray(term.name);
+  auto const &term_data = term.get_data();
+  ml_term[0]["data"] =
+      create_array({1, static_cast<size_t>(term_data.size())}, term_data);
+  auto const &coeffs         = term.get_coefficients().clone_onto_host();
+  ml_term[0]["coefficients"] = matrix_to_array(coeffs);
+
[…]
