### 2022-02-16 28720dc
Fix rotation issue in dimension mass matrix generation
