### 2022-02-18 45f0f44
Revert chaining behavior to dof size
