### 2022-03-01 9d8c44e
Add the other cases for the fokkerplanck 2d pde
