+++ src/pde/pde_fokkerplanck2_complete.hpp
@@ -43,7 +43,7 @@
 // termR2 == -1/(tau*gam(p)) f(p) * d/dz z(1-z^2) f(z)
 // ---------------------------------------------------------------------------
 
-template<typename P>
+template<typename P, PDE_case_opts user_case = PDE_case_opts::case1>
 class PDE_fokkerplanck_2d_complete : public PDE<P>
 {
 public:
@@ -59,7 +59,7 @@ private:
 
   static int constexpr num_dims_           = 2;
   static int constexpr num_sources_        = 0;
-  static int constexpr num_terms_          = 6;
+  static int constexpr num_terms_          = 3;
   static bool constexpr do_poisson_solve_  = false;
   static bool constexpr has_analytic_soln_ = false;
 
@@ -79,11 +79,46 @@ private:
     return ret;
   };
 
-  static P constexpr nuEE     = 1;
-  static P constexpr vT       = 1;
-  static P constexpr delta    = 0.3;
-  static P constexpr Z        = 5;
-  static P constexpr E        = 0.4;
+  static P constexpr nuEE  = 1;
+  static P constexpr vT    = 1;
+  static P constexpr delta = []() {
+    if constexpr (user_case == PDE_case_opts::case4)
+    {
+      return 0.3;
+    }
+    else
+    {
+      return 0.042;
+    }
+  }();
+  static P constexpr Z = []() {
+    if constexpr (user_case == PDE_case_opts::case4)
+    {
+      return 5.0;
+    }
+    else
+    {
+      return 1.0;
+    }
+  }();
+  static P constexpr E = []() {
+    if constexpr (user_case == PDE_case_opts::case1)
+    {
+      return 0.0025;
+    }
+    else if constexpr (user_case == PDE_case_opts::case2)
+    {
+      return 0.25;
+    }
+    else if constexpr (user_case == PDE_case_opts::case3)
+    {
+      return 0.0025;
+    }
+    else if constexpr (user_case == PDE_case_opts::case4)
+    {
+      return 0.4;
+    }
+  }();
   static P constexpr tau      = 1e5;
   static auto constexpr gamma = [](P p) {
     return std::sqrt(1 + std::pow(delta * p, 2));
@@ -112,7 +147,52 @@ private:
   // p dimension
 
   // initial condition in p
-  static fk::vector<P> initial_condition_p(fk::vector<P> const x, P const t = 0)
+  static fk::vector<P>
+  initial_condition_p_case1(fk::vector<P> const x, P const t = 0)
+  {
+    ignore(t);
+
+    fk::vector<P> transformed(x);
+    std::transform(x.begin(), x.end(), transformed.begin(),
+                   [](P const x_elem) -> P {
+                     if (x_elem <= 5.0)
+                     {
+                       return 3.0 / (2.0 * std::pow(5, 3));
+                     }
+                     else
+                     {
+                       return 0.0;
+                     }
+                   });
+    return transformed;
+  }
+  static fk::vector<P>
+  initial_condition_p_case2(fk::vector<P> const x, P const t = 0)
+  {
+    ignore(t);
+
+    fk::vector<P> transformed(x);
+    std::transform(x.begin(), x.end(), transformed.begin(),
+                   [](P const x_elem) -> P {
+                     return exp(-std::pow(x_elem, 2) / std::pow(2, 3));
+                   });
+    transformed.scale(2.0 / (sqrt(M_PI) * std::pow(2, 3)));
+    return transformed;
+  }
+  static fk::vector<P>
+  initial_condition_p_case3(fk::vector<P> const x, P const t = 0)
+  {
+    ignore(t);
+
[…]

[…]
