+++ src/main.cpp
@@ -157,7 +157,7 @@ int main(int argc, char **argv)
   std::array<fk::vector<prec, mem_type::view, resource::host>, 2>
       tmp_workspace = {
           fk::vector<prec, mem_type::view, resource::host>(workspace, 0,
-                                                           real_space_size),
+                                                           real_space_size - 1),
           fk::vector<prec, mem_type::view, resource::host>(
               workspace, real_space_size, real_space_size * 2 - 1)};
   // transform initial condition to realspace
@@ -192,6 +192,16 @@ int main(int argc, char **argv)
   ml_plot.init_plotting(*pde, adaptive_grid.get_table());
   ml_plot.plot_fval(*pde, adaptive_grid.get_table(), real_space,
                     analytic_solution_realspace);
+
+  // send initial condition to matlab
+  std::vector<size_t> sizes(pde->num_dims);
+  for (int i = 0; i < pde->num_dims; i++)
+  {
+    sizes[i] = pde->get_dimensions()[i].get_degree() *
+               std::pow(2, pde->get_dimensions()[i].get_level());
+  }
+  ml_plot.set_var("initial_condition",
+                  ml_plot.create_array(sizes, initial_condition));
 
   ml_plot.copy_pde(*pde);
 #endif
@@ -263,11 +273,12 @@ int main(int argc, char **argv)
 #ifdef ASGARD_USE_MATLAB
       if (opts.should_plot(i))
       {
-        auto transform_wksp = update_transform_workspace<prec>(
-            sol.size(), workspace, tmp_workspace);
-        if (analytic_solution.size() > analytic_solution_realspace.size())
+        auto const real_size = real_solution_size(*pde);
+        auto transform_wksp  = update_transform_workspace<prec>(
+            real_size, workspace, tmp_workspace);
+        if (real_size > analytic_solution_realspace.size())
         {
-          analytic_solution_realspace.resize(analytic_solution.size());
+          analytic_solution_realspace.resize(real_size);
         }
         wavelet_to_realspace<prec>(*pde, analytic_solution,
                                    adaptive_grid.get_table(), transformer,
