+++ src/pde/pde_base.hpp
@@ -514,6 +514,16 @@ public:
       expect(d.domain_max > d.domain_min);
     }
 
+    // initialize mass matrices to a default value
+    for (auto i = 0; i < num_dims; ++i)
+    {
+      auto const max_dof =
+          fm::two_raised_to(static_cast<int64_t>(cli_input.get_max_level())) *
+          degree;
+      expect(max_dof < INT_MAX);
+      update_dimension_mass_mat(i, eye<P>(max_dof));
+    }
+
     // check all sources
     for (auto const &s : sources)
     {
