+++ src/boundary_conditions_tests.cpp
@@ -151,6 +151,7 @@ TEMPLATE_TEST_CASE("problem separability", "[boundary_condition]", double,
 
     basis::wavelet_transform<TestType, resource::host> const transformer(opts,
                                                                          *pde);
+    generate_dimension_mass_mat(*pde, transformer);
     generate_all_coefficients<TestType>(*pde, transformer);
 
     /* initialize bc vector at test_time */
@@ -195,6 +196,7 @@ TEMPLATE_TEST_CASE("problem separability", "[boundary_condition]", double,
 
     basis::wavelet_transform<TestType, resource::host> const transformer(opts,
                                                                          *pde);
+    generate_dimension_mass_mat(*pde, transformer);
     generate_all_coefficients<TestType>(*pde, transformer);
 
     /* initialize bc vector at test_time */
