+++ testing/tests_general.hpp
@@ -6,7 +6,6 @@
 
 #pragma once
 
-#include "catch.hpp"
 #include "src/fast_math.hpp"
 #include "src/pde.hpp"
 #include "src/program_options.hpp"
@@ -14,6 +13,7 @@
 #include <string>
 #include <utility>
 #include <vector>
+#include <catch2/catch_all.hpp>
 
 template<typename P>
 constexpr P get_tolerance(int ulp)
@@ -68,7 +68,7 @@ void compare_vectors(std::vector<P> const &a, std::vector<P> const &b)
   if constexpr (std::is_floating_point<P>::value)
   {
     for (size_t i = 0; i < a.size(); i++)
-      if (a[i] != Approx(b[i]).epsilon(std::numeric_limits<P>::epsilon() * 2))
+      if (a[i] != Catch::Approx(b[i]).epsilon(std::numeric_limits<P>::epsilon() * 2))
         FAIL("" << a[i] << " != " << b[i]);
   }
   else
