+++ testing/tests_general.cpp
@@ -10,7 +10,6 @@
 
 #define CATCH_CONFIG_MAIN
 #include "tests_general.hpp"
-#include "catch.hpp"
 #include <string>
 #include <vector>
 

[…]
