### 2022-03-15 46358e4
Expand CI test coverage to include MPI and Scalapack cases.
