### 2022-03-15 c994715
Make the conditonal a string.
