### 2022-03-15 2ccddf6
Update coefficient and transformation tests
