### 2022-03-15 338c2d1
Add empty moment class
Signed-off-by: Steven Hahn <hahnse@ornl.gov>
