+++ CMakeLists.txt
@@ -41,7 +41,7 @@ endfunction ()
 register_project (kronmult
                   KRONMULT
                   https://github.com/project-asgard/kronmult.git
-                  1ed1a5e1b3690f66479dcd9cf003dc8a3a0de457
+                  be1ef32b9e4d75f977c066b8917e6b9adce58a03
 )
 
 #  Catch two testing framework.
@@ -54,8 +54,8 @@ register_project (Catch2
 ###############################################################################
 ## Set up the compiler and general global build options
 ###############################################################################
-set (CMAKE_CXX_STANDARD 17)
-set (CMAKE_CXX_STANDARD_REQUIRED ON)
+#set (CMAKE_CXX_STANDARD 17)
+#set (CMAKE_CXX_STANDARD_REQUIRED ON)
 
 # Set a default build type if none was specified
 # https://blog.kitware.com/cmake-and-the-default-build-type/
@@ -101,7 +101,12 @@ include(CMakeDependentOption)
 cmake_dependent_option(ASGARD_USE_SCALAPACK "Use optional scalapack linear algebra library" OFF "ASGARD_USE_MPI" OFF)
 option (ASGARD_USE_MATLAB "Optional MATLAB plotting support for asgard" OFF)
 option (ASGARD_USE_DOUBLE_PREC "enable or disable double-precision floating point" ON)
+option (ASGARD_USE_PCH "Enable precompiled header files." OFF)
 set (ASGARD_TESTING_RANKS "0" CACHE STRING "Override default number of ranks to use for testing")
+
+if (${ASGARD_USE_PCH})
+  set (USE_PCH ON CACHE BOOL "Turn on kronmult precompiled header support" FORCE)
+endif()
 
 # add scripts directory location
 set(ASGARD_SCRIPTS_DIR "${CMAKE_SOURCE_DIR}/scripts/")
@@ -142,7 +147,12 @@ endif()
 #endif ()
 
 if(ASGARD_USE_OPENMP)
-  find_package(OpenMP REQUIRED)  
+  find_package(OpenMP)
+  
+  if (NOT ${OpenMP_FOUND})
+    set (ASGARD_USE_OPENMP OFF CACHE BOOL "Optional openMP support for asgard" FORCE)
+  endif ()
+  
   if(ASGARD_USE_MKL)
     if(CMAKE_CXX_COMPILER_ID STREQUAL "AppleClang")
 	message(FATAL_ERROR 
@@ -152,160 +162,12 @@ if(ASGARD_USE_OPENMP)
   endif()
 endif()
 
-if(ASGARD_USE_CUDA)
-  find_package(CUDA 9.0 REQUIRED) # eventually want to remove this - how to set min version with enable_language?
-  include_directories(${CUDA_INCLUDE_DIRS})
-  enable_language(CUDA) 
-  set (CMAKE_CUDA_STANDARD 14)
-  set (CMAKE_CUDA_STANDARD_REQUIRED ON)
-endif()
-
-# build component to interface with Ed's kronmult lib
-##TODO: link to kronmult as interface library
-add_library(kronmult_cuda SHARED src/device/kronmult_cuda.cpp)
-if(ASGARD_USE_CUDA)
-  set_source_files_properties( src/device/kronmult_cuda.cpp PROPERTIES LANGUAGE CUDA ) # no .cu extension
-  set_target_properties( kronmult_cuda PROPERTIES CUDA_ARCHITECTURES OFF)
-  set_target_properties( kronmult_cuda PROPERTIES COMPILE_FLAGS "-arch sm_70 -g -lineinfo --ptxas-options=-O3")
-  set_target_properties( kronmult_cuda PROPERTIES LINK_FLAGS "-Wl,-rpath,${CMAKE_BINARY_DIR}")
-
[…]
