+++ src/moment_tests.cpp
@@ -3,9 +3,6 @@
 
 TEMPLATE_TEST_CASE("Multiwavelet", "[transformations]", double, float)
 {
-  std::vector<vector_func<TestType>> md_func;	
-  SECTION("Constructor")
-  {	  
-    moment<TestType> mymoment(md_func);
-  }
+  std::vector<vector_func<TestType>> md_func;
+  SECTION("Constructor") { moment<TestType> mymoment(md_func); }
 }
