### 2022-03-22 b279aaf
Fix issue in boundary condition vector, update testing data
