### 2022-03-22 5d62e49
Add template type to g_func_type
