+++ src/adapt_tests.cpp
@@ -216,8 +216,8 @@ void test_initial(parser const &problem, std::string const &gold_filepath)
   auto const quiet = true;
   basis::wavelet_transform<P, resource::host> const transformer(opts, *pde,
                                                                 quiet);
-  generate_dimension_mass_mat<P>(*pde, transformer);
   adapt::distributed_grid<P> adaptive_grid(*pde, opts);
+  generate_dimension_mass_mat<P>(*pde, transformer);
 
   auto const test =
       adaptive_grid.get_initial_condition(*pde, transformer, opts);
