### 2022-03-25 84134ae
Store pre-computed pterm coefficients for all levels
