### 2022-03-30 8495022
Change forward transform zero cutoff to use float epsilon
