### 2022-03-30 7da6d88
Add tests for fokkerplanck2d complete pterm functions
