### 2022-03-31 97fea97
createFlist member function
Signed-off-by: Steven Hahn <hahnse@ornl.gov>
