### 2022-04-05 8369237
Refactor pterm coefficient construction, add tests for pterm mass mat
