+++ src/adapt_tests.cpp
@@ -222,7 +222,7 @@ void test_initial(parser const &problem, std::string const &gold_filepath)
   auto const test =
       adaptive_grid.get_initial_condition(*pde, transformer, opts);
 
-  REQUIRE(gold.size() == test.size());
+  REQUIRE(gold.size() >= test.size());
 
   auto constexpr tol_factor = get_tolerance<P>(100);
   auto const my_subgrid     = adaptive_grid.get_subgrid(get_rank());
