+++ src/coefficients.hpp
@@ -18,4 +18,4 @@ fk::matrix<P> generate_coefficients(
     dimension<P> const &dim, term<P> const &term_1D,
     partial_term<P> const &pterm,
     basis::wavelet_transform<P, resource::host> const &transformer,
-    P const time = 0.0, bool const rotate = true);
+    int const level, P const time = 0.0, bool const rotate = true);
