### 2022-04-05 08c2193
Adjust test tolerances for single precision
