### 2022-04-19 b7dbca9
Add mac tests (#415)
* Expand CI test coverage to include MPI and Scalapack cases.

* Remove tab to fix yaml syntax error.

* Move condition into the expression so it evaluates as true not a string.

* Make the conditonal a string.
