+++ .github/workflows/ci_tests.yaml
@@ -9,21 +9,39 @@ jobs:
             matrix:
                 os: [macos-latest]
                 node: [Release, Debug]
+                features: [Default, MPI, ScaLAPACK]
         runs-on: ${{matrix.os}}
         steps:
-            - name: CPU Infor
+            - name: CPU Info
               run: |
                 sysctl -n machdep.cpu.brand_string
                 sysctl -n machdep.cpu.features
                 (echo "Number of Cores:"; sysctl -n machdep.cpu.core_count) |tr '\n' '\t' && echo ''
+            - name: Install MPI
+              if: ${{matrix.features == 'MPI'}}
+              working-directory: ${{runner.workspace}}
+              run: brew install open-mpi
+            - name: Install ScaLAPACK
+              if: ${{matrix.features == 'ScaLAPACK'}}
+              working-directory: ${{runner.workspace}}
+              run: brew install scalapack
             - name: Git checkout
               uses: actions/checkout@v1
             - name: Makedir
               working-directory: ${{runner.workspace}}/asgard
               run: cmake -E make_directory build
-            - name: Configure
+            - name: Configure Default
+              if: ${{matrix.features  == 'Default'}}
               working-directory: ${{runner.workspace}}/asgard/build
               run: cmake -DCMAKE_BUILD_TYPE=${{matrix.node}} -DASGARD_USE_OPENMP=OFF ../
+            - name: Configure MPI
+              if: ${{matrix.features == 'MPI'}}
+              working-directory: ${{runner.workspace}}/asgard/build
+              run: cmake -DCMAKE_BUILD_TYPE=${{matrix.node}} -DASGARD_USE_OPENMP=OFF -DASGARD_USE_MPI=ON ../
+            - name: Configure ScaLAPACK
+              if: ${{matrix.features == 'ScaLAPACK'}}
+              working-directory: ${{runner.workspace}}/asgard/build
+              run: cmake -DCMAKE_BUILD_TYPE=${{matrix.node}} -DASGARD_USE_OPENMP=OFF -DASGARD_USE_MPI=ON -DASGARD_USE_SCALAPACK=ON -DCMAKE_Fortran_COMPILER=/usr/local/opt/gfortran/bin/gfortran ../
             - name: Build
               working-directory: ${{runner.workspace}}/asgard/build
               run: make VERBOSE=1
