+++ src/moment.hpp
@@ -1,7 +1,8 @@
 #pragma once
-
+#include "elements.hpp"
 #include "pde/pde_base.hpp"
 #include "program_options.hpp"
+#include "tensors.hpp"
 #include <vector>
 
 template<typename P>
@@ -10,11 +11,13 @@ class moment
 public:
   moment(std::vector<vector_func<P>> md_funcs_);
   void createFlist(PDE<P> const &pde, options const &opts);
-  // createMomentVector()
+  void
+  createMomentVector(parser const &opts, elements::table const &hash_table);
+
 private:
   std::vector<vector_func<P>> md_funcs;
   std::vector<std::vector<fk::vector<P>>> fList;
-  // vector;
+  fk::vector<P> vector;
   // moment_fval_integral;
   // moment_analytic_integral;
 };
