### 2022-04-26 8f4033a
Use negate in coeffs and direct access in boundary cond
