### 2022-05-01 #3953 @arfon
:wave: @cticenhour & @drobnyjt - would either of you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html

The submission under consideration is _ASGarD: Adaptive Sparse Grid Discretization_
