### 2022-05-03 93906e1
Refactor matlab boundary condition string
