### 2022-05-05 #3953 @arfon
> Hi @arfon! I also would be available, but I do need to disclose that the first author on this work (David L Green) is a member of my PhD committee (not my dissertation advisor, however). We have not collaborated on this particular work. If this is a CoI, please let me know.

Thank you for flagging this @cticenhour. I'm happy to waive this possible conflict.
