+++ src/coefficients_tests.cpp
@@ -13,6 +13,7 @@ void test_coefficients(parser const &parse, std::string const &gold_path,
   options const opts(parse);
   basis::wavelet_transform<P, resource::host> const transformer(opts, *pde);
   P const time = 1.0;
+  generate_dimension_mass_mat(*pde, transformer);
   generate_all_coefficients(*pde, transformer, time, rotate);
 
   auto const lev_string = std::accumulate(
@@ -180,7 +181,7 @@ TEMPLATE_TEST_CASE("fokkerplanck1_pitch_E case1 terms", "[coefficients]",
 {
   auto const pde_choice = PDE_opts::fokkerplanck_1d_pitch_E_case1;
   auto const gold_path  = "../testing/generated-inputs/coefficients/"
-                         "fokkerplanck1_pitch_E_case1_coefficients";
+                         "fokkerplanck1_4p1a_coefficients";
   auto constexpr tol_factor = get_tolerance<TestType>(10);
 
   SECTION("level 4, degree 3")
@@ -214,7 +215,7 @@ TEMPLATE_TEST_CASE("fokkerplanck1_pitch_C terms", "[coefficients]", double,
 {
   auto const pde_choice = PDE_opts::fokkerplanck_1d_pitch_C;
   auto const gold_path  = "../testing/generated-inputs/coefficients/"
-                         "fokkerplanck1_pitch_C_coefficients";
+                         "fokkerplanck1_4p2_coefficients";
   TestType const tol_factor =
       std::is_same<TestType, double>::value ? 1e-14 : 1e-5;
 
@@ -278,13 +279,13 @@ TEMPLATE_TEST_CASE("fokkerplanck1_4p5 terms", "[coefficients]", double, float)
   }
 }
 
-TEMPLATE_TEST_CASE("fokkerplanck2_complete terms", "[coefficients]", double,
-                   float)
+TEMPLATE_TEST_CASE("fokkerplanck2_complete_case4 terms", "[coefficients]",
+                   double, float)
 {
   auto const gold_path = "../testing/generated-inputs/coefficients/"
                          "fokkerplanck2_complete_coefficients";
 
-  auto const pde_choice = PDE_opts::fokkerplanck_2d_complete;
+  auto const pde_choice = PDE_opts::fokkerplanck_2d_complete_case4;
   TestType const tol_factor =
       std::is_same<TestType, double>::value ? 1e-12 : 1e-3;
 
@@ -317,5 +318,49 @@ TEMPLATE_TEST_CASE("fokkerplanck2_complete terms", "[coefficients]", double,
     auto const degree = 4;
     parser const test_parse(pde_choice, levels, degree);
     test_coefficients<TestType>(test_parse, gold_path, tol_factor);
+  }
+
+  SECTION("pterm lhs mass")
+  {
+    auto const gold = fk::matrix<TestType>(
+        read_matrix_from_txt_file(std::string(gold_path) + "_lhsmass.dat"));
+
+    auto constexpr tol_factor = get_tolerance<TestType>(100);
+    auto const levels         = fk::vector<int>{4, 4};
+    int const degree          = 4;
+
+    parser const test_parse(pde_choice, levels, degree);
+    auto pde = make_PDE<TestType>(test_parse);
+    options const opts(test_parse);
+
+    basis::wavelet_transform<TestType, resource::host> const transformer(opts,
+                                                                         *pde);
+    TestType const time = 1.0;
+    generate_dimension_mass_mat(*pde, transformer);
+    generate_all_coefficients(*pde, transformer, time, true);
+
+    int row = 0;
+    for (auto i = 0; i < pde->num_dims; ++i)
+    {
+      for (auto j = 0; j < pde->num_terms; ++j)
+      {
[…]
