### 2022-05-20 f6f989f
Update Catch2 to v3.0.1
Signed-off-by: Steven Hahn <hahnse@ornl.gov>
