### 2022-06-10 3ebe0b8
Get PROJECT_SOURCE_DIR from cmake instead of  ../testing
Signed-off-by: Steven Hahn <hahnse@ornl.gov>
