+++ src/build_info.hpp.in
@@ -13,3 +13,4 @@
 #cmakedefine ASGARD_USE_DOUBLE_PREC
 
 #cmakedefine ASGARD_SCRIPTS_DIR "@ASGARD_SCRIPTS_DIR@"
+#cmakedefine ASGARD_GOLD_BASE_DIR "@ASGARD_GOLD_BASE_DIR@"

[…]
