+++ src/pde/pde_vlasov_lb_full_f.hpp
@@ -1,0 +1,450 @@
+#pragma once
+#include "pde_base.hpp"
+
+// 2D test case using continuity equation, i.e.,
+//
+//  df/dt == -v*\grad_x f + div_v( (v-u)f + theta\grad_v f)
+//
+//  BC is peridoic in x
+//  BC in v is all inflow in advection for v and Neumann for diffusion in v
+
+template<typename P>
+class PDE_vlasov_lb : public PDE<P>
+{
+public:
+  PDE_vlasov_lb(parser const &cli_input)
+      : PDE<P>(cli_input, num_dims_, num_sources_, num_terms_, dimensions_,
+               terms_, sources_, exact_vector_funcs_, exact_scalar_func_,
+               get_dt_, do_poisson_solve_, has_analytic_soln_)
+  {}
+
+private:
+  static int constexpr num_dims_           = 2;
+  static int constexpr num_sources_        = 0;
+  static int constexpr num_terms_          = 5;
+  static bool constexpr do_poisson_solve_  = false;
+  static bool constexpr has_analytic_soln_ = false;
+  static int constexpr default_degree      = 3;
+
+  static P constexpr nu = 1e3;
+
+  static fk::vector<P>
+  initial_condition_dim_x(fk::vector<P> const &x, P const t = 0)
+  {
+    // TODO: fix this - needs multi-dimensional func ability
+    P const coefficient = 1.0 / std::sqrt(2.0 * PI);
+
+    fk::vector<P> fx(x.size());
+    std::transform(x.begin(), x.end(), fx.begin(), [](P const x) -> P {
+      if (std::abs(x) > 0.5)
+      {
+        return coefficient * std::exp(-0.5 * std::pow(x, 2));
+      }
+      else
+      {
+        return 0.0;
+      }
+    });
+    return fx;
+  }
+
+  static fk::vector<P>
+  initial_condition_dim_v(fk::vector<P> const &x, P const t = 0)
+  {
+    // TODO: fix this - needs multi-dimensional func ability
+    P const coefficient = (1.0 / 8.0) / std::sqrt(2.0 * PI * (4.0 / 5.0));
+
+    fk::vector<P> fx(x.size());
+    std::transform(x.begin(), x.end(), fx.begin(), [](P const x) -> P {
+      if (x <= 0.5)
+      {
+        return coefficient * std::exp(-std::pow(x, 2) / (2.0 * (4.0 / 5.0)));
+      }
+      else
+      {
+        return 0.0;
+      }
+    });
+    return fx;
+  }
+
+  static P dV(P const x, P const time)
+  {
+    ignore(x);
+    ignore(time);
+    return 1.0;
+  }
+
+  /* Define the dimension */
+  inline static dimension<P> const dim_0 = dimension<P>(
+      -1.0, 1.0, 8, default_degree, initial_condition_dim_x, dV, "x");
+
+  inline static dimension<P> const dim_1 = dimension<P>(
+      -6.0, 6.0, 3, default_degree, initial_condition_dim_v, dV, "v");
+
+  inline static std::vector<dimension<P>> const dimensions_ = {dim_0, dim_1};
+
+  /* Define the moments */
+  static fk::vector<P> moment0_f1(fk::vector<P> const &x, P const t = 0)
+  {
+    fk::vector<P> f(x.size());
+    std::fill(f.begin(), f.end(), 1.0);
+    return f;
+  }
+
+  static fk::vector<P> moment1_f1(fk::vector<P> const &x, P const t = 0)
+  {
+    return fk::vector<P>(x);
+  }
+
+  static fk::vector<P> moment2_f1(fk::vector<P> const &x, P const t = 0)
+  {
+    fk::vector<P> f(x.size());
[…]
