### 2022-06-16 e815adf
Get PROJECT_SOURCE_DIR from cmake instead of  ../testing (#423)
Signed-off-by: Steven Hahn <hahnse@ornl.gov>
