### 2022-06-28 3969ea8
Formatting changes from clang-format-12
Signed-off-by: Steven Hahn <hahnse@ornl.gov>
