+++ src/main.cpp
@@ -226,12 +226,12 @@ int main(int argc, char **argv)
     // take a time advance step
     auto const time          = (i + 1) * pde->get_dt();
     auto const update_system = i == 0;
-    auto const method = opts.use_implicit_stepping ? time_advance::method::imp
-                                                   : time_advance::method::exp;
+    auto const method   = opts.use_implicit_stepping ? time_advance::method::imp
+                                                     : time_advance::method::exp;
     auto const time_str = opts.use_implicit_stepping ? "implicit_time_advance"
                                                      : "explicit_time_advance";
-    auto const time_id = tools::timer.start(time_str);
-    auto const sol     = time_advance::adaptive_advance(
+    auto const time_id  = tools::timer.start(time_str);
+    auto const sol      = time_advance::adaptive_advance(
         method, *pde, adaptive_grid, transformer, opts, f_val, time,
         default_workspace_MB, update_system);
     f_val.resize(sol.size()) = sol;
