### 2022-06-29 16e87af
Formatting changes from clang-format-12 (#424)
Signed-off-by: Steven Hahn <hahnse@ornl.gov>
