+++ src/tensors.hpp
@@ -240,10 +240,10 @@ public:
   template<mem_type omem, resource r_ = resrc, typename = enable_for_host<r_>>
   vector<P> operator-(vector<P, omem> const &right) const;
   template<mem_type omem, resource r_ = resrc, typename = enable_for_host<r_>>
-  P operator*(vector<P, omem> const &)const;
+  P operator*(vector<P, omem> const &) const;
 
   template<mem_type omem, resource r_ = resrc, typename = enable_for_host<r_>>
-  vector<P> operator*(matrix<P, omem> const &)const;
+  vector<P> operator*(matrix<P, omem> const &) const;
 
   template<resource r_ = resrc, typename = enable_for_host<r_>>
   vector<P> operator*(P const) const;
@@ -475,9 +475,9 @@ public:
   template<resource r_ = resrc, typename = enable_for_host<r_>>
   matrix<P> operator*(P const) const;
   template<mem_type omem, resource r_ = resrc, typename = enable_for_host<r_>>
-  vector<P> operator*(vector<P, omem> const &)const;
+  vector<P> operator*(vector<P, omem> const &) const;
   template<mem_type omem, resource r_ = resrc, typename = enable_for_host<r_>>
-  matrix<P> operator*(matrix<P, omem> const &)const;
+  matrix<P> operator*(matrix<P, omem> const &) const;
   template<mem_type omem, resource r_ = resrc, typename = enable_for_host<r_>>
   matrix<P> operator+(matrix<P, omem> const &) const;
   template<mem_type omem, resource r_ = resrc, typename = enable_for_host<r_>>
@@ -996,8 +996,8 @@ fk::vector<P, mem, resrc>::vector(vector<P, mem, resrc> const &a)
 // http://stackoverflow.com/questions/3279543/what-is-the-copy-and-swap-idiom
 //
 template<typename P, mem_type mem, resource resrc>
-fk::vector<P, mem, resrc> &fk::vector<P, mem, resrc>::
-operator=(vector<P, mem, resrc> const &a)
+fk::vector<P, mem, resrc> &
+fk::vector<P, mem, resrc>::operator=(vector<P, mem, resrc> const &a)
 {
   static_assert(mem != mem_type::const_view,
                 "cannot copy assign into const_view!");
@@ -1042,8 +1042,8 @@ fk::vector<P, mem, resrc>::vector(vector<P, mem, resrc> &&a)
 // vector move assignment
 //
 template<typename P, mem_type mem, resource resrc>
-fk::vector<P, mem, resrc> &fk::vector<P, mem, resrc>::
-operator=(vector<P, mem, resrc> &&a)
+fk::vector<P, mem, resrc> &
+fk::vector<P, mem, resrc>::operator=(vector<P, mem, resrc> &&a)
 {
   static_assert(mem != mem_type::const_view,
                 "cannot move assign into const_view!");
@@ -1089,8 +1089,8 @@ fk::vector<P, mem, resrc>::vector(vector<PP, omem> const &a)
 //
 template<typename P, mem_type mem, resource resrc>
 template<typename PP, mem_type omem, mem_type, typename, resource, typename>
-fk::vector<P, mem> &fk::vector<P, mem, resrc>::
-operator=(vector<PP, omem> const &a)
+fk::vector<P, mem> &
+fk::vector<P, mem, resrc>::operator=(vector<PP, omem> const &a)
 {
   expect(size() == a.size());
 
@@ -1124,8 +1124,8 @@ fk::vector<P, mem, resrc>::vector(vector<P, omem, resrc> const &a)
[…]
