+++ src/time_advance_tests.cpp
@@ -80,7 +80,7 @@ void time_advance_test(parser const &parse,
     auto const update_system      = i == 0;
     auto const method = opts.use_implicit_stepping ? time_advance::method::imp
                                                    : time_advance::method::exp;
-    auto const sol = time_advance::adaptive_advance(
+    auto const sol    = time_advance::adaptive_advance(
         method, *pde, adaptive_grid, transformer, opts, f_val, time,
         workspace_limit_MB, update_system);
 
