### 2022-07-02 project-asgard/asgard#425 @joglekara
JOSS Review: Build on M1 mac
Hi all,

This is part of the JOSS review at https://github.com/openjournals/joss-reviews/issues/4378.

I'm having trouble building on an M1 Mac.

I installed `openblas` and `gcc` via homebrew. I am providing both to the cmake command as follows
`cmake .. -DCMAKE_CXX_COMPILER=g++-11 -DASGARD_BLAS_PATH=/opt/homebrew/Cellar/openblas/`

Unfortunately, it is unable to find openblas
```bash
-- Found BLAS: /Library/Developer/CommandLineTools/SDKs/MacOSX12.3.sdk/System/Library/Frameworks/Accelerate.framework
-- Looking for cheev_
-- Looking for cheev_ - found
-- Found LAPACK: /Library/Developer/CommandLineTools/SDKs/MacOSX12.3.sdk/System/Library/Frameworks/Accelerate.framework;-lm;-ldl
```
I have also tried, `/opt/homebrew/Cellar/openblas/0.3.20/`.

Then, when I `make`, I see an error due to some implementation in the Apple Accelerate framework. This makes me think that we might be in the clear once I can figure out how to pass the path to the compile correctly. Any help would be appreciated.

Thank you!
