### 2022-07-05 project-asgard/asgard#425 @cianciosa
You need to set `BLA_VENDOR` to `OpenBLAS`. Otherwise it will look for the default which will be the system library. Alternatively can also set `ASGARD_BUILD_OPENBLAS=ON` and it will build its own copy of OpenBlas. However on an M1 Mac, it’s better to use the Accelerate since it has access to the built in tensor hardware.

What compile error are you are getting with Accelerate?
