### 2022-07-07 447a971
documented physics of 1d continuity equation
