### 2022-07-08 258f3c6
Add moment checking in pde
