### 2022-07-29 91a97b3
Refactor asgard to be a single library. Scalapack is support is currently broken. (#416)
Speed up tests by optimizing memory access order in tensors.

Reduce size of kronmult tests.

Enable cuda language before setting cuda properties.

Fix clang formatting in tensors.hpp

Guard the CUDA language features to only run when cuda is enabled.

Further reduce size of kronmult tests.

If def out the destructors to avoid warnings that show in clang but not gcc.

Refactor asgard to be a single library. Scalapack is support is currently broken.

Speed up tests by optimizing memory access order in tenso
[…]
