+++ src/tensors.hpp
@@ -263,7 +263,7 @@ public:
   // just get a pointer. cannot deref/assign. for e.g. blas
   // use subscript operators for general purpose access
   // this can be offsetted for views
-  P *data(int const elem = 0) const { return &data_[elem]; }
+  P *data(int const elem = 0) const { return data_ + elem; }
 
   // this is to allow specific other types to access the private ref counter of
   // owners - specifically, we want to allow a matrix<view> to be made from a
@@ -520,8 +520,8 @@ public:
   // use subscript operators for general purpose access
   P *data(int const i = 0, int const j = 0) const
   {
-    // return &data_[i * stride() + j]; // row-major
-    return &data_[j * stride() + i]; // column-major
+    // return data_ + i * stride() + j; // row-major
+    return data_ + j * stride() + i; // column-major
   }
 
   //
@@ -944,7 +944,11 @@ fk::vector<P, mem, resrc>::vector(fk::matrix<P, omem, resrc> &source,
 {}
 
 template<typename P, mem_type mem, resource resrc>
+#ifdef __clang__
+fk::vector<P, mem, resrc>::~vector<P, mem, resrc>()
+#else
 fk::vector<P, mem, resrc>::~vector()
+#endif
 {
   if constexpr (mem == mem_type::owner)
   {
@@ -1372,9 +1376,9 @@ fk::vector<P> fk::vector<P, mem, resrc>::single_column_kron(
     vector<P, omem> const &right) const
 {
   fk::vector<P> product((*this).size() * right.size());
-  for (int i = 0; i < (*this).size(); ++i)
+  for (int j = 0; j < right.size(); ++j)
   {
-    for (int j = 0; j < right.size(); ++j)
+    for (int i = 0; i < (*this).size(); ++i)
     {
       product(i * right.size() + j) = (*this)(i)*right(j);
     }
@@ -1728,7 +1732,11 @@ fk::matrix<P, mem, resrc>::matrix(fk::vector<P, omem, resrc> &source,
 
 // destructor
 template<typename P, mem_type mem, resource resrc>
+#ifdef __clang__
+fk::matrix<P, mem, resrc>::~matrix<P, mem, resrc>()
+#else
 fk::matrix<P, mem, resrc>::~matrix()
+#endif
 {
   if constexpr (mem == mem_type::owner)
   {
@@ -2017,7 +2025,7 @@ fk::matrix<P, mem, resrc>::operator=(fk::vector<P, omem> const &v)
 //
 template<typename P, mem_type mem, resource resrc>
 template<mem_type, typename, resource, typename>
-P &fk::matrix<P, mem, resrc>::operator()(int const i, int const j)
+inline P &fk::matrix<P, mem, resrc>::operator()(int const i, int const j)
 {
   expect(i < nrows() && j < ncols());
   return *(data(i, j));
@@ -2025,7 +2033,7 @@ P &fk::matrix<P, mem, resrc>::operator()(int const i, int const j)
 
 template<typename P, mem_type mem, resource resrc>
 template<resource, typename>
-P fk::matrix<P, mem, resrc>::operator()(int const i, int const j) const
+inline P fk::matrix<P, mem, resrc>::operator()(int const i, int const j) const
 {
   expect(i < nrows() && j < ncols());
   return *(data(i, j));
@@ -2276,16 +2284,24 @@ template<mem_type omem, resource, typename>
 fk::matrix<P> fk::matrix<P, mem, resrc>::kron(matrix<P, omem> const &B) const
 {
   fk::matrix<P> C(nrows() * B.nrows(), ncols() * B.ncols());
-  for (auto i = 0; i < nrows(); ++i)
+
[…]
