### 2022-07-29 7e0c499
documented physics of 1d continuity equation (#426)
* documented physics of 1d continuity equation

* clang-format

Signed-off-by: Steven Hahn <hahnse@ornl.gov>

Co-authored-by: Steven Hahn <hahnse@ornl.gov>
