### 2022-07-29 4046c0a
fix bug with uninitialized data (#427)
* fix bug with uninitialized data

* add unit test

Co-authored-by: Steven Hahn <hahnse@ornl.gov>
