### 2022-07-31 #4378 @arfon
Just checking in on this review as it has been a while since there has been any movement. @joglekara – it looks like you were having some issues building the package on your (M1) architecture. Did you manage to get past this yet?

@cticenhour – it looks like you've generated your checklist but haven't gotten much further than that. Could you updates us here on when you think you might be able to complete your review by?

Many thanks!
