### 2022-07-31 #4378 @cticenhour
@arfon Thanks for checking in - I had started the process, but failed to check off boxes here as I went along. I recently worked through getting the basic installation working on my M1 architecture - identified a couple issues with an optional dependency that I'll be passing along to the development team likely today or tomorrow.

I think I'll be able to get moving on going through the documentation this week, with a paper review either the end of this week or next week. Hoping to finally get this initial review knocked out by the end of next week.
