### 2022-08-01 project-asgard/asgard#428 @cticenhour
JOSS Review: Summary of issues building ASGarD with HDF5 on M1
## Describe the bug
This is part of my JOSS review for openjournals/joss-reviews#4378.

The build process on M1 using conda package management has been pretty straightforward for a basic installation, except when attempting to use a pre-installed HDF5, or allowing ASGarD to build it itself. Usage of HighFive and HDF5 with ASGarD was not clear in my reading of the README or Wiki documentation, so I had to work a bit blind. I've outlined issues experienced with each case below that might need to be addressed via documentation or CMake adjustments:

1. Using `-DASGARD_HDF5_PATH` to point to a pre-installed HDF5 located at `~/mambaforge3/envs/asgard/`. Specifics of the packages installed are covered below under the System heading.

    - Starting from [CMakeLists.txt](https://github.com/project-asgard/asgard/blob/4046c0a264ca1d156be1ecf5f8a30d290a0a7720/CMakeLists.txt), the `ASGARD_HDF5_PATH` variable was initially used to point ASGarD to the conda prefix for HDF5. Using
       ```
       cmake .. -DCMAKE_CXX_COMPILER=mpicxx -DASGARD_USE_MPI=YES -DASGARD_HDF5_PATH=~/mambaforge3/envs/asgard/
       ```
       led to no hint of HDF5 in the CMake output. It appeared as though this wasn't the correct/primary variable to start with, so another attempt was made adding on a reference to HighFive.
    - Adding `-DASGARD_IO_HIGHFIVE=ON` to the prior CMake command led to the following console output:
        ```
        -- using external hdf5 found at ~/mambaforge3/envs/asgard/
        -- using contrib HighFive at /Users/cticenhour/projects/asgard/contrib/highfive
        ```
        which is what I expected to see given the references to HighFive and HDF5 in the documentation. Sure enough, `io-test` was added to the test listing after build. If both of these variables are required, should `ASGARD_IO_HIGHFIVE` be set to ON whenever `ADGARD_HDF5_PATH` is set using CMake dependent options?

2. Letting ASGarD build its own HDF5

    - Again, looking at `CMakeLists.txt`, it appears as though `ASGARD_BUILD_HDF5` is the primary variable of interest here. Removing the installed HDF5 from earlier, and using:
        ```
        cmake .. -DCMAKE_CXX_COMPILER=mpicxx -DASGARD_USE_MPI=YES -DASGARD_BUILD_HDF5=ON
        ```
        we again don't have any hint of HDF5 in the cmake output. Adding `-DASGARD_IO_HIGHFIVE=ON` yields the following successful console output:
        ```
        -- using external hdf5 found at HDF5_hdf5_LIBRARY-NOTFOUND
        -- libhdf5 not found. We'll build it from source.
        -- using contrib HighFive at /Users/cticenhour/projects/asgard/contrib/highfive
        ```
        Without looking at `contrib/io.cmake` and purely at the console, it appears as though `ASGARD_BUILD_HDF5` does nothing, so it should likely also be tied to `ASGARD_IO_HIGHFIVE` as a dependency whenever it is used.

    - When the build begins, HDF5 1.10.5 is downloaded and the build begins. This fails with the error:
        ```
        checking build system type... Invalid configuration `arm64-apple-darwin20.0.0': machine `arm64-apple' not recognized
        configure: error: /bin/sh bin/config.sub arm64-apple-darwin20.0.0 failed
        ```
        I've experienced this before when working with HDF5 in the MOOSE framework, in idaholab/moose#20736, which I had to fix by adding a pre-configure step to the build for M1 platforms. I made a patch below relevant to this case:
        ```
        diff --git a/contrib/io.cmake b/contrib/io.cmake
        index 22f5c3c4..91ad9c7a 100644
        --- a/contrib/io.cmake
        +++ b/contrib/io.cmake
        @@ -44,7 +44,8 @@ function (get_hdf5)
                 PREFIX contrib/hdf5
                 URL https://support.hdfgroup.org/ftp/HDF5/releases/hdf5-1.10/hdf5-1.10.5/src/hdf5-1.10.5.tar.bz2
                 DOWNLOAD_NO_PROGRESS 1
        -        CONFIGURE_COMMAND ${CMAKE_CURRENT_BINARY_DIR}/contrib/hdf5/src/hdf5-ext/configure --prefix=${hdf5_contrib_path}
        +        CONFIGURE_COMMAND ${CMAKE_CURRENT_BINARY_DIR}/contrib/hdf5/src/hdf5-ext/autogen.sh
        +        COMMAND ${CMAKE_CURRENT_BINARY_DIR}/contrib/hdf5/src/hdf5-ext/configure --prefix=${hdf5_contrib_path}
                 BUILD_COMMAND make
                 BUILD_IN_SOURCE 1
                 INSTALL_COMMAND make install
        ```
        which allows HDF5 to reconfigure itself using `automake` and `libtool`. For Darwin ARM64 platforms using homebrew, this change should be adequate as `autogen.sh` looks for the location of `glibtool` specifically (which homebrew provides), but for conda an extra step needs to be taken. One either needs to specify the location of `libtool` using the `HDF5_LIBTOOL` variable, or create a link to `libtool` named `glibtool` for the script to pick up. Either way, HDF5 should then build successfully and pass the added `io-test`.

## To Reproduce

CMake specifics for each case are outlined above, but common attributes for the ASGarD code as cloned on my system for each are outlined below:
1. Using the `develop` branch, commit 4046c0a264ca1d156be1ecf5f8a30d290a0a7720
2. Make command: `make -j10`
4. Test invocation: `ctest -j10`

## Expected behavior
All required tests for a configuration including HDF5 (with MPI in this case) should pass successfully.

## System
 - `MacBookPro18,2`: Apple M1 Max, 64GB RAM, macOS Monterey 12.5, Command Line Tools 13.4.0.0.1.1651278267
 - Using conda package management (using packages from `conda-forge`) with the following install command:
    ```
    conda install openmpi=4.1.4 openmpi-mpicc=4.1.4 openmpi-mpicxx=4.1.4 cmake=3.23.3 hdf5=1.12.2=mpi_openmpi_hfe3cbcd_0
    ```
    - This package configuration uses Clang 13.0.1
    - The HDF5 package was removed during HDF5 build testing.
 - When allowing ASGarD to build its own HDF5, more packages were needed for a pre-configuration step:
     ```
     conda install automake=2.69 libtool=2.4.6
     ```
     - This led to:
         - Autoconf 2.69
         - Automake 1.16.5
         - Libtool 2.4.6
         - M4 1.4.18

## Additional context
I still plan to test this workflow on an Intel Mac and an AMD based Linux environment with a similar configuration (will note differences here with an edit).
