### 2022-08-05 e80c852
HDF5 CMake improvements
Improve cmake formula for highfive dependency. Check additional
variables for setting ASGARD_IO_HIGHFIVE. Fix for building HDF5
on M1 Mac.

Signed-off-by: Steven Hahn <hahnse@ornl.gov>
