### 2022-08-06 #4378 @joglekara
@arfon , thanks for checking in. Yes I'm following up with them re: M1. @cticenhour provided a nice report on the conda build process. I'm working through the Accelerate + Homebrew build.
