### 2022-08-06 project-asgard/asgard#429 @joglekara
JOSS Review: Example Usage
It would be good to include a few examples of running `asgard`. I see that there are example PDEs outlined at https://github.com/project-asgard/asgard/wiki/custom-pdes but I haven't been able to find more documentation.

This is the requirement outlined under `Documentation` in the JOSS review:
```
Example usage:

Do the authors include examples of how to use the software (ideally to solve real-world analysis problems).
```
