### 2022-08-06 project-asgard/asgard#430 @joglekara
JOSS Review: Summary and Statement of Need
The JOSS review requires a `statement of need` section [1]. In my opinion, the first paragraph is a good version of that.

It would be good to write a shorter `Summary` that is closer to a very high-level and generalized abstract, and then a `Statement of Need`. However, please feel free to be inspired by some of the accepted JOSS papers at
https://joss.theoj.org/papers/published

[1] :
```
A statement of need:

Does the paper have a section titled 'Statement of Need' that clearly states what problems the software is designed to solve and who the target audience is?
```
