### 2022-08-06 project-asgard/asgard#425 @cianciosa
In your specific case you'll need to append `/opt/homebrew/Cellar/openblas/` to the `CMAKE_PREFIX_PATH` variable.

As for accelerate it's working fine on my M1 Machine.

```
m4c@MAC130884 asgard % cd build2
m4c@MAC130884 build2 % cmake ../
-- The CXX compiler identification is AppleClang 13.1.6.13160021
-- Detecting CXX compiler ABI info
-- Check for working CXX compiler: /Applications/Xcode.app/Contents/Developer/Toolchains/XcodeDefault.xctoolchain/usr/bin/c++ - skipped
-- Detecting CXX compile features
-- Found Git: /usr/bin/git (found version "2.32.1 (Apple Git-133)")
-- Registering project kronmult
-- Could NOT find OpenMP_CXX (missing: OpenMP_CXX_FLAGS OpenMP_CXX_LIB_NAMES)
-- Could NOT find OpenMP (missing: OpenMP_CXX_FOUND)
-- Registering project Catch2
-- Setting build type to 'Debug' as none was specified.
-- Looking for sgemm_
-- Looking for sgemm_ - not found
-- Looking for C++ include pthread.h
-- Looking for C++ include pthread.h - found
-- Performing Test CMAKE_HAVE_LIBC_PTHREAD
-- Performing Test CMAKE_HAVE_LIBC_PTHREAD - Success
-- Found Threads: TRUE
-- Looking for dgemm_
-- Looking for dgemm_ - found
-- Found BLAS: /Applications/Xcode.app/Contents/Developer/Platforms/MacOSX.platform/Developer/SDKs/MacOSX12.3.sdk/System/Library/Frameworks/Accelerate.framework
-- Looking for cheev_
-- Looking for cheev_ - found
-- Found LAPACK: /Applications/Xcode.app/Contents/Developer/Platforms/MacOSX.platform/Developer/SDKs/MacOSX12.3.sdk/System/Library/Frameworks/Accelerate.framework;-lm;-ldl
-- Found LINALG: TRUE
-- Could NOT find OpenMP_CXX (missing: OpenMP_CXX_FLAGS OpenMP_CXX_LIB_NAMES)
-- Could NOT find OpenMP (missing: OpenMP_CXX_FOUND)
-- Configuring done
-- Generating done
-- Build files have been written to: /Users/m4c/Projects/asgard/build2
m4c@MAC130884 build2 % make VERBOSE=1
/Applications/CMake.app/Contents/bin/cmake -S/Users/m4c/Projects/asgard -B/Users/m4c/Projects/asgard/build2 --check-build-system CMakeFiles/Makefile.cmake 0
/Applications/CMake.app/Contents/bin/cmake -E cmake_progress_start /Users/m4c/Projects/asgard/build2/CMakeFiles /Users/m4c/Projects/asgard/build2//CMakeFiles/progress.marks
/Applications/Xcode.app/Contents/Developer/usr/bin/make  -f CMakeFiles/Makefile2 all
/Applications/Xcode.app/Contents/Developer/usr/bin/make  -f _deps/kronmult-build/CMakeFiles/kron.dir/build.make _deps/kronmult-build/CMakeFiles/kron.dir/depend
cd /Users/m4c/Projects/asgard/build2 && /Applications/CMake.app/Contents/bin/cmake -E cmake_depends "Unix Makefiles" /Users/m4c/Projects/asgard /Users/m4c/Projects/asgard/contrib/KRONMULT /Users/m4c/Projects/asgard/build2 /Users/m4c/Projects/asgard/build2/_deps/kronmult-build /Users/m4c/Projects/asgard/build2/_deps/kronmult-build/CMakeFiles/kron.dir/DependInfo.cmake --color=
/Applications/Xcode.app/Contents/Developer/usr/bin/make  -f _deps/kronmult-build/CMakeFiles/kron.dir/build.make _deps/kronmult-build/CMakeFiles/kron.dir/build
[  1%] Building CXX object _deps/kronmult-build/CMakeFiles/kron.dir/kgemm_nn_batched.cpp.o
cd /Users/m4c/Projects/asgard/build2/_deps/kronmult-build && /Applications/Xcode.app/Contents/Developer/Toolchains/XcodeDefault.xctoolchain/usr/bin/c++ -Dkron_EXPORTS -I/Users/m4c/Projects/asgard/contrib/KRONMULT -g -arch arm64 -isysroot /Applications/Xcode.app/Contents/Developer/Platforms/MacOSX.platform/Developer/SDKs/MacOSX12.3.sdk -fPIC -Wall -Wextra -Wpedantic -std=gnu++17 -MD -MT _deps/kronmult-build/CMakeFiles/kron.dir/kgemm_nn_batched.cpp.o -MF CMakeFiles/kron.dir/kgemm_nn_batched.cpp.o.d -o CMakeFiles/kron.dir/kgemm_nn_batched.cpp.o -c /Users/m4c/Projects/asgard/contrib/KRONMULT/kgemm_nn_batched.cpp
[  1%] Building CXX object _deps/kronmult-build/CMakeFiles/kron.dir/kgemm_nn.cpp.o
cd /Users/m4c/Projects/asgard/build2/_deps/kronmult-build && /Applications/Xcode.app/Contents/Developer/Toolchains/XcodeDefault.xctoolchain/usr/bin/c++ -Dkron_EXPORTS -I/Users/m4c/Projects/asgard/contrib/KRONMULT -g -arch arm64 -isysroot /Applications/Xcode.app/Contents/Developer/Platforms/MacOSX.platform/Developer/SDKs/MacOSX12.3.sdk -fPIC -Wall -Wextra -Wpedantic -std=gnu++17 -MD -MT _deps/kronmult-build/CMakeFiles/kron.dir/kgemm_nn.cpp.o -MF CMakeFiles/kron.dir/kgemm_nn.cpp.o.d -o CMakeFiles/kron.dir/kgemm_nn.cpp.o -c /Users/m4c/Projects/asgard/contrib/KRONMULT/kgemm_nn.cpp
[  2%] Building CXX object _deps/kronmult-build/CMakeFiles/kron.dir/kgemm_nt_batched.cpp.o
cd /Users/m4c/Projects/asgard/build2/_deps/kronmult-build && /Applications/Xcode.app/Contents/Developer/Toolchains/XcodeDefault.xctoolchain/usr/bin/c++ -Dkron_EXPORTS -I/Users/m4c/Projects/asgard/contrib/KRONMULT -g -arch arm64 -isysroot /Applications/Xcode.app/Contents/Developer/Platforms/MacOSX.platform/Developer/SDKs/MacOSX12.3.sdk -fPIC -Wall -Wextra -Wpedantic -std=gnu++17 -MD -MT _deps/kronmult-build/CMakeFiles/kron.dir/kgemm_nt_batched.cpp.o -MF CMakeFiles/kron.dir/kgemm_nt_batched.cpp.o.d -o CMakeFiles/kron.dir/kgemm_nt_batched.cpp.o -c /Users/m4c/Projects/asgard/contrib/KRONMULT/kgemm_nt_batched.cpp
[  2%] Building CXX object _deps/kronmult-build/CMakeFiles/kron.dir/kgemm_nt.cpp.o
cd /Users/m4c/Projects/asgard/build2/_deps/kronmult-build && /Applications/Xcode.app/Contents/Developer/Toolchains/XcodeDefault.xctoolchain/usr/bin/c++ -Dkron_EXPORTS -I/Users/m4c/Projects/asgard/contrib/KRONMULT -g -arch arm64 -isysroot /Applications/Xcode.app/Contents/Developer/Platforms/MacOSX.platform/Developer/SDKs/MacOSX12.3.sdk -fPIC -Wall -Wextra -Wpedantic -std=gnu++17 -MD -MT _deps/kronmult-build/CMakeFiles/kron.dir/kgemm_nt.cpp.o -MF CMakeFiles/kron.dir/kgemm_nt.cpp.o.d -o CMakeFiles/kron.dir/kgemm_nt.cpp.o -c /Users/m4c/Projects/asgard/contrib/KRONMULT/kgemm_nt.cpp
[  3%] Building CXX object _deps/kronmult-build/CMakeFiles/kron.dir/kronmult1_batched.cpp.o
cd /Users/m4c/Projects/asgard/build2/_deps/kronmult-build && /Applications/Xcode.app/Contents/Developer/Toolchains/XcodeDefault.xctoolchain/usr/bin/c++ -Dkron_EXPORTS -I/Users/m4c/Projects/asgard/contrib/KRONMULT -g -arch arm64 -isysroot /Applications/Xcode.app/Contents/Developer/Platforms/MacOSX.platform/Developer/SDKs/MacOSX12.3.sdk -fPIC -Wall -Wextra -Wpedantic -std=gnu++17 -MD -MT _deps/kronmult-build/CMakeFiles/kron.dir/kronmult1_batched.cpp.o -MF CMakeFiles/kron.dir/kronmult1_batched.cpp.o.d -o CMakeFiles/kron.dir/kronmult1_batched.cpp.o -c /Users/m4c/Projects/asgard/contrib/KRONMULT/kronmult1_batched.cpp
[  3%] Building CXX object _deps/kronmult-build/CMakeFiles/kron.dir/kronmult1_pbatched.cpp.o
cd /Users/m4c/Projects/asgard/build2/_deps/kronmult-build && /Applications/Xcode.app/Contents/Developer/Toolchains/XcodeDefault.xctoolchain/usr/bin/c++ -Dkron_EXPORTS -I/Users/m4c/Projects/asgard/contrib/KRONMULT -g -arch arm64 -isysroot /Applications/Xcode.app/Contents/Developer/Platforms/MacOSX.platform/Developer/SDKs/MacOSX12.3.sdk -fPIC -Wall -Wextra -Wpedantic -std=gnu++17 -MD -MT _deps/kronmult-build/CMakeFiles/kron.dir/kronmult1_pbatched.cpp.o -MF CMakeFiles/kron.dir/kronmult1_pbatched.cpp.o.d -o CMakeFiles/kron.dir/kronmult1_pbatched.cpp.o -c /Users/m4c/Projects/asgard/contrib/KRONMULT/kronmult1_pbatched.cpp
[  4%] Building CXX object _deps/kronmult-build/CMakeFiles/kron.dir/kronmult1_xbatched.cpp.o
cd /Users/m4c/Projects/asgard/build2/_deps/kronmult-build && /Applications/Xcode.app/Contents/Developer/Toolchains/XcodeDefault.xctoolchain/usr/bin/c++ -Dkron_EXPORTS -I/Users/m4c/Projects/asgard/contrib/KRONMULT -g -arch arm64 -isysroot /Applications/Xcode.app/Contents/Developer/Platforms/MacOSX.platform/Developer/SDKs/MacOSX12.3.sdk -fPIC -Wall -Wextra -Wpedantic -std=gnu++17 -MD -MT _deps/kronmult-build/CMakeFiles/kron.dir/kronmult1_xbatched.cpp.o -MF CMakeFiles/kron.dir/kronmult1_xbatched.cpp.o.d -o CMakeFiles/kron.dir/kronmult1_xbatched.cpp.o -c /Users/m4c/Projects/asgard/contrib/KRONMULT/kronmult1_xbatched.cpp
[  4%] Building CXX object _deps/kronmult-build/CMakeFiles/kron.dir/kronmult2_batched.cpp.o
cd /Users/m4c/Projects/asgard/build2/_deps/kronmult-build && /Applications/Xcode.app/Contents/Developer/Toolchains/XcodeDefault.xctoolchain/usr/bin/c++ -Dkron_EXPORTS -I/Users/m4c/Projects/asgard/contrib/KRONMULT -g -arch arm64 -isysroot /Applications/Xcode.app/Contents/Developer/Platforms/MacOSX.platform/Developer/SDKs/MacOSX12.3.sdk -fPIC -Wall -Wextra -Wpedantic -std=gnu++17 -MD -MT _deps/kronmult-build/CMakeFiles/kron.dir/kronmult2_batched.cpp.o -MF CMakeFiles/kron.dir/kronmult2_batched.cpp.o.d -o CMakeFiles/kron.dir/kronmult2_batched.cpp.o -c /Users/m4c/Projects/asgard/contrib/KRONMULT/kronmult2_batched.cpp
[  5%] Building CXX object _deps/kronmult-build/CMakeFiles/kron.dir/kronmult2_pbatched.cpp.o
cd /Users/m4c/Projects/asgard/build2/_deps/kronmult-build && /Applications/Xcode.app/Contents/Developer/Toolchains/XcodeDefault.xctoolchain/usr/bin/c++ -Dkron_EXPORTS -I/Users/m4c/Projects/asgard/contrib/KRONMULT -g -arch arm64 -isysroot /Applications/Xcode.app/Contents/Developer/Platforms/MacOSX.platform/Developer/SDKs/MacOSX12.3.sdk -fPIC -Wall -Wextra -Wpedantic -std=gnu++17 -MD -MT _deps/kronmult-build/CMakeFiles/kron.dir/kronmult2_pbatched.cpp.o -MF CMakeFiles/kron.dir/kronmult2_pbatched.cpp.o.d -o CMakeFiles/kron.dir/kronmult2_pbatched.cpp.o -c /Users/m4c/Projects/asgard/contrib/KRONMULT/kronmult2_pbatched.cpp
[  6%] Building CXX object _deps/kronmult-build/CMakeFiles/kron.dir/kronmult2_xbatched.cpp.o
cd /Users/m4c/Projects/asgard/build2/_deps/kronmult-build && /Applications/Xcode.app/Contents/Developer/Toolchains/XcodeDefault.xctoolchain/usr/bin/c++ -Dkron_EXPORTS -I/Users/m4c/Projects/asgard/contrib/KRONMULT -g -arch arm64 -isysroot /Applications/Xcode.app/Contents/Developer/Platforms/MacOSX.platform/Developer/SDKs/MacOSX12.3.sdk -fPIC -Wall -Wextra -Wpedantic -std=gnu++17 -MD -MT _deps/kronmult-build/CMakeFiles/kron.dir/kronmult2_xbatched.cpp.o -MF CMakeFiles/kron.dir/kronmult2_xbatched.cpp.o.d -o CMakeFiles/kron.dir/kronmult2_xbatched.cpp.o -c /Users/m4c/Projects/asgard/contrib/KRONMULT/kronmult2_xbatched.cpp
[  6%] Building CXX object _deps/kronmult-build/CMakeFiles/kron.dir/kronmult3_batched.cpp.o
cd /Users/m4c/Projects/asgard/build2/_deps/kronmult-build && /Applications/Xcode.app/Contents/Developer/Toolchains/XcodeDefault.xctoolchain/usr/bin/c++ -Dkron_EXPORTS -I/Users/m4c/Projects/asgard/contrib/KRONMULT -g -arch arm64 -isysroot /Applications/Xcode.app/Contents/Developer/Platforms/MacOSX.platform/Developer/SDKs/MacOSX12.3.sdk -fPIC -Wall -Wextra -Wpedantic -std=gnu++17 -MD -MT _deps/kronmult-build/CMakeFiles/kron.dir/kronmult3_batched.cpp.o -MF CMakeFiles/kron.dir/kronmult3_batched.cpp.o.d -o CMakeFiles/kron.dir/kronmult3_batched.cpp.o -c /Users/m4c/Projects/asgard/contrib/KRONMULT/kronmult3_batched.cpp
[  7%] Building CXX object _deps/kronmult-build/CMakeFiles/kron.dir/kronmult3_pbatched.cpp.o
cd /Users/m4c/Projects/asgard/build2/_deps/kronmult-build && /Applications/Xcode.app/Contents/Developer/Toolchains/XcodeDefault.xctoolchain/usr/bin/c++ -Dkron_EXPORTS -I/Users/m4c/Projects/asgard/contrib/KRONMULT -g -arch arm64 -isysroot /Applications/Xcode.app/Contents/Developer/Platforms/MacOSX.platform/Developer/SDKs/MacOSX12.3.sdk -fPIC -Wall -Wextra -Wpedantic -std=gnu++17 -MD -MT _deps/kronmult-build/CMakeFiles/kron.dir/kronmult3_pbatched.cpp.o -MF CMakeFiles/kron.dir/kronmult3_pbatched.cpp.o.d -o CMakeFiles/kron.dir/kronmult3_pbatched.cpp.o -c /Users/m4c/Projects/asgard/contrib/KRONMULT/kronmult3_pbatched.cpp
[  7%] Building CXX object _deps/kronmult-build/CMakeFiles/kron.dir/kronmult3_xbatched.cpp.o
cd /Users/m4c/Projects/asgard/build2/_deps/kronmult-build && /Applications/Xcode.app/Contents/Developer/Toolchains/XcodeDefault.xctoolchain/usr/bin/c++ -Dkron_EXPORTS -I/Users/m4c/Projects/asgard/contrib/KRONMULT -g -arch arm64 -isysroot /Applications/Xcode.app/Contents/Developer/Platforms/MacOSX.platform/Developer/SDKs/MacOSX12.3.sdk -fPIC -Wall -Wextra -Wpedantic -std=gnu++17 -MD -MT _deps/kronmult-build/CMakeFiles/kron.dir/kronmult3_xbatched.cpp.o -MF CMakeFiles/kron.dir/kronmult3_xbatched.cpp.o.d -o CMakeFiles/kron.dir/kronmult3_xbatched.cpp.o -c /Users/m4c/Projects/asgard/contrib/KRONMULT/kronmult3_xbatched.cpp
[  8%] Building CXX object _deps/kronmult-build/CMakeFiles/kron.dir/kronmult4_batched.cpp.o
cd /Users/m4c/Projects/asgard/build2/_deps/kronmult-build && /Applications/Xcode.app/Contents/Developer/Toolchains/XcodeDefault.xctoolchain/usr/bin/c++ -Dkron_EXPORTS -I/Users/m4c/Projects/asgard/contrib/KRONMULT -g -arch arm64 -isysroot /Applications/Xcode.app/Contents/Developer/Platforms/MacOSX.platform/Developer/SDKs/MacOSX12.3.sdk -fPIC -Wall -Wextra -Wpedantic -std=gnu++17 -MD -MT _deps/kronmult-build/CMakeFiles/kron.dir/kronmult4_batched.cpp.o -MF CMakeFiles/kron.dir/kronmult4_batched.cpp.o.d -o CMakeFiles/kron.dir/kronmult4_batched.cpp.o -c /Users/m4c/Projects/asgard/contrib/KRONMULT/kronmult4_batched.cpp
[  8%] Building CXX object _deps/kronmult-build/CMakeFiles/kron.dir/kronmult4_pbatched.cpp.o
cd /Users/m4c/Projects/asgard/build2/_deps/kronmult-build && /Applications/Xcode.app/Contents/Developer/Toolchains/XcodeDefault.xctoolchain/usr/bin/c++ -Dkron_EXPORTS -I/Users/m4c/Projects/asgard/contrib/KRONMULT -g -arch arm64 -isysroot /Applications/Xcode.app/Contents/Developer/Platforms/MacOSX.platform/Developer/SDKs/MacOSX12.3.sdk -fPIC -Wall -Wextra -Wpedantic -std=gnu++17 -MD -MT _deps/kronmult-build/CMakeFiles/kron.dir/kronmult4_pbatched.cpp.o -MF CMakeFiles/kron.dir/kronmult4_pbatched.cpp.o.d -o CMakeFiles/kron.dir/kronmult4_pbatched.cpp.o -c /Users/m4c/Projects/asgard/contrib/KRONMULT/kronmult4_pbatched.cpp
[  9%] Building CXX object _deps/kronmult-build/CMakeFiles/kron.dir/kronmult4_xbatched.cpp.o
cd /Users/m4c/Projects/asgard/build2/_deps/kronmult-build && /Applications/Xcode.app/Contents/Developer/Toolchains/XcodeDefault.xctoolchain/usr/bin/c++ -Dkron_EXPORTS -I/Users/m4c/Projects/asgard/contrib/KRONMULT -g -arch arm64 -isysroot /Applications/Xcode.app/Contents/Developer/Platforms/MacOSX.platform/Developer/SDKs/MacOSX12.3.sdk -fPIC -Wall -Wextra -Wpedantic -std=gnu++17 -MD -MT _deps/kronmult-build/CMakeFiles/kron.dir/kronmult4_xbatched.cpp.o -MF CMakeFiles/kron.dir/kronmult4_xbatched.cpp.o.d -o CMakeFiles/kron.dir/kronmult4_xbatched.cpp.o -c /Users/m4c/Projects/asgard/contrib/KRONMULT/kronmult4_xbatched.cpp
[  9%] Building CXX object _deps/kronmult-build/CMakeFiles/kron.dir/kronmult5_batched.cpp.o
cd /Users/m4c/Projects/asgard/build2/_deps/kronmult-build && /Applications/Xcode.app/Contents/Developer/Toolchains/XcodeDefault.xctoolchain/usr/bin/c++ -Dkron_EXPORTS -I/Users/m4c/Projects/asgard/contrib/KRONMULT -g -arch arm64 -isysroot /Applications/Xcode.app/Contents/Developer/Platforms/MacOSX.platform/Developer/SDKs/MacOSX12.3.sdk -fPIC -Wall -Wextra -Wpedantic -std=gnu++17 -MD -MT _deps/kronmult-build/CMakeFiles/kron.dir/kronmult5_batched.cpp.o -MF CMakeFiles/kron.dir/kronmult5_batched.cpp.o.d -o CMakeFiles/kron.dir/kronmult5_batched.cpp.o -c /Users/m4c/Projects/asgard/contrib/KRONMULT/kronmult5_batched.cpp
[ 10%] Building CXX object _deps/kronmult-build/CMakeFiles/kron.dir/kronmult5_pbatched.cpp.o
cd /Users/m4c/Projects/asgard/build2/_deps/kronmult-build && /Applications/Xcode.app/Contents/Developer/Toolchains/XcodeDefault.xctoolchain/usr/bin/c++ -Dkron_EXPORTS -I/Users/m4c/Projects/asgard/contrib/KRONMULT -g -arch arm64 -isysroot /Applications/Xcode.app/Contents/Developer/Platforms/MacOSX.platform/Developer/SDKs/MacOSX12.3.sdk -fPIC -Wall -Wextra -Wpedantic -std=gnu++17 -MD -MT _deps/kronmult-build/CMakeFiles/kron.dir/kronmult5_pbatched.cpp.o -MF CMakeFiles/kron.dir/kronmult5_pbatched.cpp.o.d -o CMakeFiles/kron.dir/kronmult5_pbatched.cpp.o -c /Users/m4c/Projects/asgard/contrib/KRONMULT/kronmult5_pbatched.cpp
[ 11%] Building CXX object _deps/kronmult-build/CMakeFiles/kron.dir/kronmult5_xbatched.cpp.o
cd /Users/m4c/Projects/asgard/build2/_deps/kronmult-build && /Applications/Xcode.app/Contents/Developer/Toolchains/XcodeDefault.xctoolchain/usr/bin/c++ -Dkron_EXPORTS -I/Users/m4c/Projects/asgard/contrib/KRONMULT -g -arch arm64 -isysroot /Applications/Xcode.app/Contents/Developer/Platforms/MacOSX.platform/Developer/SDKs/MacOSX12.3.sdk -fPIC -Wall -Wextra -Wpedantic -std=gnu++17 -MD -MT _deps/kronmult-build/CMakeFiles/kron.dir/kronmult5_xbatched.cpp.o -MF CMakeFiles/kron.dir/kronmult5_xbatched.cpp.o.d -o CMakeFiles/kron.dir/kronmult5_xbatched.cpp.o -c /Users/m4c/Projects/asgard/contrib/KRONMULT/kronmult5_xbatched.cpp
[ 11%] Building CXX object _deps/kronmult-build/CMakeFiles/kron.dir/kronmult6_batched.cpp.o
cd /Users/m4c/Projects/asgard/build2/_deps/kronmult-build && /Applications/Xcode.app/Contents/Developer/Toolchains/XcodeDefault.xctoolchain/usr/bin/c++ -Dkron_EXPORTS -I/Users/m4c/Projects/asgard/contrib/KRONMULT -g -arch arm64 -isysroot /Applications/Xcode.app/Contents/Developer/Platforms/MacOSX.platform/Developer/SDKs/MacOSX12.3.sdk -fPIC -Wall -Wextra -Wpedantic -std=gnu++17 -MD -MT _deps/kronmult-build/CMakeFiles/kron.dir/kronmult6_batched.cpp.o -MF CMakeFiles/kron.dir/kronmult6_batched.cpp.o.d -o CMakeFiles/kron.dir/kronmult6_batched.cpp.o -c /Users/m4c/Projects/asgard/contrib/KRONMULT/kronmult6_batched.cpp
[ 12%] Building CXX object _deps/kronmult-build/CMakeFiles/kron.dir/kronmult6_pbatched.cpp.o
cd /Users/m4c/Projects/asgard/build2/_deps/kronmult-build && /Applications/Xcode.app/Contents/Developer/Toolchains/XcodeDefault.xctoolchain/usr/bin/c++ -Dkron_EXPORTS -I/Users/m4c/Projects/asgard/contrib/KRONMULT -g -arch arm64 -isysroot /Applications/Xcode.app/Contents/Developer/Platforms/MacOSX.platform/Developer/SDKs/MacOSX12.3.sdk -fPIC -Wall -Wextra -Wpedantic -std=gnu++17 -MD -MT _deps/kronmult-build/CMakeFiles/kron.dir/kronmult6_pbatched.cpp.o -MF CMakeFiles/kron.dir/kronmult6_pbatched.cpp.o.d -o CMakeFiles/kron.dir/kronmult6_pbatched.cpp.o -c /Users/m4c/Projects/asgard/contrib/KRONMULT/kronmult6_pbatched.cpp
[ 12%] Building CXX object _deps/kronmult-build/CMakeFiles/kron.dir/kronmult6_xbatched.cpp.o
cd /Users/m4c/Projects/asgard/build2/_deps/kronmult-build && /Applications/Xcode.app/Contents/Developer/Toolchains/XcodeDefault.xctoolchain/usr/bin/c++ -Dkron_EXPORTS -I/Users/m4c/Projects/asgard/contrib/KRONMULT -g -arch arm64 -isysroot /Applications/Xcode.app/Contents/Developer/Platforms/MacOSX.platform/Developer/SDKs/MacOSX12.3.sdk -fPIC -Wall -Wextra -Wpedantic -std=gnu++17 -MD -MT _deps/kronmult-build/CMakeFiles/kron.dir/kronmult6_xbatched.cpp.o -MF CMakeFiles/kron.dir/kronmult6_xbatched.cpp.o.d -o CMakeFiles/kron.dir/kronmult6_xbatched.cpp.o -c /Users/m4c/Projects/asgard/contrib/KRONMULT/kronmult6_xbatched.cpp
[ 13%] Linking CXX shared library libkron.dylib
cd /Users/m4c/Projects/asgard/build2/_deps/kronmult-build && /Applications/CMake.app/Contents/bin/cmake -E cmake_link_script CMakeFiles/kron.dir/link.txt --verbose=1
[…]
