+++ contrib/io.cmake
@@ -32,42 +32,49 @@ function (get_hdf5)
 
   # if cmake couldn't find other hdf5, or the user asked to build it
   if (NOT HDF5_FOUND)
-    set (hdf5_contrib_path ${CMAKE_SOURCE_DIR}/contrib/HDF5)
+    set (hdf5_contrib_path ${CMAKE_SOURCE_DIR}/contrib/hdf5)
     find_library (hdf5_search hdf5
       PATHS ${hdf5_contrib_path} PATH_SUFFIXES lib lib64 NO_DEFAULT_PATH)
     if (NOT hdf5_search)
       message (STATUS "libhdf5 not found. We'll build it from source.")
-      set( HDF5_TEST_CPP OFF CACHE BOOL "" FORCE)
-      set( HDF5_TEST_EXAMPLES OFF CACHE BOOL "" FORCE)
-      set( HDF5_TEST_FORTRAN OFF CACHE BOOL "" FORCE)
-      set( HDF5_TEST_JAVA OFF CACHE BOOL "" FORCE)
-      set( HDF5_TEST_PARALLEL OFF CACHE BOOL "" FORCE)
-      set( HDF5_TEST_PASSTHROUGH_VOL OFF CACHE BOOL "" FORCE)
-      set( HDF5_TEST_SERIAL OFF CACHE BOOL "" FORCE)
-      set( HDF5_TEST_SWMR OFF CACHE BOOL "" FORCE)
-      set( HDF5_TEST_TOOLS OFF CACHE BOOL "" FORCE)
-      set( HDF5_TEST_VFD OFF CACHE BOOL "" FORCE)
-      register_project(hdf5 HDF5 https://github.com/HDFGroup/hdf5 hdf5-1_12_2)
-      unset(CMAKE_RUNTIME_OUTPUT_DIRECTORY CACHE)
-      #Mark CATCH variables as advanced.
-      get_cmake_property(_variableNames VARIABLES)
-      foreach (_variableName ${_variableNames})
-	string(FIND "${_variableName}" "HDF5_" out)
-        if("${out}" EQUAL 0)
-          mark_as_advanced(${_variableName})
-        endif()
-      endforeach()
+
+      include (ExternalProject)
+      if (DEFINED CMAKE_APPLE_SILICON_PROCESSOR and CMAKE_APPLE_SILICON_PROCESSOR STREQUAL "arm64")
+	# Get HDF5 to build on Apple silicon      
+        ExternalProject_Add (hdf5-ext
+          UPDATE_COMMAND ""
+          PREFIX contrib/hdf5
+          URL https://support.hdfgroup.org/ftp/HDF5/releases/hdf5-1.10/hdf5-1.10.5/src/hdf5-1.10.5.tar.bz2
+          DOWNLOAD_NO_PROGRESS 1
+	  CONFIGURE_COMMAND ${CMAKE_CURRENT_BINARY_DIR}/contrib/hdf5/src/hdf5-ext/autogen.sh
+          COMMAND ${CMAKE_CURRENT_BINARY_DIR}/contrib/hdf5/src/hdf5-ext/configure --prefix=${hdf5_contrib_path}
+          BUILD_COMMAND make
+          BUILD_IN_SOURCE 1
+          INSTALL_COMMAND make install
+        )
+      else()
+	ExternalProject_Add (hdf5-ext
+          UPDATE_COMMAND ""
+          PREFIX contrib/hdf5
+          URL https://support.hdfgroup.org/ftp/HDF5/releases/hdf5-1.10/hdf5-1.10.5/src/hdf5-1.10.5.tar.bz2
+          DOWNLOAD_NO_PROGRESS 1
[…]
