### 2022-08-11 047a9bf
cmake formatting
Signed-off-by: Steven Hahn <hahnse@ornl.gov>
