+++ contrib/io.cmake
@@ -40,13 +40,13 @@ function (get_hdf5)
 
       include (ExternalProject)
       if (DEFINED CMAKE_APPLE_SILICON_PROCESSOR and CMAKE_APPLE_SILICON_PROCESSOR STREQUAL "arm64")
-	# Get HDF5 to build on Apple silicon      
+        # Get HDF5 to build on Apple silicon
         ExternalProject_Add (hdf5-ext
           UPDATE_COMMAND ""
           PREFIX contrib/hdf5
           URL https://support.hdfgroup.org/ftp/HDF5/releases/hdf5-1.10/hdf5-1.10.5/src/hdf5-1.10.5.tar.bz2
           DOWNLOAD_NO_PROGRESS 1
-	  CONFIGURE_COMMAND ${CMAKE_CURRENT_BINARY_DIR}/contrib/hdf5/src/hdf5-ext/autogen.sh
+          CONFIGURE_COMMAND ${CMAKE_CURRENT_BINARY_DIR}/contrib/hdf5/src/hdf5-ext/autogen.sh
           COMMAND ${CMAKE_CURRENT_BINARY_DIR}/contrib/hdf5/src/hdf5-ext/configure --prefix=${hdf5_contrib_path}
           BUILD_COMMAND make
           BUILD_IN_SOURCE 1
