### 2022-08-23 e7f9135
Improve build and documentation associated with ASGARD_IO_HIGHFIVE (#431)
* HDF5 CMake improvements

Improve cmake formula for highfive dependency. Check additional
variables for setting ASGARD_IO_HIGHFIVE. Fix for building HDF5
on M1 Mac.

Signed-off-by: Steven Hahn <hahnse@ornl.gov>

* Build HDF5 with CMake

Signed-off-by: Steven Hahn <hahnse@ornl.gov>

* Go back to a working build

Signed-off-by: Steven Hahn <hahnse@ornl.gov>

* cmake formatting

Signed-off-by: Steven Hahn <hahnse@ornl.gov>

* fix spacing

Signed-off-by: Steven Hahn <hahnse@ornl.gov>

Signed-off-by: Steven Hahn <hahnse@ornl
[…]
