### 2022-09-18 #4378 @joglekara
Thanks for checking in. I opened a couple of issues asking for example usage and a summary / statement of need and forgot to cross-link them.

https://github.com/project-asgard/asgard/issues/429
https://github.com/project-asgard/asgard/issues/430

I will follow up on those now
