### 2022-09-21 bdf4f32
added doxygen to the build system
