### 2022-09-23 7602973
removed unused cmake variable
