### 2022-09-23 da12d86
Add sandbox file (#436)
* removed unused cmake variable

* useful file
