+++ testing/sandbox.cpp
@@ -1,0 +1,36 @@
+#include "batch.hpp"
+
+#include "build_info.hpp"
+#include "coefficients.hpp"
+#include "distribution.hpp"
+#include "elements.hpp"
+#include "tools.hpp"
+
+#ifdef ASGARD_IO_HIGHFIVE
+#include "io.hpp"
+#endif
+
+#ifdef ASGARD_USE_MPI
+#include <mpi.h>
+#endif
+
+#include "pde.hpp"
+#include "program_options.hpp"
+#include "tensors.hpp"
+#include "time_advance.hpp"
+#include "transformations.hpp"
+#include <numeric>
+
+#ifdef ASGARD_USE_DOUBLE_PREC
+using prec = double;
+#else
+using prec = float;
+#endif
+
+int main(int, char**)
+{
+  // keep this file clean for each PR
+  // allows someone to easily come here, dump code and start playing
+  // this is good for prototyping and quick-testing features/behavior
+  return 0;
+}
