### 2022-09-23 6c015b0
better permutations, generalized and without recursion
