+++ src/permutations.hpp
@@ -8,41 +8,269 @@
 // construction of the element table
 // -----------------------------------------------------------------------------
 
+/*!
+ * \defgroup permuts Permutations and index selection
+ *
+ * Collection of algorithms for generating multi-indexes
+ * with different order and selection criteria.
+ */
+
 namespace permutations
 {
-// Permutations counters
+/*!
+ * \ingroup permuts
+ * \brief Generates the list of multi-indexes that satisfy a general selection
+ * criteria.
+ *
+ * A helper method to select_indexex(), it generates a list of vectors,
+ * where the first index is in entries 0 to `num_dims-1`
+ * and all follow on indexes are appended one after the other;
+ * the order of the indexes is lexicographical ascending starting with the first
+ * dimension.
+ */
+inline std::vector<int> generate_lower_index_set(
+    size_t num_dims, std::function<bool(std::vector<int> const &index)> inside)
+{
+  size_t c   = 0;
+  bool is_in = true;
+  std::vector<int> root(num_dims, 0);
+  std::vector<int> indexes;
+  while (is_in || (c < num_dims))
+  {
+    if (is_in)
+    {
+      indexes.insert(indexes.end(), root.begin(), root.end());
+      c = 0;
+      root[c]++;
+    }
+    else
+    {
+      std::fill(root.begin(), root.begin() + c + 1, 0);
+      c++;
+      if (c == num_dims)
+        break;
+      root[c]++;
+    }
+    is_in = inside(root);
+  }
+  return indexes;
+}
+/*!
+ * \ingroup permuts
+ * \brief Generates the list of multi-indexes that satisfy a general selection
+ * criteria.
+ *
+ * Identical to generate_lower_index_set(), but returns the list sorted in
+ * levels. The first vector in the result will always be the zeroth vector, the
+ * second vector will contain the selected set of indexes from among the indexes
+ * that have one entry equal to 1, the L-th vector contains the selected indexes
+ * with sum of the entries equal to L.
+ */
+inline std::vector<std::vector<int>> generate_lower_index_level_sets(
+    size_t num_dims, std::function<bool(std::vector<int> const &index)> inside)
+{
+  size_t c = 0, level = 0;
+  bool is_in = true;
+  std::vector<int> root(num_dims, 0);
+  std::vector<std::vector<int>> indexes(1);
+  while (is_in || (c < num_dims))
+  {
+    if (is_in)
+    {
+      if (level == indexes.size())
+      {
+        indexes.push_back(std::vector<int>(root));
+      }
+      else
+      {
+        indexes[level].insert(indexes[level].end(), root.begin(), root.end());
+      }
+      c = 0;
+      root[c]++;
+      level++;
+    }
+    else
+    {
+      std::fill(root.begin(), root.begin() + c + 1, 0);
+      c++;
+      if (c == num_dims)
+        break;
+      root[c]++;
+      level = root[0];
+      for (size_t i = 1; i < root.size(); i++)
+        level += root[i];
+    }
+    is_in = inside(root);
+  }
+  return indexes;
+}
+/*!
+ * \ingroup permuts
+ * \brief Converts the output of generate_lower_index_set() into
+ * fk::matrix<int>.
+ *
+ * \param num_dims must be the same as in the call to generate_lower_index_set()
+ * \param indexes must be the output of generate_lower_index_set()
+ *
+ * \returns an fk::matrix<int> containing all the indexes.
+ */
+inline fk::matrix<int>
+transp_data(int const num_dims, std::vector<int> const &indexes)
+{
+  int const nrows = indexes.size() / num_dims;
+  fk::matrix<int> result(nrows, num_dims);
[…]
