+++ src/permutations.cpp
@@ -10,220 +10,6 @@
 
 namespace permutations
 {
-//
-// Permutation enumerators
-//
-
-// Given the number of dimensions and a limit, count the number of n-tuples
-// (where n == 'num_dims') whose non-negative elements' sum == 'limit'
-int count_equal(int const num_dims, int const limit)
-{
-  expect(num_dims > 0);
-  expect(limit >= 0);
-  if (num_dims == 1)
-  {
-    return 1;
-  }
-
-  if (num_dims == 2)
-  {
-    return limit + 1;
-  }
-
-  int count = 0;
-  for (auto i = 0; i <= limit; ++i)
-  {
-    count += count_equal(num_dims - 1, i);
-  }
-
-  return count;
-}
-
-// Given the number of dimensions and a limit, count the number of n-tuples
-// (where n == 'num_dims') whose non-negative elements' sum <= 'limit'
-// this version handles the non-uniform level case
-int count_equal_multi(fk::vector<int> const &levels, int const num_dims,
-                      int const limit)
-{
-  expect(num_dims > 0);
-  expect(levels.size() == num_dims);
-  expect(limit >= 0);
-
-  if (num_dims == 1)
-  {
-    if (levels(0) >= limit)
-    {
-      return 1;
-    }
-    return 0;
-  }
-
-  auto count = 0;
-  for (auto i = 0; i <= std::min(levels(num_dims - 1), limit); ++i)
-  {
-    auto const dim_limit = limit - i;
-    count += count_equal_multi(levels.extract(0, num_dims - 2), num_dims - 1,
-                               dim_limit);
-  }
-
-  return count;
-}
-
-// Given the number of dimensions and a limit, count the number of n-tuples
-// (where n == 'num_dims') whose non-negative elements' sum <= 'limit'
-int count_lequal(int const num_dims, int const limit)
-{
-  expect(num_dims > 0);
-  expect(limit >= 0);
-
-  int count = 0;
-  for (auto i = 0; i <= limit; ++i)
-  {
-    count += count_equal(num_dims, i);
-  }
-  return count;
-}
-
-// Given the number of dimensions and a limit, count the number of n-tuples
-// (where n == 'num_dims') whose non-negative elements' sum <= 'limit'
-// this version handles the non-uniform level case
-int count_lequal_multi(fk::vector<int> const &levels, int num_dims, int limit)
-{
-  expect(num_dims > 0);
-  expect(levels.size() == num_dims);
-  expect(limit >= 0);
-
-  auto count = 0;
-  for (auto i = 0; i <= limit; ++i)
-  {
-    count += count_equal_multi(levels, num_dims, i);
-  }
-
-  return count;
-}
-
-// Given the number of dimensions and a limit, count the number of n-tuples
-// (where n == 'num_dims') whose non-negative max element <= 'limit' (for full
-// grid only)
-int count_max(int const num_dims, int const limit)
-{
-  expect(num_dims > 0);
-  expect(limit >= 0);
-
-  return static_cast<int>(std::pow(limit + 1, num_dims));
-}
-
-//
-// Permutations builders
-//
-
-// Given the number of dimensions and a limit, produce n-tuples (where n ==
-// 'num_dims') whose elements' are non-negative and their sum == 'limit'. Each
-// tuple becomes a row of the output matrix
-fk::matrix<int>
-get_equal(int const num_dims, int const limit, bool const order_by_n)
-{
-  expect(num_dims > 0);
-  expect(limit >= 0);
-
-  int const num_tuples = count_equal(num_dims, limit);
-  fk::matrix<int> result(num_tuples, num_dims);
-
-  if (num_dims == 1)
-  {
[…]
