### 2022-09-26 4014bd9
Update Catch2 and kronmult dependencies
Signed-off-by: Steven Hahn <hahnse@ornl.gov>
