### 2022-09-28 2e5a6c7
Start removing warnings caught by -Wshadow
Signed-off-by: Steven Hahn <hahnse@ornl.gov>
