+++ src/pde/pde_continuity6.hpp
@@ -77,7 +77,7 @@ private:
     ignore(t);
     fk::vector<P> fx(x.size());
     std::transform(x.begin(), x.end(), fx.begin(),
-                   [](P const &x) { return std::cos(xarg * x); });
+                   [](P const &x_v) { return std::cos(xarg * x_v); });
     return fx;
   }
   static fk::vector<P> exact_solution_y(fk::vector<P> const x, P const t = 0)
@@ -85,7 +85,7 @@ private:
     ignore(t);
     fk::vector<P> fx(x.size());
     std::transform(x.begin(), x.end(), fx.begin(),
-                   [](P const &x) { return std::sin(yarg * x); });
+                   [](P const &x_v) { return std::sin(yarg * x_v); });
     return fx;
   }
 
@@ -94,7 +94,7 @@ private:
     ignore(t);
     fk::vector<P> fx(x.size());
     std::transform(x.begin(), x.end(), fx.begin(),
-                   [](P const &x) { return std::cos(zarg * x); });
+                   [](P const &x_v) { return std::cos(zarg * x_v); });
     return fx;
   }
   static fk::vector<P> exact_solution_vx(fk::vector<P> const x, P const t = 0)
@@ -102,7 +102,7 @@ private:
     ignore(t);
     fk::vector<P> fx(x.size());
     std::transform(x.begin(), x.end(), fx.begin(),
-                   [](P const &x) { return std::cos(vxarg * x); });
+                   [](P const &x_v) { return std::cos(vxarg * x_v); });
     return fx;
   }
   static fk::vector<P> exact_solution_vy(fk::vector<P> const x, P const t = 0)
@@ -110,7 +110,7 @@ private:
     ignore(t);
     fk::vector<P> fx(x.size());
     std::transform(x.begin(), x.end(), fx.begin(),
-                   [](P const &x) { return std::sin(vyarg * x); });
+                   [](P const &x_v) { return std::sin(vyarg * x_v); });
     return fx;
   }
 
@@ -119,7 +119,7 @@ private:
     ignore(t);
     fk::vector<P> fx(x.size());
     std::transform(x.begin(), x.end(), fx.begin(),
-                   [](P const &x) { return std::cos(vzarg * x); });
+                   [](P const &x_v) { return std::cos(vzarg * x_v); });
     return fx;
   }
 
@@ -155,7 +155,7 @@ private:
     ignore(t);
     fk::vector<P> fx(x.size());
     std::transform(x.begin(), x.end(), fx.begin(),
-                   [](P const &x) { return std::cos(xarg * x); });
+                   [](P const &x_v) { return std::cos(xarg * x_v); });
     return fx;
   }
   static fk::vector<P> source_0_y(fk::vector<P> const x, P const t = 0)
@@ -163,7 +163,7 @@ private:
     ignore(t);
     fk::vector<P> fx(x.size());
     std::transform(x.begin(), x.end(), fx.begin(),
-                   [](P const &x) { return std::sin(yarg * x); });
+                   [](P const &x_v) { return std::sin(yarg * x_v); });
     return fx;
   }
   static fk::vector<P> source_0_z(fk::vector<P> const x, P const t = 0)
@@ -171,7 +171,7 @@ private:
     ignore(t);
     fk::vector<P> fx(x.size());
     std::transform(x.begin(), x.end(), fx.begin(),
-                   [](P const &x) { return std::cos(zarg * x); });
+                   [](P const &x_v) { return std::cos(zarg * x_v); });
     return fx;
   }
   static fk::vector<P> source_0_vx(fk::vector<P> const x, P const t = 0)
@@ -179,7 +179,7 @@ private:
     ignore(t);
     fk::vector<P> fx(x.size());
[…]
