+++ src/pde/pde_continuity3.hpp
@@ -72,7 +72,7 @@ private:
     ignore(t);
     fk::vector<P> fx(x.size());
     std::transform(x.begin(), x.end(), fx.begin(),
-                   [](P const &x) { return std::cos(PI * x); });
+                   [](P const &x_v) { return std::cos(PI * x_v); });
     return fx;
   }
   static fk::vector<P> exact_solution_dim1(fk::vector<P> const x, P const t = 0)
@@ -80,7 +80,7 @@ private:
     ignore(t);
     fk::vector<P> fx(x.size());
     std::transform(x.begin(), x.end(), fx.begin(),
-                   [](P const &x) { return std::sin(2.0 * PI * x); });
+                   [](P const &x_v) { return std::sin(2.0 * PI * x_v); });
     return fx;
   }
 
@@ -89,7 +89,7 @@ private:
     ignore(t);
     fk::vector<P> fx(x.size());
     std::transform(x.begin(), x.end(), fx.begin(),
-                   [](P const &x) { return std::cos(2.0 * PI * x / 3.0); });
+                   [](P const &x_v) { return std::cos(2.0 * PI * x_v / 3.0); });
     return fx;
   }
 
@@ -111,7 +111,7 @@ private:
     ignore(t);
     fk::vector<P> fx(x.size());
     std::transform(x.begin(), x.end(), fx.begin(),
-                   [](P const &x) { return std::cos(PI * x); });
+                   [](P const &x_v) { return std::cos(PI * x_v); });
     return fx;
   }
 
@@ -120,7 +120,7 @@ private:
     ignore(t);
     fk::vector<P> fx(x.size());
     std::transform(x.begin(), x.end(), fx.begin(),
-                   [](P const &x) { return std::sin(2.0 * PI * x); });
+                   [](P const &x_v) { return std::sin(2.0 * PI * x_v); });
     return fx;
   }
 
@@ -129,7 +129,7 @@ private:
     ignore(t);
     fk::vector<P> fx(x.size());
     std::transform(x.begin(), x.end(), fx.begin(),
-                   [](P const &x) { return std::cos(2.0 * PI * x / 3.0); });
+                   [](P const &x_v) { return std::cos(2.0 * PI * x_v / 3.0); });
     return fx;
   }
 
@@ -141,7 +141,7 @@ private:
     ignore(t);
     fk::vector<P> fx(x.size());
     std::transform(x.begin(), x.end(), fx.begin(),
-                   [](P const &x) { return std::cos(PI * x); });
+                   [](P const &x_v) { return std::cos(PI * x_v); });
     return fx;
   }
 
@@ -150,7 +150,7 @@ private:
     ignore(t);
     fk::vector<P> fx(x.size());
     std::transform(x.begin(), x.end(), fx.begin(),
-                   [](P const &x) { return std::cos(2.0 * PI * x); });
+                   [](P const &x_v) { return std::cos(2.0 * PI * x_v); });
     return fx;
   }
 
@@ -159,7 +159,7 @@ private:
     ignore(t);
     fk::vector<P> fx(x.size());
     std::transform(x.begin(), x.end(), fx.begin(),
-                   [](P const &x) { return std::cos(2.0 * PI * x / 3.0); });
+                   [](P const &x_v) { return std::cos(2.0 * PI * x_v / 3.0); });
     return fx;
   }
 
@@ -174,7 +174,7 @@ private:
     ignore(t);
     fk::vector<P> fx(x.size());
     std::transform(x.begin(), x.end(), fx.begin(),
-                   [](P const &x) { return std::sin(PI * x); });
+                   [](P const &x_v) { return std::sin(PI * x_v); });
     return fx;
   }
 
[…]
