### 2022-10-04 9d8b38a
flags are not set unless doing recommended
