### 2022-10-04 d4fc509
remove redundant variables
