### 2022-10-07 4596a66
Add multi-dim function support for initial conditions
