+++ src/pde/pde_diffusion1.hpp
@@ -50,7 +50,7 @@ private:
 
   /* Define the dimension */
   inline static dimension<P> const dim_0 =
-      dimension<P>(0, 1, 3, 2, initial_condition_dim0, volume_jacobian_dV, "x");
+      dimension<P>(0, 1, 3, 2, {initial_condition_dim0}, volume_jacobian_dV, "x");
 
   inline static std::vector<dimension<P>> const dimensions_ = {dim_0};
 
