### 2022-10-08 96ad912
Cleaner cmake (#441)
* added recommended defaults option

* does not automatically build opneblas, unless specified

* flags are not set unless doing recommended

* remove redundant variables
