+++ CMakeLists.txt
@@ -84,9 +84,6 @@ else ()
   message (STATUS "CMAKE_BUILD_TYPE is ${CMAKE_BUILD_TYPE}")
 endif ()
 
-# add compiler flags we always want to use
-string(APPEND CMAKE_CXX_FLAGS " -Wall -Wextra -Wpedantic")
-
 # set up possible commandline input variable defaults (override with -D)
 include(CMakeDependentOption)
 option (ASGARD_BUILD_TESTS "Build tests for asgard" ON)
@@ -96,8 +93,6 @@ option (ASGARD_BUILD_HDF5 "Download and build our own HDF5/HighFive" OFF)
 option (ASGARD_HIGHFIVE_PATH "Optional location of existing HighFive library" "")
 option (ASGARD_HDF5_PATH "Optional location of lib/ containing HDF5" "")
 option (ASGARD_BUILD_PROFILE_DEPS "Download and build our own tools (e.g.  graphviz)" OFF)
-option (ASGARD_LAPACK_PATH "optional location of include/ and lib/ containing LAPACK" "")
-option (ASGARD_BLAS_PATH "optional location of include/ and lib/ containing BLAS" "")
 option (ASGARD_PROFILE_GPROF "enable profiling support for using gprof" "")
 option (ASGARD_PROFILE_XRAY "enable profiling support for using LLVM XRAY" "")
 option (ASGARD_PROFILE_GPERF_CPU "enable cpu profiling support using gperftools" "")
@@ -106,7 +101,7 @@ option (ASGARD_PROFILE_PERF "enable profiling support for using linux perf" "")
 option (ASGARD_PROFILE_VALGRIND "enable profiling support for using valgrind" "")
 option (ASGARD_GRAPHVIZ_PATH "optional location of bin/ containing dot executable" "")
 option (ASGARD_USE_CUDA "Optional CUDA support for asgard" OFF)
-option (ASGARD_USE_OPENMP "Optional openMP support for asgard" ON)
+option (ASGARD_USE_OPENMP "Optional OpenMP support for asgard" OFF)
 option (ASGARD_USE_MPI "Optional distributed computing support for asgard" OFF)
 cmake_dependent_option (ASGARD_USE_SCALAPACK "Use optional scalapack linear algebra library" OFF "ASGARD_USE_MPI" OFF)
 option (ASGARD_USE_MATLAB "Optional MATLAB plotting support for asgard" OFF)
@@ -114,21 +109,20 @@ option (ASGARD_USE_DOUBLE_PREC "enable or disable double-precision floating poin
 option (ASGARD_USE_PCH "Enable precompiled header files." OFF)
 set (ASGARD_TESTING_RANKS "0" CACHE STRING "Override default number of ranks to use for testing")
 
+option (ASGARD_RECOMMENDED_DEFAULTS "Enable OpenMP, set some flags, download OpenBLAS if system BLAS is missing." OFF)
 option (ASGARD_BUILD_DOCS "Build the documentation." OFF)
 
 if (${ASGARD_USE_PCH})
   set (USE_PCH ON CACHE BOOL "Turn on kronmult precompiled header support" FORCE)
 endif()
 
+if (ASGARD_RECOMMENDED_DEFAULTS)
+  # add compiler flags we always want to use
+  string(APPEND CMAKE_CXX_FLAGS " -Wall -Wextra -Wpedantic")
+endif()
+
 # add scripts directory location
[…]
