### 2022-10-10 913d87b
Fix -Wshadow warnings
Signed-off-by: Steven Hahn <hahnse@ornl.gov>
