+++ src/coefficients_tests.cpp
@@ -321,9 +321,8 @@ TEMPLATE_TEST_CASE("fokkerplanck2_complete_case4 terms", "[coefficients]",
     auto const gold = fk::matrix<TestType>(
         read_matrix_from_txt_file(std::string(gold_path) + "_lhsmass.dat"));
 
-    auto constexpr tol_factor = get_tolerance<TestType>(100);
-    auto const levels         = fk::vector<int>{4, 4};
-    int const degree          = 4;
+    auto const levels = fk::vector<int>{4, 4};
+    int const degree  = 4;
 
     parser const test_parse(pde_choice, levels, degree);
     auto pde = make_PDE<TestType>(test_parse);
@@ -352,7 +351,7 @@ TEMPLATE_TEST_CASE("fokkerplanck2_complete_case4 terms", "[coefficients]",
           fk::matrix<TestType> gold_mass(
               gold.extract_submatrix(row, 0, dof, dof));
 
-          rmse_comparison(mass, gold_mass, tol_factor);
+          rmse_comparison(mass, gold_mass, get_tolerance<TestType>(100));
 
           row += dof;
         }
