+++ src/tensors_tests.cpp
@@ -629,19 +629,19 @@ TEMPLATE_TEST_CASE("fk::vector operators", "[tensors]", double, float, int)
     fk::vector<TestType> test_vect(gold);
     fk::vector<TestType, mem_type::view> const test_vect_v(test_vect);
     fk::vector<TestType, mem_type::const_view> const test_vect_cv(test_vect);
-    fk::vector<TestType> const gold{290, 490, 690};
+    fk::vector<TestType> const gold_result{290, 490, 690};
 
-    REQUIRE((test_vect * test_mat) == gold);
-    REQUIRE((test_vect_v * test_mat) == gold);
-    REQUIRE((test_vect_cv * test_mat) == gold);
+    REQUIRE((test_vect * test_mat) == gold_result);
+    REQUIRE((test_vect_v * test_mat) == gold_result);
+    REQUIRE((test_vect_cv * test_mat) == gold_result);
 
-    REQUIRE((test_vect * test_mat_v) == gold);
-    REQUIRE((test_vect_v * test_mat_v) == gold);
-    REQUIRE((test_vect_cv * test_mat_v) == gold);
+    REQUIRE((test_vect * test_mat_v) == gold_result);
+    REQUIRE((test_vect_v * test_mat_v) == gold_result);
+    REQUIRE((test_vect_cv * test_mat_v) == gold_result);
 
-    REQUIRE((test_vect * test_mat_cv) == gold);
-    REQUIRE((test_vect_v * test_mat_cv) == gold);
-    REQUIRE((test_vect_cv * test_mat_cv) == gold);
+    REQUIRE((test_vect * test_mat_cv) == gold_result);
+    REQUIRE((test_vect_v * test_mat_cv) == gold_result);
+    REQUIRE((test_vect_cv * test_mat_cv) == gold_result);
   }
 
   SECTION("vector*scalar operator")
@@ -883,18 +883,13 @@ TEMPLATE_TEST_CASE("fk::vector utilities", "[tensors]", double, float, int)
     // because you can't concat on an owner w/ outstanding views
     fk::vector<TestType> empty_copy(empty);
     fk::vector<TestType, mem_type::view> empty_v(empty_copy);
-    fk::vector<TestType> gold_copy(gold);
-    fk::vector<TestType, mem_type::view> gold_v(gold_copy);
+    fk::vector<TestType> gold_data_copy(gold);
+    fk::vector<TestType, mem_type::view> gold_data_v(gold_data_copy);
 
     REQUIRE(empty.concat(gold) == gold);
     empty.resize(0);
-    REQUIRE(empty.concat(gold_v) == gold);
+    REQUIRE(empty.concat(gold_data_v) == gold);
     empty.resize(0);
-
-    // REQUIRE(empty_v.concat(gold) == gold);
-    // empty_v.resize(0);
-    // REQUIRE(empty_v.concat(gold_v) == gold);
-    // empty_v.resize(0);
 
     // non-const gold copy I can concat with
     fk::vector<TestType> gold_2(gold);
@@ -902,9 +897,6 @@ TEMPLATE_TEST_CASE("fk::vector utilities", "[tensors]", double, float, int)
     gold_2.resize(gold.size()) = gold;
     REQUIRE(gold_2.concat(empty_v) == gold);
     gold_2.resize(gold.size()) = gold;
-    // REQUIRE(gold_copy_v.concat(empty) == gold);
-    // gold_copy_v = gold_copy;
-    // REQUIRE(gold_copy_v.concat(empty_v) == gold);
   }
 
   SECTION("vector set")
@@ -2366,20 +2358,20 @@ TEMPLATE_TEST_CASE("fk::matrix operators", "[tensors]", double, float, int)
[…]
