+++ src/lib_dispatch_tests.cpp
@@ -177,12 +177,12 @@ TEMPLATE_TEST_CASE("matrix-matrix multiply (lib_dispatch::gemm)",
     std::fill(result.begin(), result.end(), 1.0);
 
     fk::matrix<TestType> const gold = [&] {
-      fk::matrix<TestType> ans = (in1 * in2) * 2.0;
+      fk::matrix<TestType> answer = (in1 * in2) * 2.0;
       std::transform(
-          ans.begin(), ans.end(), ans.begin(),
+          answer.begin(), answer.end(), answer.begin(),
           [](TestType const elem) -> TestType { return elem + 1.0; });
 
-      return ans;
+      return answer;
     }();
 
     TestType alpha     = 2.0;
@@ -210,12 +210,12 @@ TEMPLATE_TEST_CASE("matrix-matrix multiply (lib_dispatch::gemm)",
           result.clone_onto_device());
 
       fk::matrix<TestType> const gold = [&] {
-        fk::matrix<TestType> ans = (in1 * in2) * 2.0;
+        fk::matrix<TestType> answer = (in1 * in2) * 2.0;
         std::transform(
-            ans.begin(), ans.end(), ans.begin(),
+            answer.begin(), answer.end(), answer.begin(),
             [](TestType const elem) -> TestType { return elem + 1.0; });
 
-        return ans;
+        return answer;
       }();
 
       TestType alpha     = 2.0;
@@ -419,12 +419,12 @@ TEMPLATE_TEST_CASE("matrix-vector multiply (lib_dispatch::gemv)",
     std::fill(result.begin(), result.end(), 1.0);
 
     fk::vector<TestType> const gold = [&] {
-      fk::vector<TestType> ans = (A * x) * 2.0;
+      fk::vector<TestType> answer = (A * x) * 2.0;
       std::transform(
-          ans.begin(), ans.end(), ans.begin(),
+          answer.begin(), answer.end(), answer.begin(),
           [](TestType const elem) -> TestType { return elem + 1.0; });
 
-      return ans;
+      return answer;
     }();
     TestType alpha     = 2.0;
     TestType beta      = 1.0;
@@ -449,12 +449,12 @@ TEMPLATE_TEST_CASE("matrix-vector multiply (lib_dispatch::gemv)",
           result.clone_onto_device());
 
       fk::vector<TestType> const gold = [&] {
-        fk::vector<TestType> ans = (A * x) * 2.0;
+        fk::vector<TestType> answer = (A * x) * 2.0;
         std::transform(
-            ans.begin(), ans.end(), ans.begin(),
+            answer.begin(), answer.end(), answer.begin(),
             [](TestType const elem) -> TestType { return elem + 1.0; });
 
-        return ans;
+        return answer;
       }();
 
       TestType alpha     = 2.0;
@@ -857,8 +857,8 @@ void test_batched_gemm(int const m, int const n, int const k, int const lda,
   std::vector<std::vector<fk::matrix<P, mem_type::owner, resrc>>> const
       matrices = [=]() {
         // {a, b, c, gold}
-        std::vector<std::vector<fk::matrix<P, mem_type::owner, resrc>>>
-            matrices(4);
+        std::vector<std::vector<fk::matrix<P, mem_type::owner, resrc>>> output(
+            4);
 
         std::random_device rd;
         std::mt19937 mersenne_engine(rd());
@@ -878,15 +878,15 @@ void test_batched_gemm(int const m, int const n, int const k, int const lda,
 
           if constexpr (resrc == resource::host)
           {
-            matrices[0].push_back(a);
-            matrices[1].push_back(b);
-            matrices[2].push_back(c);
+            output[0].push_back(a);
+            output[1].push_back(b);
+            output[2].push_back(c);
           }
           else
           {
[…]
