+++ src/pde/pde_base.hpp
@@ -81,12 +81,13 @@ public:
   vector_func<P> const initial_condition;
   g_func_type<P> const moment_dV;
   std::string const name;
-  dimension(P const domain_min, P const domain_max, int const level,
-            int const degree, vector_func<P> const initial_condition,
-            g_func_type<P> const moment_dV, std::string const name)
+  dimension(P const d_min, P const d_max, int const level, int const degree,
+            vector_func<P> const initial_condition_in,
+            g_func_type<P> const moment_dV_in, std::string const name_in)
 
-      : domain_min(domain_min), domain_max(domain_max),
-        initial_condition(initial_condition), moment_dV(moment_dV), name(name)
+      : domain_min(d_min), domain_max(d_max),
+        initial_condition(initial_condition_in), moment_dV(moment_dV_in),
+        name(name_in)
   {
     set_level(level);
     set_degree(degree);
@@ -162,27 +163,28 @@ public:
 
   static P null_scalar_func(P const p) { return p; }
 
-  partial_term(coefficient_type const coeff_type,
-               g_func_type<P> const g_func        = null_gfunc,
-               g_func_type<P> const lhs_mass_func = null_gfunc,
-               flux_type const flux               = flux_type::central,
-               boundary_condition const left      = boundary_condition::neumann,
-               boundary_condition const right     = boundary_condition::neumann,
-               homogeneity const left_homo        = homogeneity::homogeneous,
-               homogeneity const right_homo       = homogeneity::homogeneous,
-               std::vector<vector_func<P>> const left_bc_funcs = {},
-               scalar_func<P> const left_bc_time_func = null_scalar_func,
-               std::vector<vector_func<P>> const right_bc_funcs = {},
-               scalar_func<P> const right_bc_time_func = null_scalar_func,
-               g_func_type<P> const dv_func            = null_gfunc)
+  partial_term(coefficient_type const coeff_type_in,
+               g_func_type<P> const g_func_in        = null_gfunc,
+               g_func_type<P> const lhs_mass_func_in = null_gfunc,
+               flux_type const flux_in               = flux_type::central,
+               boundary_condition const left_in  = boundary_condition::neumann,
+               boundary_condition const right_in = boundary_condition::neumann,
+               homogeneity const left_homo_in    = homogeneity::homogeneous,
+               homogeneity const right_homo_in   = homogeneity::homogeneous,
+               std::vector<vector_func<P>> const left_bc_funcs_in = {},
+               scalar_func<P> const left_bc_time_func_in = null_scalar_func,
+               std::vector<vector_func<P>> const right_bc_funcs_in = {},
+               scalar_func<P> const right_bc_time_func_in = null_scalar_func,
+               g_func_type<P> const dv_func_in            = null_gfunc)
 
-      : coeff_type(coeff_type), g_func(g_func), lhs_mass_func(lhs_mass_func),
-        flux(set_flux(flux)), left(left), right(right),
-        ileft(set_bilinear_boundary(left)),
-        iright(set_bilinear_boundary(right)), left_homo(left_homo),
-        right_homo(right_homo), left_bc_funcs(left_bc_funcs),
-        right_bc_funcs(right_bc_funcs), left_bc_time_func(left_bc_time_func),
-        right_bc_time_func(right_bc_time_func), dv_func(dv_func)
+      : coeff_type(coeff_type_in), g_func(g_func_in),
+        lhs_mass_func(lhs_mass_func_in), flux(set_flux(flux_in)), left(left_in),
[…]
