### 2022-10-14 be90b84
Add moment reduced matrix to moment class
