+++ src/moment.cpp
@@ -1,4 +1,5 @@
 #include "moment.hpp"
+#include "elements.hpp"
 #include "transformations.hpp"
 
 template<typename P>
@@ -17,10 +18,6 @@ void moment<P>::createFlist(PDE<P> const &pde, options const &opts)
   std::size_t num_dims = dims.size();
 
   this->fList.resize(num_md_funcs);
-  for (auto &elem : this->fList)
-  {
-    elem.resize(num_dims);
-  }
 
   basis::wavelet_transform<P, resource::host> const transformer(opts, pde);
 
@@ -29,8 +26,8 @@ void moment<P>::createFlist(PDE<P> const &pde, options const &opts)
     auto const &md_func = this->md_funcs[s];
     for (std::size_t d = 0; d < num_dims; ++d)
     {
-      fList[s][d] = forward_transform<P>(
-          dims[d], md_func, dims[d].volume_jacobian_dV, transformer);
+      fList[s].push_back(forward_transform<P>(
+          dims[d], md_func, dims[d].volume_jacobian_dV, transformer));
     }
   }
 }
@@ -41,16 +38,103 @@ template<typename P>
 void moment<P>::createMomentVector(parser const &opts,
                                    elements::table const &hash_table)
 {
+  // check that fList has been constructed
+  expect(this->fList.size() > 0);
+
   if (this->vector.size() == 0 || opts.do_adapt_levels())
   {
     distribution_plan const plan = get_plan(get_num_ranks(), hash_table);
-    auto rank = get_rank();
-    this->vector = combine_dimensions(opts.get_degree(), hash_table, plan.at(rank).row_start, plan.at(rank).row_stop, this->fList[0]);
+    auto rank                    = get_rank();
+    auto tmp = combine_dimensions(opts.get_degree(), hash_table,
+                                  plan.at(rank).row_start,
+                                  plan.at(rank).row_stop, this->fList[0]);
+    this->vector.resize(tmp.size());
+    this->vector      = std::move(tmp);
     auto num_md_funcs = md_funcs.size();
     for (std::size_t s = 1; s < num_md_funcs; ++s)
     {
-      auto tmp = combine_dimensions(opts.get_degree(), hash_table, plan.at(rank).row_start, plan.at(rank).row_stop, this->fList[s]);
-      std::transform(tmp.begin(), tmp.end(), this->vector.begin(), this->vector.begin(), std::plus<>{});
+      tmp = combine_dimensions(opts.get_degree(), hash_table,
+                               plan.at(rank).row_start, plan.at(rank).row_stop,
+                               this->fList[s]);
+      std::transform(tmp.begin(), tmp.end(), this->vector.begin(),
+                     this->vector.begin(), std::plus<>{});
+    }
+  }
+}
+
+// helpers for converting linear coordinates into operator matrix indices
+inline fk::vector<int> linearize(fk::vector<int> const &coords)
+{
+  fk::vector<int> linear(coords.size() / 2);
+  for (int i = 0; i < linear.size(); ++i)
+  {
+    linear(i) = elements::get_1d_index(coords(i), coords(i + linear.size()));
+  }
+  return linear;
+}
+
+template<typename P>
+inline fk::vector<int>
+linear_coords_to_indices(PDE<P> const &pde, int const degree,
+                         fk::vector<int> const &coords)
+{
+  fk::vector<int> indices(coords.size());
+  for (int d = 0; d < pde.num_dims; ++d)
+  {
+    indices(d) = coords(d) * degree;
+  }
+  return indices;
+}
+
+template<typename P>
+void moment<P>::createMomentReducedMatrix(PDE<P> const &pde,
[…]
