+++ README.md
@@ -22,7 +22,7 @@ Issues are a great way to discuss all aspects of the ASGarD project, whether it
 is to ask a general question, request a new feature, or propose a contribution
 to the code base.
 
-The ASGarD project is led by David Green at Oak Ridge
+The ASGarD project was initiated by David Green at Oak Ridge
 National Laboratory.
 
 For technical questions, contact Miroslav Stoyanov (stoyanovmk@ornl.gov) at
