+++ src/pde/pde_vlasov_lb_full_f.hpp
@@ -29,16 +29,14 @@ private:
   static P constexpr nu = 1e3;
 
   static fk::vector<P>
-  initial_condition_dim_x(fk::vector<P> const &x, P const t = 0)
+  initial_condition_dim_x_0(fk::vector<P> const &x, P const t = 0)
   {
-    // TODO: fix this - needs multi-dimensional func ability
-    P const coefficient = 1.0 / std::sqrt(2.0 * PI);
-
+    ignore(t);
     fk::vector<P> fx(x.size());
     std::transform(x.begin(), x.end(), fx.begin(), [](P const x) -> P {
       if (std::abs(x) > 0.5)
       {
-        return coefficient * std::exp(-0.5 * std::pow(x, 2));
+        return 1.0;
       }
       else
       {
@@ -49,22 +47,50 @@ private:
   }
 
   static fk::vector<P>
-  initial_condition_dim_v(fk::vector<P> const &x, P const t = 0)
+  initial_condition_dim_x_1(fk::vector<P> const &x, P const t = 0)
   {
-    // TODO: fix this - needs multi-dimensional func ability
-    P const coefficient = (1.0 / 8.0) / std::sqrt(2.0 * PI * (4.0 / 5.0));
-
+    ignore(t);
     fk::vector<P> fx(x.size());
     std::transform(x.begin(), x.end(), fx.begin(), [](P const x) -> P {
-      if (x <= 0.5)
+      if (std::abs(x) <= 0.5)
       {
-        return coefficient * std::exp(-std::pow(x, 2) / (2.0 * (4.0 / 5.0)));
+        return 1.0;
       }
       else
       {
         return 0.0;
       }
     });
+    return fx;
+  }
+
+  static fk::vector<P>
+  initial_condition_dim_v_0(fk::vector<P> const &x, P const t = 0)
+  {
+    ignore(t);
+
+    P const coefficient = 1.0 / std::sqrt(2.0 * PI);
+
+    fk::vector<P> fx(x.size());
+    std::transform(x.begin(), x.end(), fx.begin(),
+                   [coefficient](P const x) -> P {
+                     return coefficient * std::exp(-std::pow(x, 2) / 2.0);
+                   });
+    return fx;
+  }
+
+  static fk::vector<P>
+  initial_condition_dim_v_1(fk::vector<P> const &x, P const t = 0)
+  {
+    ignore(t);
+
+    P const coefficient = (1.0 / 8.0) / std::sqrt(2.0 * PI * (4.0 / 5.0));
+
+    fk::vector<P> fx(x.size());
+    std::transform(
+        x.begin(), x.end(), fx.begin(), [coefficient](P const x) -> P {
+          return coefficient * std::exp(-std::pow(x, 2) / (2.0 * (4.0 / 5.0)));
+        });
     return fx;
   }
 
@@ -77,16 +103,20 @@ private:
 
   /* Define the dimension */
   inline static dimension<P> const dim_0 = dimension<P>(
-      -1.0, 1.0, 8, default_degree, {initial_condition_dim_x}, dV, "x");
+      -1.0, 1.0, 8, default_degree,
+      {initial_condition_dim_x_0, initial_condition_dim_x_1}, dV, "x");
 
   inline static dimension<P> const dim_1 = dimension<P>(
-      -6.0, 6.0, 3, default_degree, {initial_condition_dim_v}, dV, "v");
+      -6.0, 6.0, 3, default_degree,
+      {initial_condition_dim_v_0, initial_condition_dim_v_1}, dV, "v");
 
   inline static std::vector<dimension<P>> const dimensions_ = {dim_0, dim_1};
 
   /* Define the moments */
[…]
