### 2022-10-18 d54d051
Document vector class
Signed-off-by: Steven Hahn <hahnse@ornl.gov>
