+++ src/tensors.hpp
@@ -8,6 +8,7 @@
 #include "lib_dispatch.hpp"
 #include "tools.hpp"
 
+#include <filesystem>
 #include <memory>
 #include <string>
 #include <vector>
@@ -83,6 +84,11 @@ using enable_for_device = std::enable_if_t<resrc == resource::device>;
 // device views can only be created from a device owner, and host
 // views can only be constructor from host owners
 
+/*!
+ * \defgroup tensors
+ *
+ * One- and two-dimensional tensors managing memory allocation and destruction.
+ */
 namespace fk
 {
 // forward declarations
@@ -93,6 +99,8 @@ template<typename P, mem_type mem = mem_type::owner,
          resource resrc = resource::host>
 class matrix;
 
+/*! One-dimensional tensor managing memory allocation and destruction.
+ */
 template<typename P, mem_type mem, resource resrc>
 class vector
 {
@@ -103,113 +111,183 @@ class vector
 public:
   template<mem_type m_ = mem, typename = enable_for_owner<m_>>
   vector();
+  /*! constructor
+   * \param size size of newly constructed vector.
+   */
   template<mem_type m_ = mem, typename = enable_for_owner<m_>>
   explicit vector(int const size);
+  /*! constructor
+   * \param list initial values of vector.
+   */
   template<mem_type m_ = mem, typename = enable_for_owner<m_>>
   vector(std::initializer_list<P> list);
+  /*! copy constructor
+   * \param other vector
+   */
   template<mem_type m_ = mem, typename = enable_for_owner<m_>,
            resource r_ = resrc, typename = enable_for_host<r_>>
-  vector(std::vector<P> const &);
-
+  vector(std::vector<P> const &other);
+  /*! copy constructor
+   * \param other vector
+   */
   template<mem_type m_ = mem, typename = enable_for_owner<m_>>
-  vector(fk::matrix<P, mem_type::owner, resrc> const &);
-
+  vector(fk::matrix<P, mem_type::owner, resrc> const &other);
+  /*! copy constructor returns non-owning view of data
+   * \param vec vector owning data array
+   * \param start_index first element contained in view (inclusive)
+   * \param stop_index last element contained in view (inclusive)
+   */
   template<mem_type m_ = mem, typename = enable_for_view<m_>, mem_type omem,
            mem_type om_ = omem, typename = disable_for_const_view<om_>>
   explicit vector(fk::vector<P, omem, resrc> &vec, int const start_index,
                   int const stop_index);
+  /*! copy constructor returns const non-owning view of data
+   * \param vec vector owning data array
+   * \param start_index first element contained in view (inclusive)
+   * \param stop_index last element contained in view (inclusive)
+   */
   template<mem_type m_ = mem, typename = enable_for_const_view<m_>,
            mem_type omem>
   explicit vector(fk::vector<P, omem, resrc> const &vec, int const start_index,
                   int const stop_index);
-
-  // overloads for default case - whole vector
+  /*! copy constructor
+   * overloads for default case - whole vector
+   * \param other vector
+   */
   template<mem_type m_ = mem, typename = enable_for_view<m_>, mem_type omem,
            mem_type om_ = omem, typename = disable_for_const_view<om_>>
   explicit vector(fk::vector<P, omem, resrc> &owner);
 
+  /*! copy constructor
+   * overloads for default case - whole vector
+   * \param other vector
+   */
   template<mem_type m_ = mem, typename = enable_for_const_view<m_>,
            mem_type omem>
   explicit vector(fk::vector<P, omem, resrc> const &owner);
 
-  // create vector view from matrix
-  // const view version
+  /*! create vector view from matrix
+   *  const view version
+   * \param source matrix owning data array
+   * \param col_index column contained in view
[…]
