### 2022-10-18 c18b8bb
removed the unused data term from the term class
