+++ src/pde/pde_fokkerplanck2_complete.hpp
@@ -467,16 +467,12 @@ private:
       partial_term<P>::null_scalar_func, dV_z);
 
   // 2. combine partial terms into single dimension term
-  inline static term<P> const c1_term_p =
-      term<P>(false,           // time-dependent
-              fk::vector<P>(), // additional data vector
-              "C1_p",          // name
-              {c1_pterm1, c1_pterm2});
-  inline static term<P> const c1_term_z =
-      term<P>(false,           // time-dependent
-              fk::vector<P>(), // additional data vector
-              "C1_z",          // name
-              {c1_pterm3, c1_pterm3});
+  inline static term<P> const c1_term_p = term<P>(false,  // time-dependent
+                                                  "C1_p", // name
+                                                  {c1_pterm1, c1_pterm2});
+  inline static term<P> const c1_term_z = term<P>(false,  // time-dependent
+                                                  "C1_z", // name
+                                                  {c1_pterm3, c1_pterm3});
 
   // 3. combine single dimension terms into multi dimension term
   inline static std::vector<term<P>> const termC1 = {c1_term_p, c1_term_z};
@@ -509,16 +505,12 @@ private:
       partial_term<P>::null_scalar_func, dV_z);
 
   // 2. combine partial terms into single dimension term
-  inline static term<P> const c2_term_p =
-      term<P>(false,           // time-dependent
-              fk::vector<P>(), // additional data vector
-              "C2_p",          // name
-              {c2_pterm1});
-  inline static term<P> const c2_term_z =
-      term<P>(false,           // time-dependent
-              fk::vector<P>(), // additional data vector
-              "C2_z",          // name
-              {c2_pterm2});
+  inline static term<P> const c2_term_p = term<P>(false,  // time-dependent
+                                                  "C2_p", // name
+                                                  {c2_pterm1});
+  inline static term<P> const c2_term_z = term<P>(false,  // time-dependent
+                                                  "C2_z", // name
+                                                  {c2_pterm2});
 
   // 3. combine single dimension terms into multi dimension term
   inline static std::vector<term<P>> const termC2 = {c2_term_p, c2_term_z};
@@ -567,16 +559,12 @@ private:
       partial_term<P>::null_scalar_func, dV_z3);
 
   // 2. combine partial terms into single dimension term
-  inline static term<P> const c3_term_p =
-      term<P>(false,           // time-dependent
-              fk::vector<P>(), // additional data vector
-              "C3_p",          // name
-              {c3_pterm1, c3_pterm1});
-  inline static term<P> const c3_term_z =
-      term<P>(false,           // time-dependent
-              fk::vector<P>(), // additional data vector
-              "C3_z",          // name
-              {c3_pterm2, c3_pterm3});
+  inline static term<P> const c3_term_p = term<P>(false,  // time-dependent
+                                                  "C3_p", // name
+                                                  {c3_pterm1, c3_pterm1});
+  inline static term<P> const c3_term_z = term<P>(false,  // time-dependent
+                                                  "C3_z", // name
+                                                  {c3_pterm2, c3_pterm3});
 
   // 3. combine single dimension terms into multi dimension term
   inline static std::vector<term<P>> const termC3 = {c3_term_p, c3_term_z};
@@ -620,16 +608,12 @@ private:
       partial_term<P>::null_scalar_func, dV_z);
 
   // 2. combine partial terms into single dimension term
-  inline static term<P> const e1_term_p =
-      term<P>(false,           // time-dependent
-              fk::vector<P>(), // additional data vector
-              "E1_p",          // name
-              {e1_pterm1});
[…]
