+++ src/pde/pde_fokkerplanck1_4p3.hpp
@@ -149,9 +149,11 @@ private:
       homogeneity::homogeneous, {}, partial_term<P>::null_scalar_func, {},
       partial_term<P>::null_scalar_func, dV_func);
 
-  inline static term<P> const term0_dim0_ = term<P>(false,  // time-dependent
-                                                    "d_dx", // name
-                                                    {partial_term_0});
+  inline static term<P> const term0_dim0_ =
+      term<P>(false,           // time-dependent
+              fk::vector<P>(), // additional data vector
+              "d_dx",          // name
+              {partial_term_0});
 
   inline static std::vector<term<P>> const terms0_ = {term0_dim0_};
 
