+++ src/pde/pde_continuity1.hpp
@@ -179,9 +179,11 @@ private:
       boundary_condition::periodic, boundary_condition::periodic);
 
   // define terms (1 in this case)
-  inline static term<P> term0_dim0_ = term<P>(false,  // time-dependent
-                                              "d_dx", // name
-                                              {partial_term_0});
+  inline static term<P> term0_dim0_ =
+      term<P>(false,           // time-dependent
+              fk::vector<P>(), // additional data vector
+              "d_dx",          // name
+              {partial_term_0});
 
   inline static std::vector<term<P>> const terms0_ = {term0_dim0_};
 
