+++ src/pde/pde_continuity2.hpp
@@ -230,18 +230,22 @@ private:
       coefficient_type::div, g_func_t0_d0, g_func_identity, flux_type::downwind,
       boundary_condition::periodic, boundary_condition::periodic);
 
-  inline static term<P> const term0_dim0_ = term<P>(false, // time-dependent
-                                                    "v_x.d_dx", // name
-                                                    {partial_term_t0_d0});
+  inline static term<P> const term0_dim0_ =
+      term<P>(false,           // time-dependent
+              fk::vector<P>(), // additional data vector
+              "v_x.d_dx",      // name
+              {partial_term_t0_d0});
 
   inline static partial_term<P> const partial_term_t0_d1 =
       partial_term<P>(coefficient_type::mass, g_func_identity, g_func_identity,
                       flux_type::central, boundary_condition::periodic,
                       boundary_condition::periodic);
 
-  inline static term<P> const term0_dim1_ = term<P>(false,   // time-dependent
-                                                    "massY", // name
-                                                    {partial_term_t0_d1});
+  inline static term<P> const term0_dim1_ =
+      term<P>(false,           // time-dependent
+              fk::vector<P>(), // additional data vector
+              "massY",         // name
+              {partial_term_t0_d1});
 
   inline static std::vector<term<P>> const terms0_ = {term0_dim0_, term0_dim1_};
 
@@ -252,17 +256,21 @@ private:
                       flux_type::central, boundary_condition::periodic,
                       boundary_condition::periodic);
 
-  inline static term<P> const term1_dim0_ = term<P>(false,   // time-dependent
-                                                    "massX", // name
-                                                    {partial_term_t1_d0});
+  inline static term<P> const term1_dim0_ =
+      term<P>(false,           // time-dependent
+              fk::vector<P>(), // additional data vector
+              "massX",         // name
+              {partial_term_t1_d0});
 
   inline static partial_term<P> const partial_term_t1_d1 = partial_term<P>(
       coefficient_type::div, g_func_t1_d1, g_func_identity, flux_type::downwind,
       boundary_condition::periodic, boundary_condition::periodic);
 
-  inline static term<P> const term1_dim1_ = term<P>(false, // time-dependent
-                                                    "v_y.d_dy", // name
-                                                    {partial_term_t1_d1});
+  inline static term<P> const term1_dim1_ =
+      term<P>(false,           // time-dependent
+              fk::vector<P>(), // additional data vector
+              "v_y.d_dy",      // name
+              {partial_term_t1_d1});
 
   inline static std::vector<term<P>> const terms1_ = {term1_dim0_, term1_dim1_};
 
