### 2022-10-21 37489c8
Fix multi dim functions in moments
