+++ src/pde/pde_vlasov_lb_full_f.hpp
@@ -139,11 +139,11 @@ private:
   }
 
   inline static moment<P> const moment0 = moment<P>(
-      std::vector<vector_func<P>>({moment0_f1, moment0_f1, moment0_f1}));
+      std::vector<md_func_type<P>>({{moment0_f1, moment0_f1, moment0_f1}}));
   inline static moment<P> const moment1 = moment<P>(
-      std::vector<vector_func<P>>({moment0_f1, moment1_f1, moment0_f1}));
+      std::vector<md_func_type<P>>({{moment0_f1, moment1_f1, moment0_f1}}));
   inline static moment<P> const moment2 = moment<P>(
-      std::vector<vector_func<P>>({moment0_f1, moment2_f1, moment0_f1}));
+      std::vector<md_func_type<P>>({{moment0_f1, moment2_f1, moment0_f1}}));
 
   inline static std::vector<moment<P>> const moments_ = {moment0, moment1,
                                                          moment2};
@@ -254,17 +254,13 @@ private:
       flux_type::central, boundary_condition::periodic,
       boundary_condition::periodic);
 
-  inline static term<P> const term_e1x =
-      term<P>(false,           // time-dependent
-              fk::vector<P>(), // additional data vector
-              "E1_x",          // name
-              {e1_pterm_x});
+  inline static term<P> const term_e1x = term<P>(false,  // time-dependent
+                                                 "E1_x", // name
+                                                 {e1_pterm_x});
 
-  inline static term<P> const term_e1v =
-      term<P>(false,           // time-dependent
-              fk::vector<P>(), // additional data vector
-              "E1_v",          // name
-              {e1_pterm_v});
+  inline static term<P> const term_e1v = term<P>(false,  // time-dependent
+                                                 "E1_v", // name
+                                                 {e1_pterm_v});
 
   inline static std::vector<term<P>> const terms_1 = {term_e1x, term_e1v};
 
@@ -298,17 +294,13 @@ private:
       flux_type::central, boundary_condition::periodic,
       boundary_condition::periodic);
 
-  inline static term<P> const term_e2x =
-      term<P>(false,           // time-dependent
-              fk::vector<P>(), // additional data vector
-              "E2_x",          // name
-              {e2_pterm_x});
+  inline static term<P> const term_e2x = term<P>(false,  // time-dependent
+                                                 "E2_x", // name
+                                                 {e2_pterm_x});
 
-  inline static term<P> const term_e2v =
-      term<P>(false,           // time-dependent
-              fk::vector<P>(), // additional data vector
-              "E2_v",          // name
-              {e2_pterm_v});
+  inline static term<P> const term_e2v = term<P>(false,  // time-dependent
+                                                 "E2_v", // name
+                                                 {e2_pterm_v});
 
   inline static std::vector<term<P>> const terms_2 = {term_e2x, term_e2v};
 
@@ -338,17 +330,13 @@ private:
                       flux_type::downwind, boundary_condition::dirichlet,
                       boundary_condition::dirichlet);
 
-  inline static term<P> const term_i1x =
-      term<P>(false,           // time-dependent
-              fk::vector<P>(), // additional data vector
-              "I1_x",          // name
-              {i1_pterm_x});
+  inline static term<P> const term_i1x = term<P>(false,  // time-dependent
+                                                 "I1_x", // name
+                                                 {i1_pterm_x});
 
-  inline static term<P> const term_i1v =
-      term<P>(false,           // time-dependent
-              fk::vector<P>(), // additional data vector
-              "I1_v",          // name
-              {i1_pterm_v});
+  inline static term<P> const term_i1v = term<P>(false,  // time-dependent
+                                                 "I1_v", // name
[…]
