### 2022-10-21 1a7647a
Add operator[] to fk::vector
Signed-off-by: Steven Hahn <hahnse@ornl.gov>
