### 2022-10-28 #4378 @joglekara
Hey @danielskatz , see https://github.com/openjournals/joss-reviews/issues/4378#issuecomment-1250159180 .

In general, the repo needs some more documentation and examples. To be specific, there are a couple of issues I wrote that will help me keep going through the checklist.
