### 2022-10-28 #4378 @danielskatz
👋 @quantumsteve - How are you doing on the issues mentioned by @joglekara?
