### 2022-10-29 344e2ae
Refactor moment reduced matrix
