+++ src/pde/pde_fokkerplanck1_4p4.hpp
@@ -102,7 +102,7 @@ private:
     return 1.0;
   }
 
-  static P moment_dV(P const x, P const time)
+  static P volume_jacobian_dV(P const x, P const time)
   {
     // suppress compiler warnings
     ignore(x);
@@ -163,7 +163,7 @@ private:
                    2,      // levels
                    2,      // degree
                    f0_vec, // initial condition
-                   moment_dV,
+                   volume_jacobian_dV,
                    "x"); // name
 
   inline static std::vector<dimension<P>> const dimensions_ = {dim0_};
