+++ src/pde/pde_fokkerplanck1_4p3.hpp
@@ -89,7 +89,7 @@ private:
     return 1.0;
   }
 
-  static P moment_dV(P const x, P const time)
+  static P volume_jacobian_dV(P const x, P const time)
   {
     // suppress compiler warnings
     ignore(x);
@@ -120,7 +120,7 @@ private:
                    2,                      // levels
                    2,                      // degree
                    initial_condition_dim0, // initial condition
-                   moment_dV,
+                   volume_jacobian_dV,
                    "x"); // name
 
   inline static std::vector<dimension<P>> const dimensions_ = {dim0_};

[…]
