+++ src/pde/pde_fokkerplanck1_pitch_E.hpp
@@ -100,7 +100,7 @@ private:
     return 1.0;
   }
 
-  static P moment_dV(P const x, P const time)
+  static P volume_jacobian_dV(P const x, P const time)
   {
     // suppress compiler warnings
     ignore(x);
@@ -159,7 +159,7 @@ private:
                    2,                      // levels
                    2,                      // degree
                    initial_condition_dim0, // initial condition
-                   moment_dV,
+                   volume_jacobian_dV,
                    "x"); // name
 
   inline static std::vector<dimension<P>> const dimensions_ = {dim0_};
