+++ src/pde/pde_continuity2.hpp
@@ -83,7 +83,7 @@ private:
     return fx;
   }
 
-  static P moment_dV_dim0(P const x, P const time)
+  static P volume_jacobian_dV_dim0(P const x, P const time)
   {
     // suppress compiler warnings
     ignore(x);
@@ -91,7 +91,7 @@ private:
     return 1.0;
   }
 
-  static P moment_dV_dim1(P const x, P const time)
+  static P volume_jacobian_dV_dim1(P const x, P const time)
   {
     // suppress compiler warnings
     ignore(x);
@@ -205,22 +205,22 @@ private:
 
   // define dimensions
   inline static dimension<P> const dim0_ =
-      dimension<P>(-1.0,                   // domain min
-                   1.0,                    // domain max
-                   2,                      // levels
-                   2,                      // degree
-                   initial_condition_dim0, // initial condition
-                   moment_dV_dim0,         // volume portion
-                   "x");                   // name
+      dimension<P>(-1.0,                    // domain min
+                   1.0,                     // domain max
+                   2,                       // levels
+                   2,                       // degree
+                   initial_condition_dim0,  // initial condition
+                   volume_jacobian_dV_dim0, // volume portion
+                   "x");                    // name
 
   inline static dimension<P> const dim1_ =
-      dimension<P>(-2.0,                   // domain min
-                   2.0,                    // domain max
-                   2,                      // levels
-                   2,                      // degree
-                   initial_condition_dim1, // initial condition
-                   moment_dV_dim1,         // volume portion
-                   "y");                   // name
+      dimension<P>(-2.0,                    // domain min
+                   2.0,                     // domain max
+                   2,                       // levels
+                   2,                       // degree
+                   initial_condition_dim1,  // initial condition
+                   volume_jacobian_dV_dim1, // volume portion
+                   "y");                    // name
 
   inline static std::vector<dimension<P>> const dimensions_ = {dim0_, dim1_};
 
