### 2022-11-04 71c2573
Document vector class (#462)
* Document vector class

Signed-off-by: Steven Hahn <hahnse@ornl.gov>
