### 2022-11-07 3d54fb1
Fix MATLAB option in CMake
