### 2022-11-07 bdc8698
Fix MPI issue in moment tests
