+++ src/moment_tests.cpp
@@ -1,41 +1,51 @@
+#include "adapt.hpp"
 #include "basis.hpp"
 #include "coefficients.hpp"
 #include "moment.hpp"
 #include "tests_general.hpp"
-#include "time_advance.hpp"
 
 static auto const moment_base_dir = gold_base_dir / "moment";
+
+struct distribution_test_init
+{
+  distribution_test_init() { initialize_distribution(); }
+  ~distribution_test_init() { finalize_distribution(); }
+};
+
+#ifdef ASGARD_USE_MPI
+static distribution_test_init const distrib_test_info;
+#endif
 
 TEMPLATE_TEST_CASE("Multiwavelet", "[transformations]", double, float)
 {
   std::string const pde_choice = "diffusion_2";
   fk::vector<int> const levels{5, 5};
-  auto const degree = 4;
-  auto const cfl = 0.01;
-  auto const full_grid = false;
+  auto const degree               = 4;
+  auto const cfl                  = 0.01;
+  auto const full_grid            = false;
   static auto constexpr num_steps = 5;
-  auto const use_implicit = true;
-  auto const do_adapt_levels = true;
-  auto const adapt_threshold = 0.5e-1;
+  auto const use_implicit         = true;
+  auto const do_adapt_levels      = true;
+  auto const adapt_threshold      = 0.5e-1;
 
   parser const parse(pde_choice, levels, degree, cfl, full_grid,
-                       parser::DEFAULT_MAX_LEVEL, num_steps, use_implicit,
-                       do_adapt_levels, adapt_threshold);
+                     parser::DEFAULT_MAX_LEVEL, num_steps, use_implicit,
+                     do_adapt_levels, adapt_threshold);
 
   auto pde = make_PDE<TestType>(parse);
   options const opts(parse);
   elements::table const check(opts, *pde);
 
   adapt::distributed_grid adaptive_grid(*pde, opts);
-  basis::wavelet_transform<TestType, resource::host> const transformer(opts, *pde);
+  basis::wavelet_transform<TestType, resource::host> const transformer(opts,
+                                                                       *pde);
 
-    // -- set coeffs
+  // -- set coeffs
   generate_all_coefficients(*pde, transformer);
 
   // -- generate initial condition vector.
-  auto const initial_condition = adaptive_grid.get_initial_condition(*pde, transformer, opts);
-
-
+  auto const initial_condition =
+      adaptive_grid.get_initial_condition(*pde, transformer, opts);
 
   std::vector<vector_func<TestType>> md_func;
   SECTION("Constructor") { moment<TestType> mymoment({md_func}); }
@@ -93,101 +103,3 @@ TEMPLATE_TEST_CASE("CreateMomentReducedMatrix", "[moments]", double, float)
                     tol_factor);
   }
 }
-
-/*function test_moment_2D_spherical(testCase)
-
-args = {'lev',5,'deg',4,'case',3};
-opts = OPTS(args);
-pde = diffusion2(opts);
-
-num_dims = numel(pde.dimensions);
-[elements, elements_idx]    = hash_table_sparse_nD (pde.get_lev_vec, opts.max_lev, opts.grid_type);
-hash_table.elements         = elements;
[…]
