### 2022-11-08 366bfb6
Fix bug in determining max level across dims
