### 2022-11-08 902bcb8
Fix undefined degree in moments when no cli degree given
