+++ src/moment.cpp
@@ -35,7 +35,7 @@ void moment<P>::createFlist(PDE<P> const &pde, options const &opts)
 // Actually contstructs the moment vector using fList.
 // Calculate only if adapt is true or the vector field is empty
 template<typename P>
-void moment<P>::createMomentVector(parser const &opts,
+void moment<P>::createMomentVector(PDE<P> const &pde, parser const &opts,
                                    elements::table const &hash_table)
 {
   // check that fList has been constructed
@@ -45,17 +45,16 @@ void moment<P>::createMomentVector(parser const &opts,
   {
     distribution_plan const plan = get_plan(get_num_ranks(), hash_table);
     auto rank                    = get_rank();
-    auto tmp = combine_dimensions(opts.get_degree(), hash_table,
-                                  plan.at(rank).row_start,
+    int const degree             = pde.get_dimensions()[0].get_degree();
+    auto tmp = combine_dimensions(degree, hash_table, plan.at(rank).row_start,
                                   plan.at(rank).row_stop, this->fList[0]);
     this->vector.resize(tmp.size());
     this->vector      = std::move(tmp);
     auto num_md_funcs = md_funcs.size();
     for (std::size_t s = 1; s < num_md_funcs; ++s)
     {
-      tmp = combine_dimensions(opts.get_degree(), hash_table,
-                               plan.at(rank).row_start, plan.at(rank).row_stop,
-                               this->fList[s]);
+      tmp = combine_dimensions(degree, hash_table, plan.at(rank).row_start,
+                               plan.at(rank).row_stop, this->fList[s]);
       std::transform(tmp.begin(), tmp.end(), this->vector.begin(),
                      this->vector.begin(), std::plus<>{});
     }
