+++ src/pde/pde_vlasov_lb_full_f.hpp
@@ -104,7 +104,7 @@ private:
 
   /* Define the dimension */
   inline static dimension<P> const dim_0 = dimension<P>(
-      -1.0, 1.0, 8, default_degree,
+      -1.0, 1.0, 4, default_degree,
       {initial_condition_dim_x_0, initial_condition_dim_x_1}, dV, "x");
 
   inline static dimension<P> const dim_1 = dimension<P>(
