+++ src/moment_tests.cpp
@@ -90,7 +90,7 @@ TEMPLATE_TEST_CASE("CreateMomentReducedMatrix", "[moments]", double, float)
   for (size_t i = 0; i < moments.size(); ++i)
   {
     moments[i].createFlist(*pde, opts);
-    moments[i].createMomentVector(parse, check);
+    moments[i].createMomentVector(*pde, parse, check);
     moments[i].createMomentReducedMatrix(*pde, check);
 
     auto const gold_filename =
