+++ src/scalapack_matrix_info_tests.cpp
@@ -5,6 +5,8 @@
 
 #include <array>
 
+using namespace asgard;
+
 int main(int argc, char *argv[])
 {
   initialize_distribution();
