+++ src/quadrature.hpp
@@ -3,6 +3,8 @@
 #include "tensors.hpp"
 #include <array>
 
+namespace asgard
+{
 enum class legendre_normalization
 {
   unnormalized,
@@ -25,3 +27,4 @@ template<typename P>
 std::array<fk::vector<P>, 2>
 legendre_weights(int const degree, P const lower_bound, P const upper_bound,
                  bool const use_degree_points = false);
+} // namespace asgard
