+++ src/pde/pde_fokkerplanck1_4p3.hpp
@@ -12,6 +12,8 @@
 #include "../tensors.hpp"
 #include "pde_base.hpp"
 
+namespace asgard
+{
 // ---------------------------------------------------------------------------
 //
 // the "fokkerplanck 1d - problem 4.3" pde
@@ -167,3 +169,4 @@ private:
   inline static scalar_func<P> const exact_scalar_func_ =
       analytic_solution_time;
 };
+} // namespace asgard
