### 2022-11-16 754ea8d
Allow RHS to be a matrix
Signed-off-by: Steven Hahn <hahnse@ornl.gov>
