### 2022-11-17 adb4af5
Allow RHS to be a matrix (#466)
Signed-off-by: Steven Hahn <hahnse@ornl.gov>
