+++ src/moment_tests.cpp
@@ -6,6 +6,8 @@
 
 static auto const moment_base_dir = gold_base_dir / "moment";
 
+using namespace asgard;
+
 struct distribution_test_init
 {
   distribution_test_init() { initialize_distribution(); }
