+++ src/time_advance.hpp
@@ -53,7 +53,8 @@ template<typename P>
 fk::vector<P>
 imex_advance(PDE<P> const &pde, adapt::distributed_grid<P> const &adaptive_grid,
              basis::wavelet_transform<P, resource::host> const &transformer,
-             std::array<unscaled_bc_parts<P>, 2> const &unscaled_parts,
+             std::array<boundary_conditions::unscaled_bc_parts<P>, 2> const
+                 &unscaled_parts,
              fk::vector<P> const &x_orig, P const time, solve_opts const solver,
              bool const update_system = true);
 
