### 2022-11-18 b2e7f93
Add constructor overload for dimension class
