+++ src/pde/pde_fokkerplanck1_pitch_C.hpp
@@ -149,11 +149,11 @@ private:
 
   // define dimensions
   inline static dimension<P> const dim0_ =
-      dimension<P>(-1.0,                     // domain min
-                   1.0,                      // domain max
-                   2,                        // levels
-                   2,                        // degree
-                   {initial_condition_dim0}, // initial condition
+      dimension<P>(-1.0,                   // domain min
+                   1.0,                    // domain max
+                   2,                      // levels
+                   2,                      // degree
+                   initial_condition_dim0, // initial condition
                    volume_jacobian_dV,
                    "x"); // name
 
