### 2022-11-18 177a334
Helper to set initial conditions.
