### 2022-11-18 b058bcc
Modified to function to take dims instead of pde input.
