### 2022-11-18 db39010
updated the error checking and constructors
