### 2022-11-23 d82fef2
[wip] working on the initial conditions
