+++ src/adapt.cpp
@@ -104,6 +104,25 @@ static void update_levels(elements::table const &adapted_table, PDE<P> &pde,
     }
   }
 }
+template<typename P>
+static void update_levels(elements::table const &adapted_table, std::vector<dimension<P>> &dims,
+                          int const num_terms, std::vector<std::vector<term<P>>> &terms,
+                          bool const rechain = false)
+{
+  auto const new_levels =
+      get_levels(adapted_table, dims.size());
+  for (auto i = 0; i < static_cast<int>(new_levels.size()); ++i)
+  {
+    dims[i].set_level(new_levels[i]);
+    if (rechain)
+    {
+      for (auto j = 0; j < num_terms; ++j)
+      {
+        terms[j][i].rechain_coefficients(dims[i]);
+      }
+    }
+  }
+}
 
 template<typename P>
 distributed_grid<P>::distributed_grid(std::vector<dimension<P>> const &dims,
@@ -114,15 +133,85 @@ distributed_grid<P>::distributed_grid(std::vector<dimension<P>> const &dims,
 }
 
 // FIXME assumes uniform degree across levels
+// template<typename P>
+// fk::vector<P> distributed_grid<P>::get_initial_condition(
+//     PDE<P> &pde, basis::wavelet_transform<P, resource::host> const &transformer,
+//     options const &cli_opts)
+// {
+//   // get unrefined condition
+//   std::vector<vector_func<P>> v_functions;
+//   for (auto const &dim : pde.get_dimensions())
+//   {
+//     v_functions.push_back(dim.initial_condition);
+//   }
+//   P const time             = 0;
+//   auto const initial_unref = [this, &v_functions, &pde, &transformer, time]() {
+//     auto const subgrid = this->get_subgrid(get_rank());
+//     // TODO temp add scalar time func to initial conditions with multi-D func PR
+//     auto const mult =
+//         pde.has_analytic_soln ? pde.exact_time(time) : static_cast<P>(1.0);
+//     return transform_and_combine_dimensions(
+//         pde, v_functions, this->get_table(), transformer, subgrid.col_start,
+//         subgrid.col_stop, pde.get_dimensions()[0].get_degree(), time, mult);
+//   }();
+//
+//   if (!cli_opts.do_adapt_levels)
+//   {
+//     return initial_unref;
+//   }
+//
+//   // refine
+//   fk::vector<P> refine_y(initial_unref);
+//   auto refining = true;
+//   while (refining)
+//   {
+//     auto const old_y   = fk::vector<P>(refine_y);
+//     auto const refined = this->refine(old_y, cli_opts);
+//     refining           = old_y.size() != refined.size();
+//     update_levels(this->get_table(), pde);
+//
+//     // reproject
+//     auto const reprojected = [this, &v_functions, &pde, &transformer, time]() {
+//       auto const subgrid = this->get_subgrid(get_rank());
+//       auto const mult =
+//           pde.has_analytic_soln ? pde.exact_time(time) : static_cast<P>(1.0);
+//       return transform_and_combine_dimensions(
+//           pde, v_functions, this->get_table(), transformer, subgrid.col_start,
+//           subgrid.col_stop, pde.get_dimensions()[0].get_degree(), time, mult);
+//     }();
+//     refine_y.resize(reprojected.size()) = reprojected;
+//   }
+//
+//   // coarsen
+//   auto const coarse_y = this->coarsen(refine_y, cli_opts);
+//   update_levels(this->get_table(), pde);
+//
+//   // reproject
+//   auto const adapted_y = [this, &v_functions, &pde, &transformer, time]() {
+//     auto const subgrid = this->get_subgrid(get_rank());
+//     auto const mult =
[…]
