### 2022-11-29 3891baa
Fix -Wpedantic error on asgard namespace
