+++ src/transformations.cpp
@@ -90,13 +90,15 @@ recursive_kron(std::vector<fk::matrix<P, mem_type::view>> &kron_matrices,
 template<typename P>
 std::vector<fk::matrix<P>> gen_realspace_transform(
     PDE<P> const &pde,
-    basis::wavelet_transform<P, resource::host> const &transformer)
+    basis::wavelet_transform<P, resource::host> const &transformer,
+    int const num_dims)
 {
   /* contains a basis matrix for each dimension */
   std::vector<fk::matrix<P>> real_space_transform;
-  real_space_transform.reserve(pde.num_dims);
+  expect(pde.num_dims >= num_dims);
+  real_space_transform.reserve(num_dims);
 
-  for (int i = 0; i < pde.num_dims; i++)
+  for (int i = 0; i < num_dims; i++)
   {
     /* get the ith dimension */
     dimension<P> const &d    = pde.get_dimensions()[i];
@@ -129,27 +131,74 @@ std::vector<fk::matrix<P>> gen_realspace_transform(
 }
 
 template<typename P>
+fk::vector<P>
+gen_realspace_nodes(int const degree, int const level, P const min, P const max)
+{
+  int const n        = pow(2, level);
+  int const mat_dims = degree * n;
+  P const h          = (max - min) / n;
+  auto const lgwt    = legendre_weights(degree, -1.0, 1.0, true);
+  auto const roots   = lgwt[0];
+
+  unsigned int const dof = roots.size();
+
+  fk::vector<P> nodes(mat_dims);
+  for (int i = 0; i < n; i++)
+  {
+    auto p_val = legendre(roots, degree, legendre_normalization::lin);
+
+    p_val[0] = p_val[0] * sqrt(1.0 / h);
+
+    std::vector<P> xi(dof);
+    for (std::size_t j = 0; j < dof; j++)
+    {
+      xi[j] = (0.5 * (roots(j) + 1.0) + i) * h + min;
+    }
+
+    std::vector<int> Iu(degree);
+    for (int j = 0, je = degree - 1; j < je; j++)
+    {
+      Iu[j] = dof * i + j + 1;
+    }
+    Iu[degree - 1] = dof * (i + 1);
+
+    for (std::size_t j = 0; j < dof; j++)
+    {
+      expect(j <= Iu.size());
+      nodes(Iu[j] - 1) = xi[j];
+    }
+  }
+
+  return nodes;
+}
+
+template<typename P>
 void wavelet_to_realspace(
     PDE<P> const &pde, fk::vector<P> const &wave_space,
     elements::table const &table,
     basis::wavelet_transform<P, resource::host> const &transformer,
     int const memory_limit_MB,
     std::array<fk::vector<P, mem_type::view, resource::host>, 2> &workspace,
-    fk::vector<P> &real_space)
+    fk::vector<P> &real_space, int const transform_dims)
 {
   expect(memory_limit_MB > 0);
 
   std::vector<batch_chain<P, resource::host>> chain;
 
+  // generate transform for each dimension, or only for up to transform_dims
+  // dimensions
+  int const actual_num_dims =
+      transform_dims > 0 ? transform_dims : pde.num_dims;
+
   /* generate the wavelet-to-real-space transformation matrices for each
    * dimension */
   std::vector<fk::matrix<P>> real_space_transform =
-      gen_realspace_transform(pde, transformer);
+      gen_realspace_transform(pde, transformer, actual_num_dims);
 
   // FIXME Assume the degree in the first dimension is equal across all the
   // remaining dimensions
   auto const stride =
-      std::pow(pde.get_dimensions()[0].get_degree(), pde.num_dims);
[…]
