+++ src/time_advance.cpp
@@ -2,6 +2,7 @@
 #include "adapt.hpp"
 #include "batch.hpp"
 #include "boundary_conditions.hpp"
+#include "coefficients.hpp"
 #include "distribution.hpp"
 #include "elements.hpp"
 #include "fast_math.hpp"
@@ -57,9 +58,14 @@ adaptive_advance(method const step_method, PDE<P> &pde,
                ? explicit_advance(pde, adaptive_grid, transformer, program_opts,
                                   unscaled_parts, x_orig, workspace_size_MB,
                                   time)
-               : implicit_advance(pde, adaptive_grid, transformer,
-                                  unscaled_parts, x_orig, time,
-                                  program_opts.solver, update_system);
+               : ((step_method == method::imp)
+                      ? implicit_advance(pde, adaptive_grid, transformer,
+                                         unscaled_parts, x_orig, time,
+                                         program_opts.solver, update_system)
+                      : imex_advance(pde, adaptive_grid, transformer,
+                                     program_opts, unscaled_parts, x_orig,
+                                     workspace_size_MB, time,
+                                     program_opts.solver, update_system));
   }
 
   // coarsen
@@ -83,8 +89,13 @@ adaptive_advance(method const step_method, PDE<P> &pde,
         (step_method == method::exp)
             ? explicit_advance(pde, adaptive_grid, transformer, program_opts,
                                unscaled_parts, y, workspace_size_MB, time)
-            : implicit_advance(pde, adaptive_grid, transformer, unscaled_parts,
-                               y, time, program_opts.solver, refining);
+            : ((step_method == method::imp)
+                   ? implicit_advance(pde, adaptive_grid, transformer,
+                                      unscaled_parts, y, time,
+                                      program_opts.solver, refining)
+                   : imex_advance(pde, adaptive_grid, transformer, program_opts,
+                                  unscaled_parts, y, workspace_size_MB, time,
+                                  program_opts.solver, refining));
 
     auto const old_plan = adaptive_grid.get_distrib_plan();
     old_size            = adaptive_grid.size();
@@ -392,13 +403,17 @@ implicit_advance(PDE<P> const &pde,
 // current solution vector x. on exit, the next solution vector is stored in fx.
 template<typename P>
 fk::vector<P>
-imex_advance(PDE<P> const &pde, adapt::distributed_grid<P> const &adaptive_grid,
+imex_advance(PDE<P> &pde, adapt::distributed_grid<P> const &adaptive_grid,
              basis::wavelet_transform<P, resource::host> const &transformer,
+             options const &program_opts,
              std::array<boundary_conditions::unscaled_bc_parts<P>, 2> const
                  &unscaled_parts,
-             fk::vector<P> const &x_orig, P const time, solve_opts const solver,
-             bool const update_system)
+             fk::vector<P> const &x_orig, int const workspace_size_MB,
+             P const time, solve_opts const solver, bool const update_system)
 {
+  ignore(unscaled_parts);
+  ignore(solver);
+
   // BEFE = 0 case
   expect(time >= 0);
 
[…]
