### 2022-11-29 a773b8f
Fix issues with term selection option
