### 2022-11-29 d83d02e
[wip] fixing the fields to match dimension requirements
