+++ src/asgard_field.hpp
@@ -18,10 +18,13 @@
 namespace asgard
 {
 
+/*!
+ * \brief Copies the dimensions and modifies them based on the cli-parameters
+ */
 template<typename precision>
 std::vector<dimension_description<precision>>
 cli_apply_level_degree_correction(parser const &cli_input,
-                                  std::vector<dimension_description<precision>> const dimensions)
+                                  std::vector<dimension_description<precision>> const &dimensions)
 {
   size_t num_dims = dimensions.size();
   std::vector<int> levels(dimensions.size()), degrees(dimensions.size());
@@ -93,7 +96,7 @@ inline void verify_unique_strings(std::vector<std::string> const &names) {
 
 template<typename precision>
 struct dimension_set {
-  dimension_set(parser const &cli_input, std::vector<dimension_description<precision>> const dimensions)
+  dimension_set(parser const &cli_input, std::vector<dimension_description<precision>> const &dimensions)
     : list(cli_apply_level_degree_correction(cli_input, dimensions))
   {
     std::vector<std::string> names(list.size());
@@ -123,35 +126,54 @@ struct field_description
   field_description(std::string const &dimension_name,
                     vector_func<precision> const initial_condition,
                     vector_func<precision> const exact_solution,
-                    g_func_type<precision> const volume_jacobian_dV_in, // MIRO: maybe this should be part of the dimension_description
                     std::string const &field_name
                     )
-    : field_description(std::vector<std::string>{dimension_name}, {initial_condition}, {exact_solution}, {volume_jacobian_dV_in}, field_name)
+    : field_description(std::vector<std::string>{dimension_name}, {initial_condition}, {exact_solution}, field_name)
+    {}
+
+  field_description(std::string const &dimension_name,
+                    vector_func<precision> const initial_condition,
+                    std::string const &field_name
+                    )
+    : field_description(std::vector<std::string>{dimension_name}, {initial_condition}, field_name)
     {}
 
   field_description(std::vector<std::string> const &dimension_names,
                     std::vector<vector_func<precision>> const &initial_conditions,
                     std::vector<vector_func<precision>> const &exact_solution,
-                    std::vector<g_func_type<precision>> const &volume_jacobian_dV_in, // MIRO: maybe this should be part of the dimension_description
                     std::string const &field_name
                     )
       // the const-ref constructor copies and calls the r-value constructor
       : field_description(std::vector<std::string>(dimension_names),
                           std::vector<vector_func<precision>>(initial_conditions),
                           std::vector<vector_func<precision>>(exact_solution),
-                          std::vector<g_func_type<precision>>(volume_jacobian_dV_in),
+                          std::string(field_name))
+  {}
+    field_description(std::vector<std::string> const &dimension_names,
+                      std::vector<vector_func<precision>> const &initial_conditions,
+                      std::string const &field_name
+                      )
+      // the const-ref constructor copies and calls the r-value constructor
+      : field_description(std::vector<std::string>(dimension_names),
+                          std::vector<vector_func<precision>>(initial_conditions),
                           std::string(field_name))
   {}
 
   field_description(std::vector<std::string> &&dimensions,
                     std::vector<vector_func<precision>> &&initial_conditions,
+                    std::string &&field_name
+                    )
+  // I think I should use std::forward here
+      : field_description(std::move(dimensions), std::move(initial_conditions), {}, std::move(field_name))
+  {}
[…]
