+++ src/asgard_discretization_tests.cpp
@@ -1,0 +1,10 @@
+#include "asgard_discretization.hpp"
+
+using namespace asgard;
+
+
+//int main(int argc, char *argv[])
+int main(int, char**)
+{
+  return 0;
+}

[…]
