### 2022-11-30 fd8922c
[wip] adding discretization class
