### 2022-12-01 0cf2783
Remove unused variant of build_system_matrix
Signed-off-by: Steven Hahn <hahnse@ornl.gov>
