+++ src/batch.cpp
@@ -621,102 +621,6 @@ void build_system_matrix(PDE<P> const &pde, elements::table const &elem_table,
   }
 }
 
-// function to allocate and build implicit system.
-// given a problem instance (pde/elem table)
-// does not utilize batching, here because it shares underlying structure and
-// routines with explicit time advance
-template<typename P>
-void build_system_matrix(PDE<P> const &pde, elements::table const &elem_table,
-                         fk::matrix<P> &A)
-{
-  // assume uniform degree for now
-  int const degree    = pde.get_dimensions()[0].get_degree();
-  int const elem_size = static_cast<int>(std::pow(degree, pde.num_dims));
-  int const A_size    = elem_size * elem_table.size();
-
-  expect(A.ncols() == A_size && A.nrows() == A_size);
-
-  using key_type = std::pair<int, int>;
-  using val_type = fk::matrix<P, mem_type::owner, resource::host>;
-  std::map<key_type, val_type> coef_cache;
-
-  expect(A.ncols() == A_size && A.nrows() == A_size);
-
-  // copy coefficients to host for subsequent use
-  for (int k = 0; k < pde.num_terms; ++k)
-  {
-    for (int d = 0; d < pde.num_dims; d++)
-    {
-      coef_cache.emplace(key_type(k, d),
-                         pde.get_coefficients(k, d).clone_onto_host());
-    }
-  }
-
-  // loop over elements
-  for (auto i = 0; i < elem_table.size(); ++i)
-  {
-    // first, get linearized indices for this element
-    //
-    // calculate from the level/cell indices for each
-    // dimension
-    fk::vector<int> const coords = elem_table.get_coords(i);
-    expect(coords.size() == pde.num_dims * 2);
-    fk::vector<int> const elem_indices = linearize(coords);
-
-    int const global_row = i * elem_size;
-
-    // calculate the row portion of the
-    // operator position used for this
-    // element's gemm calls
-    fk::vector<int> const operator_row =
-        linear_coords_to_indices(pde, degree, elem_indices);
-
-    // loop over connected elements. for now, we assume
-    // full connectivity
-    for (int j = 0; j < elem_table.size(); ++j)
-    {
-      // get linearized indices for this connected element
-      fk::vector<int> const coords_nD = elem_table.get_coords(j);
-      expect(coords_nD.size() == pde.num_dims * 2);
-      fk::vector<int> const connected_indices = linearize(coords_nD);
-
-      // calculate the col portion of the
-      // operator position used for this
-      // element's gemm calls
-      fk::vector<int> const operator_col =
-          linear_coords_to_indices(pde, degree, connected_indices);
-
-      for (int k = 0; k < pde.num_terms; ++k)
-      {
-        std::vector<fk::matrix<P>> kron_vals;
-        fk::matrix<P> kron0(1, 1);
-        kron0(0, 0) = 1.0;
-        kron_vals.push_back(kron0);
-        for (int d = 0; d < pde.num_dims; d++)
-        {
-          fk::matrix<P, mem_type::view> op_view = fk::matrix<P, mem_type::view>(
-              coef_cache[key_type(k, d)], operator_row(d),
-              operator_row(d) + degree - 1, operator_col(d),
-              operator_col(d) + degree - 1);
-          fk::matrix<P> k_new = kron_vals[d].kron(op_view);
-          kron_vals.push_back(k_new);
-        }
-
-        // calculate the position of this element in the
-        // global system matrix
-        int const global_col = j * elem_size;
-        auto const &k_tmp    = kron_vals.back();
-
-        fk::matrix<P, mem_type::view> A_view(
[…]
