### 2022-12-01 75531fd
Remove unused variant of build_system_matrix (#471)
Signed-off-by: Steven Hahn <hahnse@ornl.gov>
