### 2022-12-02 8d1ff02
[wip] getting the initial conditions
