+++ src/asgard_field.hpp
@@ -120,58 +120,67 @@ struct dimension_set {
 };
 
 
+enum class field_mode {
+  evolution, conservation
+};
+
 template<typename precision>
 struct field_description
 {
-  field_description(std::string const &dimension_name,
+  field_description(field_mode fmode,
+                    std::string const &dimension_name,
                     vector_func<precision> const initial_condition,
                     vector_func<precision> const exact_solution,
                     std::string const &field_name
                     )
-    : field_description(std::vector<std::string>{dimension_name}, {initial_condition}, {exact_solution}, field_name)
+    : field_description(fmode, std::vector<std::string>{dimension_name}, {initial_condition}, {exact_solution}, field_name)
     {}
 
-  field_description(std::string const &dimension_name,
+  field_description(field_mode fmode,
+                    std::string const &dimension_name,
                     vector_func<precision> const initial_condition,
                     std::string const &field_name
                     )
-    : field_description(std::vector<std::string>{dimension_name}, {initial_condition}, field_name)
+    : field_description(fmode, std::vector<std::string>{dimension_name}, {initial_condition}, field_name)
     {}
 
-  field_description(std::vector<std::string> const &dimension_names,
+  field_description(field_mode fmode,
+                    std::vector<std::string> const &dimension_names,
                     std::vector<vector_func<precision>> const &initial_conditions,
                     std::vector<vector_func<precision>> const &exact_solution,
                     std::string const &field_name
                     )
-      // the const-ref constructor copies and calls the r-value constructor
-      : field_description(std::vector<std::string>(dimension_names),
+      : field_description(fmode,
+                          std::vector<std::string>(dimension_names),
                           std::vector<vector_func<precision>>(initial_conditions),
                           std::vector<vector_func<precision>>(exact_solution),
                           std::string(field_name))
   {}
-    field_description(std::vector<std::string> const &dimension_names,
+    field_description(field_mode fmode,
+                      std::vector<std::string> const &dimension_names,
                       std::vector<vector_func<precision>> const &initial_conditions,
                       std::string const &field_name
                       )
-      // the const-ref constructor copies and calls the r-value constructor
-      : field_description(std::vector<std::string>(dimension_names),
+      : field_description(fmode,
+                          std::vector<std::string>(dimension_names),
                           std::vector<vector_func<precision>>(initial_conditions),
                           std::string(field_name))
   {}
 
-  field_description(std::vector<std::string> &&dimensions,
+  field_description(field_mode fmode,
+                    std::vector<std::string> &&dimensions,
                     std::vector<vector_func<precision>> &&initial_conditions,
                     std::string &&field_name
                     )
-  // I think I should use std::forward here
-      : field_description(std::move(dimensions), std::move(initial_conditions), {}, std::move(field_name))
+      : field_description(fmode, std::move(dimensions), std::move(initial_conditions), {}, std::move(field_name))
   {}
-  field_description(std::vector<std::string> &&dimensions,
+  field_description(field_mode fmode,
+                    std::vector<std::string> &&dimensions,
                     std::vector<vector_func<precision>> &&initial_conditions,
                     std::vector<vector_func<precision>> &&exact_solution,
                     std::string &&field_name
                     )
-      : d_names(std::move(dimensions)),
+      : mode(fmode), d_names(std::move(dimensions)),
         init_cond(std::move(initial_conditions)), exact(std::move(exact_solution)),
         name(std::move(field_name))
   {
@@ -204,65 +213,12 @@ struct field_description
[…]
