### 2022-12-03 1678bef
[wip] initial conditions, getting a vector
