+++ src/basis.hpp
@@ -34,8 +34,13 @@ template<typename P, resource resrc>
 class wavelet_transform
 {
 public:
-  wavelet_transform(options const &program_opts, PDE<P> const &pde,
+  wavelet_transform(options const &program_opts, int const max_degree,
                     bool const quiet = true);
+
+  wavelet_transform(options const &program_opts, PDE<P> const &pde,
+                    bool const quiet = true)
+    : wavelet_transform(program_opts, pde.get_dimensions()[0].get_degree(), quiet)
+  {}
 
   // apply the fmwt matrix to coefficients
   template<mem_type omem>
