+++ src/pde/pde_base.hpp
@@ -11,6 +11,7 @@
 #include <typeinfo>
 #include <vector>
 
+#include "../asgard_dimension.hpp"
 #include "../fast_math.hpp"
 #include "../matlab_utilities.hpp"
 #include "../moment.hpp"
@@ -27,20 +28,6 @@ namespace asgard
 // FIXME we plan a major rework of this component in the future
 // for RAII compliance and readability
 
-// same pi used by matlab
-static constexpr double const PI = 3.141592653589793;
-
-// for passing around vector/scalar-valued functions used by the PDE
-template<typename P>
-using vector_func = std::function<fk::vector<P>(fk::vector<P> const, P const)>;
-template<typename P>
-using scalar_func = std::function<P(P const)>;
-
-template<typename P>
-using g_func_type = std::function<P(P const, P const)>;
-
-template<typename P>
-using md_func_type = std::vector<vector_func<P>>;
 //----------------------------------------------------------------------------
 //
 // Define member classes of the PDE type: dimension, term, source
@@ -77,111 +64,9 @@ auto const element_segment_size = [](auto const &pde) {
 template<typename P>
 class PDE;
 
-/*!
- * \brief Contains the user provided description of a dimension.
- *
- * The \b dimension_name will be used to identify when the dimension is used for each field.
- */
 template<typename P>
-struct dimension_description {
-  dimension_description(P const domain_min, P const domain_max,
-                        int const grid_level, int const basis_degree,
-                        std::string const dimension_name,
-                        g_func_type<P> const volume_jacobian_dV = [](P const, P const)->P{ return 1.0; }) :
-    d_min(domain_min), d_max(domain_max),
-    level(grid_level), degree(basis_degree),
-    name(dimension_name),
-    jacobian(volume_jacobian_dV)
-  {
-    // note that the constructor that is user facing will use the most descriptive and clear names
-    //      also the longest names for each variable/parameter
-    //      the internal variables will use short-hand names
-    expect(d_max > d_min);
-    expect(level > 0);
-    expect(degree > 1); // must fix this one, degree should count from zero
-    expect(name.length() > 0);
-  }
-  P const d_min;
-  P const d_max;
-  int const level;
-  int const degree;
-  std::string const name;
-  g_func_type<P> const jacobian;
-};
-
 class moment;
 
-// Note (from Miro): there is nothing inside this class that requires internal consistency of the input data
-// Why have private members when you can freely modify every one of them?
-// Make a struct to keep related data together and have simple direct access to the data.
-template<typename P>
-struct dimension
-{
-  P const domain_min;
-  P const domain_max;
-  std::vector<vector_func<P>> const initial_condition;
-  g_func_type<P> const volume_jacobian_dV;
-  std::string const name;
-  dimension(P const d_min, P const d_max, int const level, int const degree,
-            vector_func<P> const initial_condition_in,
-            g_func_type<P> const volume_jacobian_dV_in,
-            std::string const name_in)
-
-      : dimension(d_min, d_max, level, degree,
-                  std::vector<vector_func<P>>({initial_condition_in}),
-                  volume_jacobian_dV_in, name_in)
-  {}
-
-  dimension(P const d_min, P const d_max, int const level, int const degree,
-            std::vector<vector_func<P>> const initial_condition_in,
-            g_func_type<P> const volume_jacobian_dV_in,
-            std::string const name_in)
-
-      : domain_min(d_min), domain_max(d_max),
[…]
