### 2022-12-06 938cc3e
[added methods that work on view]
