### 2022-12-07 c7f58d1
global state
