### 2022-12-09 298b387
Convert from wavelet to realspace for plotting
Signed-off-by: Steven Hahn <hahnse@ornl.gov>
