+++ src/matlab_plot.cpp
@@ -1,5 +1,6 @@
 #include "matlab_plot.hpp"
 
+#include "asgard_dimension.hpp"
 #include "quadrature.hpp"
 #include <cmath>
 #include <type_traits>
@@ -236,15 +237,23 @@ template<typename P>
 fk::vector<P> matlab_plot::gen_elem_coords(PDE<P> const &pde,
                                            elements::table const &table) const
 {
-  int const ndims = pde.num_dims;
+  return gen_elem_coords(pde.get_dimensions(), table);
+}
+
+template<typename P>
+fk::vector<P>
+matlab_plot::gen_elem_coords(std::vector<dimension<P>> const &dims,
+                             elements::table const &table) const
+{
+  int const ndims = dims.size();
 
   fk::vector<P> center_coords(ndims * table.size());
 
   // Iterate over dimensions first since matlab needs col-major order
   for (int d = 0; d < ndims; d++)
   {
-    P const max = pde.get_dimensions()[d].domain_max;
-    P const min = pde.get_dimensions()[d].domain_min;
+    P const max = dims[d].domain_max;
+    P const min = dims[d].domain_min;
     P const rng = max - min;
 
     for (int i = 0; i < table.size(); i++)
@@ -277,20 +286,74 @@ fk::vector<P> matlab_plot::gen_elem_coords(PDE<P> const &pde,
       center_coords(d * table.size() + i) = x;
     }
   }
+  return center_coords;
+}
 
+template<typename P>
+fk::vector<P>
+matlab_plot::gen_elem_coords(std::vector<dimension_description<P>> const &dims,
+                             elements::table const &table) const
+{
+  int const ndims = dims.size();
+
+  fk::vector<P> center_coords(ndims * table.size());
+
+  // Iterate over dimensions first since matlab needs col-major order
+  for (int d = 0; d < ndims; d++)
+  {
+    P const max = dims[d].d_max;
+    P const min = dims[d].d_min;
+    P const rng = max - min;
+
+    for (int i = 0; i < table.size(); i++)
+    {
+      fk::vector<int> const &coords = table.get_coords(i);
+
+      int const lev = coords(d);
+      int const pos = coords(ndims + d);
+
+      expect(lev >= 0);
+      expect(pos >= 0);
+
+      P x0;
+      if (lev > 1)
+      {
+        P const s = pow(2, lev - 1) - 1.0;
+        P const h = 1.0 / (pow(2, lev - 1));
+        P const w = 1.0 - h;
+        P const o = 0.5 * h;
+
+        x0 = pos / s * w + o;
+      }
+      else
+      {
+        x0 = pos + 0.5;
+      }
+
+      P const x = x0 * rng + min;
+
+      center_coords(d * table.size() + i) = x;
+    }
+  }
   return center_coords;
 }
 
 template<typename P>
 void matlab_plot::init_plotting(PDE<P> const &pde, elements::table const &table)
 {
+  init_plotting(pde.get_dimensions(), table);
+}
+
+template<typename P>
+void matlab_plot::init_plotting(std::vector<dimension<P>> const &dims,
+                                elements::table const &table)
+{
   // Generates cell array of nodes and element coordinates needed for plotting
-  sol_sizes_ = get_soln_sizes(pde);
+  sol_sizes_ = get_soln_sizes(dims);
 
-  nodes_          = std::vector<matlab::data::Array>(pde.num_dims);
-  auto const dims = pde.get_dimensions();
+  nodes_ = std::vector<matlab::data::Array>(dims.size());
 
-  for (int i = 0; i < pde.num_dims; i++)
+  for (int i = 0; i < dims.size(); i++)
   {
     auto const &dim       = dims[i];
[…]
