+++ src/transformations.cpp
@@ -85,21 +85,29 @@ recursive_kron(std::vector<fk::matrix<P, mem_type::view>> &kron_matrices,
   }
 }
 
-/* given a pde, for each dimension create a matrix where the columns are
-   legendre basis functions evaluated at the roots */
 template<typename P>
 std::vector<fk::matrix<P>> gen_realspace_transform(
     PDE<P> const &pde,
     basis::wavelet_transform<P, resource::host> const &transformer)
 {
+  return gen_realspace_transform(pde.get_dimensions(), transformer);
+}
+
+/* given a pde, for each dimension create a matrix where the columns are
+   legendre basis functions evaluated at the roots */
+template<typename P>
+std::vector<fk::matrix<P>> gen_realspace_transform(
+    std::vector<dimension<P>> const &dims,
+    basis::wavelet_transform<P, resource::host> const &transformer)
+{
   /* contains a basis matrix for each dimension */
   std::vector<fk::matrix<P>> real_space_transform;
-  real_space_transform.reserve(pde.num_dims);
+  real_space_transform.reserve(dims.size());
 
-  for (int i = 0; i < pde.num_dims; i++)
+  for (int i = 0; i < dims.size(); i++)
   {
     /* get the ith dimension */
-    dimension<P> const &d    = pde.get_dimensions()[i];
+    dimension<P> const &d    = dims[i];
     int const level          = d.get_level();
     int const n_segments     = fm::two_raised_to(level);
     int const deg_freedom_1d = d.get_degree() * n_segments;
@@ -171,8 +179,61 @@ gen_realspace_nodes(int const degree, int const level, P const min, P const max)
 }
 
 template<typename P>
+std::vector<fk::matrix<P>> gen_realspace_transform(
+    std::vector<dimension_description<P>> const &dims,
+    basis::wavelet_transform<P, resource::host> const &transformer)
+{
+  /* contains a basis matrix for each dimension */
+  std::vector<fk::matrix<P>> real_space_transform;
+  real_space_transform.reserve(dims.size());
+
+  for (int i = 0; i < dims.size(); i++)
+  {
+    /* get the ith dimension */
+    dimension_description<P> const &d = dims[i];
+    int const level                   = d.level;
+    int const n_segments              = fm::two_raised_to(level);
+    int const deg_freedom_1d          = d.degree * n_segments;
+    P const normalize                 = (d.d_max - d.d_min) / n_segments;
+    fk::matrix<P> dimension_transform(deg_freedom_1d, deg_freedom_1d);
+    /* create matrix of Legendre polynomial basis functions evaluated at the
+     * roots */
+    bool const use_degree_quad = true;
+    auto const roots = legendre_weights<P>(d.degree, -1, 1, use_degree_quad)[0];
+    /* normalized legendre transformation matrix. Column i is legendre
+       polynomial of degree i. element (i, j) is polynomial evaluated at jth
+       root of the highest degree polynomial */
+    fk::matrix<P> const basis = legendre<P>(roots, d.degree)[0] *
+                                (static_cast<P>(1.0) / std::sqrt(normalize));
+    /* set submatrices of dimension_transform */
+    for (int j = 0; j < n_segments; j++)
+    {
+      int const diagonal_pos = d.degree * j;
+      dimension_transform.set_submatrix(diagonal_pos, diagonal_pos, basis);
+    }
+    real_space_transform.push_back(transformer.apply(dimension_transform, level,
+                                                     basis::side::right,
+                                                     basis::transpose::trans));
+  }
+  return real_space_transform;
+}
+
+template<typename P>
 void wavelet_to_realspace(
[…]

[…]
