### 2022-12-09 6e8500c
pde system with initial conditions
