+++ src/asgard_pde_system.hpp
@@ -7,7 +7,7 @@ namespace asgard
 
 template<typename precision>
 std::vector<field<precision>>
-make_fields(std::vector<field_description<precision>> const &descriptions)
+make_fields_vector(std::vector<field_description<precision>> const &descriptions)
 {
   std::vector<field<precision>> result;
   result.reserve(descriptions.size());
@@ -23,7 +23,8 @@ public:
              std::vector<dimension_description<precision>> const &dimensions,
              std::vector<field_description<precision>> const &field_terms
              )
-    : cli(cli_input), dims(cli, dimensions), fields(make_fields(field_terms)),
+    : cli(cli_input), dims(cli, dimensions),
+      fields(make_fields_vector(field_terms)),
       transformer(cli_input, dimensions.front().degree, true)
   {
     static_assert(std::is_same<precision, float>::value
@@ -39,12 +40,14 @@ public:
     for(size_t i=0; i<field_names.size(); i++) field_names[i] = fields[i].name;
     verify_unique_strings(field_names);
 
+    // eventually we will give the user a more fine-grained control to the discretization
+    // this method is the default, i.e., use default discretization
+    // currently, we only allow default discretization
     finalizes_discretization();
   }
 
   void finalizes_discretization()
   {
-    // TODO: move the automatic discretization into a method that finalizes discretizations
     for(size_t i=0; i<fields.size(); i++)
     {
       // for each field, see if we already have discretization for this field
@@ -76,7 +79,17 @@ public:
       total_size += size;
     }
 
-    state = fk::vector<precision>(total_size);
+    state.resize(total_size);
+  }
+
+  void load_initial_conditions()
+  {
+    for(auto const &f : fields) {
+      grids[f.grid_index].get_initial_conditions(
+        f,
+        fk::vector<precision, mem_type::view>(state, f.global_begin, f.global_end-1)
+        );
+    }
   }
 
   fk::vector<precision, mem_type::const_view>
@@ -85,8 +98,15 @@ public:
     size_t findex = 0;
     while(findex < fields.size() && fields[findex].name != name)
       findex++;
+
+    auto f = std::find_if(fields.cbegin(), fields.cend(),
+                          [&](field<precision> const &candidate)->bool{ return (candidate.name == name); });
+
+    if (f == fields.end())
+      throw std::runtime_error("field name '" + name + "' is not unrecognized");
+
     return fk::vector<precision, mem_type::const_view>
-      (state, fields[findex].global_begin, fields[findex].global_end-1);
+      (state, f->global_begin, f->global_end-1);
   }
 
   // TODO: will work with operators similar to field
