+++ src/asgard_discretization.hpp
@@ -12,7 +12,7 @@ struct field_discretization {
                        dimension_set<precision> const &d_set,
                        asgard::basis::wavelet_transform<precision, resrc> &wavelet_transformer,
                        std::vector<std::string> const &d_names)
-    : cli(cli_input), transformer(wavelet_transformer), num_dof(0)
+    : cli(cli_input), transformer(wavelet_transformer), state_size(0)
   {
     dims.reserve(d_names.size());
     for(size_t i=0; i<d_names.size(); i++)
@@ -20,7 +20,7 @@ struct field_discretization {
 
     grid = std::make_unique<adapt::distributed_grid<precision>>(cli_input, dims);
     auto const subgrid = grid->get_subgrid(get_rank());
-    num_dof = (subgrid.col_stop - subgrid.col_start + 1) * std::pow(dims.front().degree_, dims.size());
+    state_size = (subgrid.col_stop - subgrid.col_start + 1) * std::pow(dims.front().degree_, dims.size());
   }
 
   bool can_discretize(field<precision> const &field) {
@@ -33,16 +33,16 @@ struct field_discretization {
     return true;
   }
 
-  int64_t size() const { return num_dof; }
+  int64_t size() const { return state_size; }
 
   void get_initial_conditions(field<precision> const &field, fk::vector<precision, mem_type::view> result)
   {
-    expect(result.size() == size());
+    expect(result.size() == state_size);
     grid->get_initial_condition(cli, dims, field.init_cond, 1.0, transformer, result);
   }
   fk::vector<precision> get_initial_conditions(field_description<precision> const &field)
   {
-    fk::vector<precision> result(num_dof);
+    fk::vector<precision> result(state_size);
     get_initial_conditions(field, fk::vector<precision, mem_type::view>(result));
     return result;
   }
@@ -52,7 +52,7 @@ struct field_discretization {
   asgard::basis::wavelet_transform<precision, resrc> &transformer;
 
   std::unique_ptr<adapt::distributed_grid<precision>> grid;
-  int64_t num_dof;
+  int64_t state_size;
 
   std::vector<dimension<precision>> dims;
 
