### 2022-12-09 eddf74e
fixed testing and removed commented out code
