### 2022-12-09 a083dec
fix clang compiler issues
