+++ src/elements.cpp
@@ -360,7 +360,7 @@ fk::matrix<int> table::get_cell_index_set(fk::vector<int> const &level_tuple)
   return cell_index_set;
 }
 
-template table::table(options const &program_opts, PDE<float> const &pde);
-template table::table(options const &program_opts, PDE<double> const &pde);
+template table::table(options const &opts, std::vector<dimension<float>> const &dims);
+template table::table(options const &opts, std::vector<dimension<double>> const &dims);
 
 } // namespace asgard::elements
