### 2022-12-13 979861b
Remove interactive matlab test
