### 2022-12-13 8ca8bd6
Use switch statement for time method selection
