+++ src/pde/pde_vlasov_lb_full_f.hpp
@@ -271,11 +271,11 @@ private:
 
   inline static term<P> const term_i1x = term<P>(false,  // time-dependent
                                                  "I1_x", // name
-                                                 {i1_pterm_x});
+                                                 {i1_pterm_x}, imex_flag::imp);
 
   inline static term<P> const term_i1v = term<P>(false,  // time-dependent
                                                  "I1_v", // name
-                                                 {i1_pterm_v});
+                                                 {i1_pterm_v}, imex_flag::imp);
 
   inline static std::vector<term<P>> const terms_3 = {term_i1x, term_i1v};
 
@@ -308,11 +308,11 @@ private:
 
   inline static term<P> const term_i2x = term<P>(false,  // time-dependent
                                                  "I2_x", // name
-                                                 {i2_pterm_x});
+                                                 {i2_pterm_x}, imex_flag::imp);
 
   inline static term<P> const term_i2v = term<P>(false,  // time-dependent
                                                  "I2_v", // name
-                                                 {i2_pterm_v});
+                                                 {i2_pterm_v}, imex_flag::imp);
 
   inline static std::vector<term<P>> const terms_4 = {term_i2x, term_i2v};
 
@@ -347,9 +347,10 @@ private:
       flux_type::central, boundary_condition::periodic,
       boundary_condition::periodic);
 
-  inline static term<P> const term_i3x = term<P>(false,  // time-dependent
-                                                 "I3_x", // name
-                                                 {i3_pterm_x1, i3_pterm_x2});
+  inline static term<P> const term_i3x =
+      term<P>(false,  // time-dependent
+              "I3_x", // name
+              {i3_pterm_x1, i3_pterm_x2}, imex_flag::imp);
 
   static P i3_g3(P const x, P const time = 0)
   {
