+++ src/kronmult.cpp
@@ -276,7 +276,8 @@ fk::vector<P, mem_type::view, resource::device>
 execute(PDE<P> const &pde, elements::table const &elem_table,
         options const &program_opts, element_subgrid const &my_subgrid,
         fk::vector<P, mem_type::const_view, resource::device> const &x,
-        fk::vector<P, mem_type::view, resource::device> &fx)
+        fk::vector<P, mem_type::view, resource::device> &fx,
+        imex_flag const &flag)
 {
   // FIXME code relies on uniform degree across dimensions
   auto const degree     = pde.get_dimensions()[0].get_degree();
@@ -311,14 +312,29 @@ execute(PDE<P> const &pde, elements::table const &elem_table,
   fk::matrix<P, mem_type::const_view> const blah(coeff, 0, real_size - 1, 0,
                                                  real_size - 1);
 
-  fk::vector<P *> const operators = [&pde, lda] {
+  fk::vector<P *> const operators = [&pde, lda, &program_opts, &flag] {
     fk::vector<P *> builder(pde.num_terms * pde.num_dims);
     for (int i = 0; i < pde.num_terms; ++i)
     {
       for (int j = 0; j < pde.num_dims; ++j)
       {
-        builder(i * pde.num_dims + j) = pde.get_coefficients(i, j).data();
-        expect(pde.get_coefficients(i, j).nrows() == lda);
+        // If using IMEX, include terms that only match the imex flag that
+        // matches this stage
+        if (!program_opts.use_imex_stepping ||
+            (program_opts.use_imex_stepping &&
+             pde.get_terms()[i][j].flag == flag))
+        {
+          builder(i * pde.num_dims + j) = pde.get_coefficients(i, j).data();
+          expect(pde.get_coefficients(i, j).nrows() == lda);
+        }
+        else
+        {
+          // fill term portion with 0s for excluded terms
+          builder(i * pde.num_dims + j) =
+              fk::matrix<P>(pde.get_coefficients(i, j).ncols(),
+                            pde.get_coefficients(i, j).nrows())
+                  .data();
+        }
       }
     }
     return builder;
@@ -356,7 +372,8 @@ fk::vector<P, mem_type::owner, resource::host>
 execute(PDE<P> const &pde, elements::table const &elem_table,
         options const &program_opts, element_subgrid const &my_subgrid,
         int const workspace_size_MB,
-        fk::vector<P, mem_type::owner, resource::host> const &x)
+        fk::vector<P, mem_type::owner, resource::host> const &x,
+        imex_flag const &flag)
 {
   auto const grids = decompose(pde, elem_table, my_subgrid, workspace_size_MB);
 
@@ -380,7 +397,7 @@ execute(PDE<P> const &pde, elements::table const &elem_table,
     fk::vector<P, mem_type::view, resource::device> fx_dev_grid(
         fx_dev, row_start * deg_to_dim, (row_end + 1) * deg_to_dim - 1);
     fx_dev_grid = kronmult::execute(pde, elem_table, program_opts, grid,
-                                    x_dev_grid, fx_dev_grid);
+                                    x_dev_grid, fx_dev_grid, flag);
   }
   return fx_dev.clone_onto_host();
 }
@@ -397,12 +414,14 @@ template fk::vector<float, mem_type::owner, resource::host>
 execute(PDE<float> const &pde, elements::table const &elem_table,
         options const &program_options, element_subgrid const &my_subgrid,
         int const workspace_size_MB,
-        fk::vector<float, mem_type::owner, resource::host> const &x);
+        fk::vector<float, mem_type::owner, resource::host> const &x,
+        imex_flag const &flag);
 
 template fk::vector<double, mem_type::owner, resource::host>
[…]
