+++ src/asgard_field.hpp
@@ -19,7 +19,19 @@ namespace asgard
 {
 
 /*!
- * \brief Copies the dimensions and modifies them based on the cli-parameters
+ * \internal
+ * \ingroup AsgardPDESystem
+ * \brief Creates a copy of the dimensions and modifies them based on the cli-parameters
+ *
+ * \tparam precision is either \b float or \b double
+ * \param cli_input is a parser with the current command line arguments
+ * \param dimensions is a user-provided default values for the dimensions of the PDE system
+ *
+ * \returns a copy of the dimensions, but with the command line corrections applied
+ *
+ * \throws runtime_error if sanity check fails on the parser data or the values of the levels and order
+ *
+ * \endinternal
  */
 template<typename precision>
 std::vector<dimension_description<precision>>
@@ -80,7 +92,15 @@ cli_apply_level_degree_correction(parser const &cli_input,
 }
 
 /*!
+ * \internal
+ * \ingroup AsgardPDESystem
  * \brief Throws an exception if there are repeated entries among the names.
+ *
+ * \param names is a list of strings that need a sanity check
+ *
+ * \return \b true if there are no unique entries among the strings, and \b false if repeated entries are found
+ *
+ * \endinternal
  */
 inline bool check_unique_strings(std::vector<std::string> const &names) {
   size_t num_dims = names.size();
@@ -95,8 +115,27 @@ inline bool check_unique_strings(std::vector<std::string> const &names) {
   return true;
 }
 
+/*!
+ * \internal
+ * \ingroup AsgardPDESystem
+ * \brief Wrapper for an arbitrary set of dimension_description objects.
+ *
+ * \tparam precision is either float or double
+ *
+ * Holds a vector of dimension_description objects and provides methods to extract a specific description from the list.
+ *
+ * \endinternal
+ */
 template<typename precision>
 struct dimension_set {
+  /*!
+   * \brief Creates a new set of dimensions from the provided list modified by the command line arguments.
+   *
+   * \param cli_input is a parser of the command line arguments used to modify the default values in dimensions
+   * \param dimensions is a list of dimensions with default data provided by the user
+   *
+   * \throws runtime_error if there are entries with the same name
+   */
   dimension_set(parser const &cli_input, std::vector<dimension_description<precision>> const &dimensions)
     : list(cli_apply_level_degree_correction(cli_input, dimensions))
   {
@@ -108,7 +147,17 @@ struct dimension_set {
       throw std::runtime_error("dimensions should have unique names");
   }
 
-  dimension_description<precision> operator() (std::string const &name) const
+  /*!
+   * \brief Returns the dimension_description for the dimension with the given name.
+   *
+   * \param name is the name to search among the dimensions in the set
+   *
+   * \returns const-reference to the dimension_description with the same name,
+   *          the descriptions have already been updated with the command line arguments.
+   *
+   * \throws runtime_error if the name is not in the list of dimensions
+   */
+  dimension_description<precision> const& operator() (std::string const &name) const
   {
     for(size_t i=0; i<list.size(); i++)
     {
@@ -118,17 +167,44 @@ struct dimension_set {
     throw std::runtime_error(std::string("invalid dimension name: '") + name + "', has not been defined.");
   }
 
+  //! \brief Contains the vector of dimension_description that has been updated by the command line arguments.
   std::vector<dimension_description<precision>> const list;
 };
 
-
+/*!
+ * \ingroup AsgardPDESystem
+ * \brief Defines the type of each field, specifically whether it has a time-differential operator or not.
+ *
+ * TODO: See ... Will add a link to the math document here ...
+ */
 enum class field_mode {
-  evolution, closure
+  //! \brief Each evolution equation will have an implicit time derivative operator added its equation.
+  evolution,
[…]

[…]
