+++ src/asgard_dimension.hpp
@@ -38,6 +38,30 @@ using g_func_type = std::function<P(P const, P const)>;
 
 
 /*!
+ * \internal
+ * \ingroup AsgardPDESystem
+ * \brief Throws an exception if there are repeated entries among the names.
+ *
+ * \param names is a list of strings that need a sanity check
+ *
+ * \return \b true if there are no unique entries among the strings, and \b false if repeated entries are found
+ *
+ * \endinternal
+ */
+inline bool check_unique_strings(std::vector<std::string> const &names) {
+  size_t num_dims = names.size();
+  for(size_t i=0; i<num_dims; i++)
+  {
+    for(size_t j=i+1; j<num_dims; j++)
+    {
+      if (names[i] == names[j])
+        return false;
+    }
+  }
+  return true;
+}
+
+/*!
  * \ingroup AsgardPDESystem
  * \brief Contains the user provided description of a dimension.
  *
@@ -96,6 +120,135 @@ struct dimension_description {
 /*!
  * \internal
  * \ingroup AsgardPDESystem
+ * \brief Creates a copy of the dimensions and modifies them based on the cli-parameters
+ *
+ * \tparam precision is either \b float or \b double
+ * \param cli_input is a parser with the current command line arguments
+ * \param dimensions is a user-provided default values for the dimensions of the PDE system
+ *
+ * \returns a copy of the dimensions, but with the command line corrections applied
+ *
+ * \throws runtime_error if sanity check fails on the parser data or the values of the levels and order
+ *
+ * \endinternal
+ */
+template<typename precision>
+std::vector<dimension_description<precision>>
+cli_apply_level_degree_correction(parser const &cli_input,
+                                  std::vector<dimension_description<precision>> const &dimensions)
+{
+  size_t num_dims = dimensions.size();
+  std::vector<int> levels(dimensions.size()), degrees(dimensions.size());
+  for(size_t i=0; i<num_dims; i++)
+  {
+    levels[i] = dimensions[i].level;
+    degrees[i] = dimensions[i].degree;
+  }
+
+  // modify for appropriate level/degree
+  // if default lev/degree not used
+  auto const user_levels = cli_input.get_starting_levels().size();
+  if (user_levels != 0 && user_levels != static_cast<int>(num_dims))
+  {
+    throw std::runtime_error(
+        std::string("failed to parse dimension-many starting levels - parsed ")
+        + std::to_string(user_levels) + " levels");
+  }
+  if (user_levels == static_cast<int>(num_dims))
+  {
+    auto counter = 0;
+    for (int &l : levels)
+    {
+      l = cli_input.get_starting_levels()(counter++);
+      expect(l > 1);
+    }
+  }
+  auto const cli_degree = cli_input.get_degree();
+  if (cli_degree != parser::NO_USER_VALUE)
+  {
+    expect(cli_degree > 0);
+    for (int &d : degrees) d = cli_degree;
+  }
+
+  // check all dimensions
+  for(size_t i=0; i<dimensions.size(); i++)
+  {
+    expect(degrees[i] > 0);
+    expect(levels[i] > 1);
+  }
+
+  std::vector<dimension_description<precision>> result;
+  result.reserve(num_dims);
+  for(size_t i=0; i<num_dims; i++)
+  {
+    result.push_back(
+      dimension_description<precision>(dimensions[i].d_min, dimensions[i].d_max,
+                                       levels[i], degrees[i],
+                                       dimensions[i].name)
+                     );
+  }
+  return result;
+}
+
+/*!
+ * \internal
+ * \ingroup AsgardPDESystem
+ * \brief Wrapper for an arbitrary set of dimension_description objects.
+ *
+ * \tparam precision is either float or double
+ *
[…]
