+++ src/asgard_field_tests.cpp
@@ -1,59 +1,63 @@
 #include "asgard_field.hpp"
 
+#include "tests_general.hpp"
+
 using namespace asgard;
 
-static fk::vector<float> ic_x(fk::vector<float> const &x,
-                              float const = 0)
+template<typename precision>
+struct volume {
+  static precision jacobi(precision const, precision const) {
+    return 1.0;
+  }
+};
+
+template<typename precision>
+struct initial {
+  static fk::vector<precision> x(fk::vector<precision> const &x,
+                                 precision const = 0)
+  {
+    fk::vector<precision> fx(x.size());
+    for(int i=0; i<fx.size(); i++) fx[i] = 1.0;
+    return fx;
+  }
+  static fk::vector<precision> y(fk::vector<precision> const &x,
+                                 precision const = 0)
+  {
+    fk::vector<precision> fx(x.size());
+    for(int i=0; i<fx.size(); i++) fx[i] = x[i];
+    return fx;
+  }
+  static fk::vector<precision> v(fk::vector<precision> const &x,
+                                 precision const = 0)
+  {
+    fk::vector<precision> fx(x.size());
+    for(int i=0; i<fx.size(); i++) fx[i] = x[i] * x[i];
+    return fx;
+  }
+};
+
+TEMPLATE_TEST_CASE("testing construction of a basic field_description", "[field]", float, double)
 {
-  fk::vector<float> fx(x.size());
-  for(int i=0; i<fx.size(); i++) fx[i] = 1.0;
-  return fx;
-}
-static fk::vector<float> ic_y(fk::vector<float> const &x,
-                             float const = 0)
-{
-  fk::vector<float> fx(x.size());
-  for(int i=0; i<fx.size(); i++) fx[i] = x[i];
-  return fx;
-}
-static fk::vector<float> ic_vel(fk::vector<float> const &x,
-                             float const = 0)
-{
-  fk::vector<float> fx(x.size());
-  for(int i=0; i<fx.size(); i++) fx[i] = x[i] * x[i];
-  return fx;
-}
+  parser const cli_input = make_empty_parser();
 
-static float vol_jac_dV(float const, float const)
-{
-  return 1.0;
-}
-
-int main(int argc, char *argv[])
-{
-
-  parser const cli_input(argc, argv);
-
-  float min0 = 0.0, min1 = 1.0;
-  int level = 2, degree = 2;
-
-  dimension_description<float> dim_0 =
-      dimension_description<float>(min0, min1, level, degree, "x");
-  dimension_description<float> dim_1 =
-      dimension_description<float>(min0, min1, level, degree, "y");
-  dimension_description<float> dim_2 =
-      dimension_description<float>(min0, min1, level, degree, "vx", vol_jac_dV);
-  dimension_description<float> dim_3 =
-      dimension_description<float>(min0, min1, level, degree, "vy");
-
-  field_description<float> field_1d(field_mode::evolution, "x", ic_x, ic_x, "projected to 1d");
-  field_description<float> pos_field(field_mode::closure, {"x", "y"}, {ic_x, ic_y}, {ic_x, ic_y}, "position");
-  field_description<float> vel_only_field(field_mode::evolution, {"vx", "vy"}, {ic_vel, ic_vel}, {ic_vel, ic_vel}, "vel-only");
-  field_description<float> mixed_field(field_mode::closure,
+  field_description<TestType> field_1d(field_mode::evolution, "x", initial<TestType>::x, initial<TestType>::x, "projected to 1d");
+  field_description<TestType> pos_field(field_mode::evolution, {"x", "y"},
+                                        {initial<TestType>::x, initial<TestType>::y},
+                                        {initial<TestType>::x, initial<TestType>::y},
[…]
