+++ src/asgard_pde_system_tests.cpp
@@ -38,10 +38,7 @@ struct initial {
 
 TEMPLATE_TEST_CASE("testing construction of a basic pde_system", "[pde]", float, double)
 {
-
-  std::vector<char> ename = {'a', 's'};
-  char *ename_data = ename.data();
-  parser const cli_input(1, &ename_data); // dummy parser
+  parser const cli_input = make_empty_parser();
 
   TestType min0 = 0.0, min1 = 1.0;
   int level = 2, degree = 2;
