### 2022-12-21 f23decf
fixed mpi testing
