+++ src/pde/pde_base.hpp
@@ -67,7 +67,6 @@ class PDE;
 template<typename P>
 class moment;
 
-
 enum class coefficient_type
 {
   grad,
@@ -546,13 +545,10 @@ public:
   {
     return dimensions_;
   }
-  std::vector<dimension<P>> &get_dimensions()
-  {
-    return dimensions_;
-  }
+  std::vector<dimension<P>> &get_dimensions() { return dimensions_; }
 
   term_set<P> const &get_terms() const { return terms_; }
-  term_set<P> &get_terms(){ return terms_; }
+  term_set<P> &get_terms() { return terms_; }
 
   fk::matrix<P, mem_type::owner, resource::device> const &
   get_coefficients(int const term, int const dim) const
