+++ src/adapt.hpp
@@ -39,36 +39,36 @@ template<typename P>
 class distributed_grid
 {
 public:
-  distributed_grid(options const &cli_options, std::vector<dimension<P>> const &dims);
+  distributed_grid(options const &cli_options,
+                   std::vector<dimension<P>> const &dims);
 
-  distributed_grid(PDE<P> const &pde, options const &cli_options) : distributed_grid(cli_options, pde.get_dimensions()){}
+  distributed_grid(PDE<P> const &pde, options const &cli_options)
+      : distributed_grid(cli_options, pde.get_dimensions())
+  {}
 
   // driver routines
   fk::vector<P> get_initial_condition(
-    PDE<P> &pde,
-    basis::wavelet_transform<P, resource::host> const &transformer,
-    options const &cli_opts)
+      PDE<P> &pde,
+      basis::wavelet_transform<P, resource::host> const &transformer,
+      options const &cli_opts)
   {
-    return
-    this->get_initial_condition(pde.get_dimensions(),
-                                pde.has_analytic_soln ? pde.exact_time(0.0) : static_cast<P>(1.0),
-                                pde.num_terms, pde.get_terms(), transformer, cli_opts);
+    return this->get_initial_condition(
+        pde.get_dimensions(),
+        pde.has_analytic_soln ? pde.exact_time(0.0) : static_cast<P>(1.0),
+        pde.num_terms, pde.get_terms(), transformer, cli_opts);
   }
 
   fk::vector<P> get_initial_condition(
-    std::vector<dimension<P>> &dims,
-    P const mult,
-    int const num_terms, std::vector<std::vector<term<P>>> &terms,
-    basis::wavelet_transform<P, resource::host> const &transformer,
-    options const &cli_opts);
+      std::vector<dimension<P>> &dims, P const mult, int const num_terms,
+      std::vector<std::vector<term<P>>> &terms,
+      basis::wavelet_transform<P, resource::host> const &transformer,
+      options const &cli_opts);
 
   void get_initial_condition(
-    options const &cli_opts,
-    std::vector<dimension<P>> &dims,
-    std::vector<vector_func<P>> const &v_functions,
-    P const mult,
-    basis::wavelet_transform<P, resource::host> const &transformer,
-    fk::vector<P, mem_type::view> result);
+      options const &cli_opts, std::vector<dimension<P>> &dims,
+      std::vector<vector_func<P>> const &v_functions, P const mult,
+      basis::wavelet_transform<P, resource::host> const &transformer,
+      fk::vector<P, mem_type::view> result);
 
   fk::vector<P> coarsen_solution(PDE<P> &pde, fk::vector<P> const &x,
                                  options const &cli_opts);
