+++ src/asgard_field.hpp
@@ -17,14 +17,14 @@
 
 namespace asgard
 {
-
 /*!
  * \ingroup AsgardPDESystem
  * \brief Defines the type of each field, specifically whether it has a time-differential operator or not.
  *
  * TODO: See ... Will add a link to the math document here ...
  */
-enum class field_mode {
+enum class field_mode
+{
   //! \brief Each evolution equation will have an implicit time derivative operator added its equation.
   evolution,
   //! \brief Fields with no time derivative, i.e., fields that depend explicitly on the field_mode::evolution fields.
@@ -36,7 +36,8 @@ enum class field_mode {
  * \brief Simple container holding the basic properties of a field.
  *
  * The user has to crate a description object for each field of the PDE system.
- * A vector holding all of these objects will be passed into the constructor of asgard::pde_system
+ * A vector holding all of these objects will be passed into the constructor of
+ * asgard::pde_system
  *
  * \tparam precision is either \b float or \b double and indicates whether to use 32-bit or 64-bit floating point precision
  */
@@ -52,25 +53,23 @@ struct field_description
    * \param exact_solution is an optional parameter that defines the exact solution for the field
    * \param field_name a unique identifier for this field
    */
-  field_description(field_mode fmode,
-                    std::string const &dimension_name,
+  field_description(field_mode fmode, std::string const &dimension_name,
                     vector_func<precision> const initial_condition,
                     vector_func<precision> const exact_solution,
-                    std::string const &field_name
-                    )
-    : field_description(fmode, std::vector<std::string>{dimension_name}, {initial_condition}, {exact_solution}, field_name)
-    {}
+                    std::string const &field_name)
+      : field_description(fmode, std::vector<std::string>{dimension_name},
+                          {initial_condition}, {exact_solution}, field_name)
+  {}
 
   /*!
    * \brief Constructs a new 1D field without a known exact solution and the following parameters.
    */
-  field_description(field_mode fmode,
-                    std::string const &dimension_name,
+  field_description(field_mode fmode, std::string const &dimension_name,
                     vector_func<precision> const initial_condition,
-                    std::string const &field_name
-                    )
-    : field_description(fmode, std::vector<std::string>{dimension_name}, {initial_condition}, field_name)
-    {}
+                    std::string const &field_name)
+      : field_description(fmode, std::vector<std::string>{dimension_name},
+                          {initial_condition}, field_name)
+  {}
 
   /*!
    * \brief Constructs a new multidimensional field with known exact solution and the following parameters.
@@ -78,28 +77,28 @@ struct field_description
    * \param fmode identifies the field as either field_mode::evolution or field_mode::closure
    * \param dimension_names is a list of the dimensions associated with the field
    * \param initial_conditions provides initial condition for this field,
-   *        note that the initial condition is a product of the 1D initial conditions
-   *        as per the separability assumption in the code
+   *        note that the initial condition is a product of the 1D initial
+   * conditions as per the separability assumption in the code
    * \param exact_solution defines the exact solution for the field (assumed separable),
    *        if the exact solution is not knows, then this can be an empty vector
    * \param field_name a unique identifier for this field
    *
    * \throws runtime_error if there is a mismatch in the sizes of dimension_names,
-   *         initial_conditions and exact_solution or if dimension_names contains repeated entries.
+   *         initial_conditions and exact_solution or if dimension_names
+   * contains repeated entries.
    */
-  field_description(field_mode fmode,
[…]
