+++ src/kronmult.hpp
@@ -20,6 +20,6 @@ execute(PDE<P> const &pde, elements::table const &elem_table,
         options const &program_options, element_subgrid const &my_subgrid,
         int const workspace_size_MB,
         fk::vector<P, mem_type::owner, resource::host> const &x,
-        imex_flag const &flag = imex_flag::exp);
+        imex_flag const imex = imex_flag::unspecified);
 
 } // namespace asgard::kronmult
