+++ src/pde/pde_vlasov_lb_full_f.hpp
@@ -197,13 +197,15 @@ private:
       flux_type::central, boundary_condition::periodic,
       boundary_condition::periodic);
 
-  inline static term<P> const term_e1x = term<P>(false,  // time-dependent
-                                                 "E1_x", // name
-                                                 {e1_pterm_x});
+  inline static term<P> const term_e1x =
+      term<P>(false,  // time-dependent
+              "E1_x", // name
+              {e1_pterm_x}, imex_flag::imex_explicit);
 
-  inline static term<P> const term_e1v = term<P>(false,  // time-dependent
-                                                 "E1_v", // name
-                                                 {e1_pterm_v});
+  inline static term<P> const term_e1v =
+      term<P>(false,  // time-dependent
+              "E1_v", // name
+              {e1_pterm_v}, imex_flag::imex_explicit);
 
   inline static std::vector<term<P>> const terms_1 = {term_e1x, term_e1v};
 
@@ -233,13 +235,15 @@ private:
       flux_type::central, boundary_condition::periodic,
       boundary_condition::periodic);
 
-  inline static term<P> const term_e2x = term<P>(false,  // time-dependent
-                                                 "E2_x", // name
-                                                 {e2_pterm_x});
+  inline static term<P> const term_e2x =
+      term<P>(false,  // time-dependent
+              "E2_x", // name
+              {e2_pterm_x}, imex_flag::imex_explicit);
 
-  inline static term<P> const term_e2v = term<P>(false,  // time-dependent
-                                                 "E2_v", // name
-                                                 {e2_pterm_v});
+  inline static term<P> const term_e2v =
+      term<P>(false,  // time-dependent
+              "E2_v", // name
+              {e2_pterm_v}, imex_flag::imex_explicit);
 
   inline static std::vector<term<P>> const terms_2 = {term_e2x, term_e2v};
 
@@ -269,13 +273,15 @@ private:
                       flux_type::downwind, boundary_condition::dirichlet,
                       boundary_condition::dirichlet);
 
-  inline static term<P> const term_i1x = term<P>(false,  // time-dependent
-                                                 "I1_x", // name
-                                                 {i1_pterm_x}, imex_flag::imp);
+  inline static term<P> const term_i1x =
+      term<P>(false,  // time-dependent
+              "I1_x", // name
+              {i1_pterm_x}, imex_flag::imex_implicit);
 
-  inline static term<P> const term_i1v = term<P>(false,  // time-dependent
-                                                 "I1_v", // name
-                                                 {i1_pterm_v}, imex_flag::imp);
+  inline static term<P> const term_i1v =
+      term<P>(false,  // time-dependent
+              "I1_v", // name
+              {i1_pterm_v}, imex_flag::imex_implicit);
 
   inline static std::vector<term<P>> const terms_3 = {term_i1x, term_i1v};
 
@@ -306,13 +312,15 @@ private:
                       flux_type::central, boundary_condition::dirichlet,
                       boundary_condition::dirichlet);
 
-  inline static term<P> const term_i2x = term<P>(false,  // time-dependent
-                                                 "I2_x", // name
-                                                 {i2_pterm_x}, imex_flag::imp);
+  inline static term<P> const term_i2x =
+      term<P>(false,  // time-dependent
+              "I2_x", // name
+              {i2_pterm_x}, imex_flag::imex_implicit);
 
-  inline static term<P> const term_i2v = term<P>(false,  // time-dependent
-                                                 "I2_v", // name
-                                                 {i2_pterm_v}, imex_flag::imp);
[…]
