### 2023-01-06 79af2db
Use scal for setting up system matrix
