### 2023-01-06 ecc745c
Update src/asgard_field.hpp
Co-authored-by: Steven Hahn <hahnse@ornl.gov>
