### 2023-01-06 a96abcc
Update src/asgard_field.hpp
Co-authored-by: Steven Hahn <hahnse@ornl.gov>
