+++ src/adapt.cpp
@@ -210,7 +210,7 @@ fk::vector<P> distributed_grid<P>::get_initial_condition(
 
 template<typename P>
 void distributed_grid<P>::get_initial_condition(
-    options const &cli_opts, std::vector<dimension<P>> &dims,
+    options const &cli_opts, std::vector<dimension<P>> const &dims,
     std::vector<vector_func<P>> const &v_functions, P const mult,
     basis::wavelet_transform<P, resource::host> const &transformer,
     fk::vector<P, mem_type::view> result)
