+++ src/asgard_discretization_tests.cpp
@@ -25,24 +25,9 @@ struct initial
   }
 };
 
-struct distribution_test_init
-{
-  void set_my_rank(const int rank) { my_rank = rank; }
-  void set_num_ranks(const int size) { num_ranks = size; }
-  int get_my_rank() const { return my_rank; }
-  int get_num_ranks() const { return num_ranks; }
-
-private:
-  int my_rank;
-  int num_ranks;
-};
-static distribution_test_init distrib_test_info;
-
 int main(int argc, char *argv[])
 {
-  auto const [rank, total_ranks] = initialize_distribution();
-  distrib_test_info.set_my_rank(rank);
-  distrib_test_info.set_num_ranks(total_ranks);
+  initialize_distribution();
 
   int result = Catch::Session().run(argc, argv);
 
