### 2023-01-08 4802b5a
Use runtime_error and add param check in main
