### 2023-01-09 d928e93
Use default constructor for parameter
