### 2023-01-09 7c18999
Field class (#474)
* field class initial implementaiotn
