+++ src/asgard_field.cpp
@@ -1,0 +1,1 @@
+#include "asgard_field.hpp"
