### 2023-01-10 3ae8d39
Remove dimension limit added to wavelet transform
