### 2023-01-10 e417e50
Rename num_terms_in and add test
