+++ src/pde/pde_vlasov_lb_full_f.hpp
@@ -18,7 +18,11 @@ public:
       : PDE<P>(cli_input, num_dims_, num_sources_, num_terms_, dimensions_,
                terms_, sources_, exact_vector_funcs_, exact_scalar_func_,
                get_dt_, do_poisson_solve_, has_analytic_soln_, moments_)
-  {}
+  {
+    param_manager.add_parameter(parameter<P>{"n", n});
+    param_manager.add_parameter(parameter<P>{"u", u});
+    param_manager.add_parameter(parameter<P>{"theta", theta});
+  }
 
 private:
   static int constexpr num_dims_           = 2;
@@ -193,13 +197,15 @@ private:
       flux_type::central, boundary_condition::periodic,
       boundary_condition::periodic);
 
-  inline static term<P> const term_e1x = term<P>(false,  // time-dependent
-                                                 "E1_x", // name
-                                                 {e1_pterm_x});
+  inline static term<P> const term_e1x =
+      term<P>(false,  // time-dependent
+              "E1_x", // name
+              {e1_pterm_x}, imex_flag::imex_explicit);
 
-  inline static term<P> const term_e1v = term<P>(false,  // time-dependent
-                                                 "E1_v", // name
-                                                 {e1_pterm_v});
+  inline static term<P> const term_e1v =
+      term<P>(false,  // time-dependent
+              "E1_v", // name
+              {e1_pterm_v}, imex_flag::imex_explicit);
 
   inline static std::vector<term<P>> const terms_1 = {term_e1x, term_e1v};
 
@@ -229,13 +235,15 @@ private:
       flux_type::central, boundary_condition::periodic,
       boundary_condition::periodic);
 
-  inline static term<P> const term_e2x = term<P>(false,  // time-dependent
-                                                 "E2_x", // name
-                                                 {e2_pterm_x});
+  inline static term<P> const term_e2x =
+      term<P>(false,  // time-dependent
+              "E2_x", // name
+              {e2_pterm_x}, imex_flag::imex_explicit);
 
-  inline static term<P> const term_e2v = term<P>(false,  // time-dependent
-                                                 "E2_v", // name
-                                                 {e2_pterm_v});
+  inline static term<P> const term_e2v =
+      term<P>(false,  // time-dependent
+              "E2_v", // name
+              {e2_pterm_v}, imex_flag::imex_explicit);
 
   inline static std::vector<term<P>> const terms_2 = {term_e2x, term_e2v};
 
@@ -265,13 +273,15 @@ private:
                       flux_type::downwind, boundary_condition::dirichlet,
                       boundary_condition::dirichlet);
 
-  inline static term<P> const term_i1x = term<P>(false,  // time-dependent
-                                                 "I1_x", // name
-                                                 {i1_pterm_x});
+  inline static term<P> const term_i1x =
+      term<P>(false,  // time-dependent
+              "I1_x", // name
+              {i1_pterm_x}, imex_flag::imex_implicit);
 
-  inline static term<P> const term_i1v = term<P>(false,  // time-dependent
-                                                 "I1_v", // name
-                                                 {i1_pterm_v});
+  inline static term<P> const term_i1v =
+      term<P>(false,  // time-dependent
+              "I1_v", // name
+              {i1_pterm_v}, imex_flag::imex_implicit);
 
   inline static std::vector<term<P>> const terms_3 = {term_i1x, term_i1v};
 
@@ -280,9 +290,9 @@ private:
   //
   static P i2_g1(P const x, P const time = 0)
   {
-    ignore(x);
-    ignore(time);
-    return -u(x);
+    auto param = param_manager.get_parameter("u");
+    expect(param != nullptr);
[…]
