### 2023-01-10 fbed0c7
trowing a fit about the use of exit and clarifying a method
