### 2023-01-10 0240e1c
Fix term selection option (#476)
Fix term selection option
