+++ src/pde/pde_base.hpp
@@ -451,7 +451,7 @@ class PDE
 {
 public:
   PDE(parser const &cli_input, int const num_dims_in, int const num_sources_in,
-      int const num_terms_in, std::vector<dimension<P>> const dimensions,
+      int const max_num_terms, std::vector<dimension<P>> const dimensions,
       term_set<P> const terms, std::vector<source<P>> const sources_in,
       std::vector<vector_func<P>> const exact_vector_funcs_in,
       scalar_func<P> const exact_time_in, dt_func<P> const get_dt,
@@ -459,7 +459,7 @@ public:
       bool const has_analytic_soln_in         = false,
       std::vector<moment<P>> const moments_in = {})
       : num_dims(num_dims_in), num_sources(num_sources_in),
-        num_terms(get_num_terms(cli_input, num_terms_in)),
+        num_terms(get_num_terms(cli_input, max_num_terms)),
         max_level(get_max_level(cli_input, dimensions)), sources(sources_in),
         exact_vector_funcs(exact_vector_funcs_in), moments(moments_in),
         exact_time(exact_time_in), do_poisson_solve(do_poisson_solve_in),
@@ -471,7 +471,7 @@ public:
     expect(num_terms > 0);
 
     expect(dimensions.size() == static_cast<unsigned>(num_dims));
-    expect(terms.size() == static_cast<unsigned>(num_terms));
+    expect(terms.size() == static_cast<unsigned>(max_num_terms));
     expect(sources.size() == static_cast<unsigned>(num_sources));
 
     // ensure analytic solution functions were provided if this flag is set
@@ -504,7 +504,7 @@ public:
     if (num_active_terms != 0)
     {
       auto const active_terms = cli_input.get_active_terms();
-      for (auto i = num_terms - 1; i >= 0; --i)
+      for (auto i = max_num_terms - 1; i >= 0; --i)
       {
         if (active_terms(i) == 0)
         {
@@ -738,7 +738,7 @@ public:
   }
 
 private:
-  int get_num_terms(parser const &cli_input, int const num_terms_in) const
+  int get_num_terms(parser const &cli_input, int const max_num_terms) const
   {
     // returns either the number of terms set in the PDE specification, or the
     // number of terms toggled on by the user
@@ -746,25 +746,25 @@ private:
 
     // verify that the CLI input matches the spec before altering the num_terms
     // we have
-    if (num_active_terms != 0 && num_active_terms != num_terms_in)
+    if (num_active_terms != 0 && num_active_terms != max_num_terms)
     {
-      std::cerr << "failed to parse dimension-many active terms - parsed "
-                << num_active_terms << " terms, expected " << num_terms << "\n";
-      exit(1);
+      throw std::runtime_error(
+          std::string("failed to parse dimension-many active terms - parsed ") +
+          std::to_string(num_active_terms) + " terms, expected " +
+          std::to_string(max_num_terms));
     }
-    if (num_active_terms == num_terms_in)
+    // if nothing specified in the cli, use the default max_num_terms
+    if (num_active_terms == 0)
+      return max_num_terms;
+    // terms specified in the cli, parse the new number of terms
+    auto const active_terms = cli_input.get_active_terms();
+    int new_num_terms =
+        std::accumulate(active_terms.begin(), active_terms.end(), 0);
+    if (new_num_terms == 0)
     {
-      auto const active_terms = cli_input.get_active_terms();
-      int new_num_terms =
-          std::accumulate(active_terms.begin(), active_terms.end(), 0);
-      if (new_num_terms == 0)
-      {
-        std::cerr << "must have at least one term enabled\n";
-        exit(1);
-      }
-      return new_num_terms;
[…]
