### 2023-01-11 c50229e
Address comments from code review
Signed-off-by: Steven Hahn <hahnse@ornl.gov>
