### 2023-01-11 1e8988f
Fix clang warnings
Signed-off-by: Steven Hahn <hahnse@ornl.gov>
