### 2023-01-11 88b6b62
update variable names
Signed-off-by: Steven Hahn <hahnse@ornl.gov>
