+++ src/asgard_discretization_tests.cpp
@@ -1,7 +1,7 @@
 #include "asgard_discretization.hpp"
 #include "build_info.hpp"
-#include "transformations.hpp"
 #include "tests_general.hpp"
+#include "transformations.hpp"
 
 using namespace asgard;
 
