### 2023-01-11 02f0fd5
rename real_solution_size
Signed-off-by: Steven Hahn <hahnse@ornl.gov>
