### 2023-01-12 e1d95f7
Convert from wavelet to real space for plotting (#479)
* Convert from wavelet to realspace for plotting

* Address comments from code review
* Fix clang warnings
* fix failing test
* update variable names
* rename real_solution_size

Signed-off-by: Steven Hahn <hahnse@ornl.gov>
