### 2023-01-12 2dbcb40
Fix compiler errors building with -DASGARD_USE_MATLAB
Signed-off-by: Steven Hahn <hahnse@ornl.gov>
