### 2023-01-12 99fe77a
Fix compiler errors building with -DASGARD_USE_MATLAB (#480)
* Fix compiler errors building with -DASGARD_USE_MATLAB

* fix warnings

Signed-off-by: Steven Hahn <hahnse@ornl.gov>
