+++ src/main.cpp
@@ -189,7 +189,7 @@ int main(int argc, char **argv)
   ml_plot.connect(cli_input.get_ml_session_string());
   asgard::node_out() << "  connected to MATLAB" << '\n';
 
-  asgard::fk::vector<prec> analytic_solution_realspace(real_space_size);
+  asgard::fk::vector<prec> analytic_solution_realspace(dense_size);
   if (pde->has_analytic_soln)
   {
     // generate the analytic solution at t=0
@@ -301,8 +301,7 @@ int main(int argc, char **argv)
 #ifdef ASGARD_USE_MATLAB
       if (opts.should_plot(i))
       {
-        auto const dense_size = asgard::dense_space_size(*pde);
-        auto transform_wksp   = asgard::update_transform_workspace<prec>(
+        auto transform_wksp = asgard::update_transform_workspace<prec>(
             dense_size, workspace, tmp_workspace);
         if (dense_size > analytic_solution_realspace.size())
         {
@@ -324,8 +323,7 @@ int main(int argc, char **argv)
     if (opts.should_output_realspace(i) || opts.should_plot(i))
     {
       // resize transform workspaces if grid size changed due to adaptivity
-      auto const dense_size = dense_space_size(*pde);
-      auto transform_wksp   = asgard::update_transform_workspace<prec>(
+      auto transform_wksp = asgard::update_transform_workspace<prec>(
           dense_size, workspace, tmp_workspace);
       real_space.resize(dense_size);
 
