### 2023-01-13 e83d98e
Draft 1D Poisson solver.
