+++ testing/sandbox.cpp
@@ -27,10 +27,181 @@ using prec = double;
 using prec = float;
 #endif
 
+#include <Accelerate/Accelerate.h>
+
+using namespace asgard;
+
+void poisson_solver( fk::vector<prec> const &source, fk::vector<prec> &phi,       fk::vector<prec> &E, double h, int degree, int N_elements, double x_min, double x_max, double phi_min, double phi_max )
+{
+    
+    // Solving: - phi_xx = source Using Linear Finite Elements
+    // Boundary Cconditions: phi(x_min)=phi_min and phi(x_max)=phi_max
+    // Returns phi and E = - Phi_x in Gauss-Legendre Nodes
+    
+    int INFO;
+    int NRHS = 1;
+    int N_nodes = N_elements + 1;
+    fk :: vector<prec> b   (N_nodes); // Source Vector //
+    fk :: vector<prec> A_D (N_nodes);
+    fk :: vector<prec> A_DL(N_nodes-1);
+    fk :: vector<prec> A_DU(N_nodes-1);
+    
+    auto const lgwt = legendre_weights( degree+1, -1.0, 1.0, true );
+    
+    // Set the source vector //
+    
+    for ( int i = 1; i < N_nodes-1; i++ )
+    {
+        b[i] = 0.0;
+        for ( int q = 0; q < degree+1; q++ )
+        {
+            b[i] += 0.25 * h * lgwt[1][q]
+                    * (   source[(i-1)*(degree+1)+q] * ( 1.0 + lgwt[0][q] )
+                        + source[(i  )*(degree+1)+q] * ( 1.0 - lgwt[0][q] ) );
+        }
+    }
+    
+    b[0]         = 0.0;
+    b[N_nodes-1] = 0.0;
+    
+    // Set the Matrix //
+    
+    for ( int i = 1; i < N_nodes-1; i++ )
+    {
+        A_D[i] = 2.0 / h;
+    }
+    
+    A_D[0]         = 1.0;
+    A_D[N_nodes-1] = 1.0;
+    
+    for ( int i = 0; i < N_nodes-1; i++ )
+    {
+        A_DL[i] = - 1.0 / h;
+        A_DU[i] = - 1.0 / h;
+    }
+    
+    A_DL[N_nodes-2] = 0.0;
+    A_DU[0]         = 0.0;
+    
+    // Linear Solve //
+    
+    dgtsv_( &N_nodes, &NRHS, A_DL.data(), A_D.data(), A_DU.data(), b.data(), &N_nodes, &INFO );
+    
+    // Set Potential and Electric Field in DG Nodes //
+    
+    for ( int i = 0; i < N_elements; i++ )
+    {
+        for ( int q = 0; q < degree+1; q++ )
+        {
+            
+            int k      = i * (degree+1) + q;
+            double x_k = ( x_min + i * h ) + 0.5 * h * ( 1.0 + lgwt[0][q] );
+            double dg  = ( phi_max - phi_min ) / ( x_max - x_min );
+            double g_k = phi_min + dg * ( x_k - x_min );
+            
+            phi[k] = 0.5 * (   b[i]   * ( 1.0 - lgwt[0][q] )
+                             + b[i+1] * ( 1.0 + lgwt[0][q] ) ) + g_k;
+            
+            E[k] = - ( b[i+1] - b[i] ) / h - dg;
+            
+        }
+    }
+    
+}
+
 int main(int, char**)
 {
   // keep this file clean for each PR
   // allows someone to easily come here, dump code and start playing
   // this is good for prototyping and quick-testing features/behavior
+    
+    int    const N_elements = 16;
+    int    const N_nodes    = N_elements + 1;
+    int    const degree     = 2;
+    double const x_min = 0.0;
+    double const x_max = 1.0;
+    double const phi_min = 0.0;
[…]
