+++ src/batch.hpp
@@ -149,5 +149,7 @@ private:
 // doesn't use batches, but does use many of the same helpers/structure
 template<typename P>
 void build_system_matrix(PDE<P> const &pde, elements::table const &elem_table,
-                         fk::matrix<P> &A, element_subgrid const &grid);
+                         fk::matrix<P> &A, element_subgrid const &grid,
+                         bool const use_imex  = false,
+                         imex_flag const imex = imex_flag::unspecified);
 } // namespace asgard
