### 2023-01-17 9f1911e
Add moment matrix functions for 1x2v and 1x3v
