### 2023-01-17 c6b4338
Edits due to comments from Miro.
