+++ testing/sandbox.cpp
@@ -31,24 +31,24 @@ using prec = float;
 
 using namespace asgard;
 
-void poisson_solver( fk::vector<prec> const &source, fk::vector<prec> &phi,       fk::vector<prec> &E, double h, int degree, int N_elements, double x_min, double x_max, double phi_min, double phi_max )
+void poisson_solver( fk::vector<prec> const &source, fk::vector<prec> &phi, fk::vector<prec> &E, int degree, int N_elements, double x_min, double x_max, double phi_min, double phi_max )
 {
     
     // Solving: - phi_xx = source Using Linear Finite Elements
     // Boundary Cconditions: phi(x_min)=phi_min and phi(x_max)=phi_max
     // Returns phi and E = - Phi_x in Gauss-Legendre Nodes
     
-    int INFO;
-    int NRHS = 1;
-    int N_nodes = N_elements + 1;
-    fk :: vector<prec> b   (N_nodes); // Source Vector //
-    fk :: vector<prec> A_D (N_nodes);
-    fk :: vector<prec> A_DL(N_nodes-1);
-    fk :: vector<prec> A_DU(N_nodes-1);
+    double h = ( x_max - x_min ) / (double) N_elements;
     
     auto const lgwt = legendre_weights( degree+1, -1.0, 1.0, true );
     
+    int N_nodes = N_elements + 1;
+    
     // Set the source vector //
+    
+    fk :: vector<prec> b(N_nodes);
+    
+    b[0] = 0.0;
     
     for ( int i = 1; i < N_nodes-1; i++ )
     {
@@ -60,19 +60,28 @@ void poisson_solver( fk::vector<prec> const &source, fk::vector<prec> &phi,     
                         + source[(i  )*(degree+1)+q] * ( 1.0 - lgwt[0][q] ) );
         }
     }
-    
-    b[0]         = 0.0;
+        
     b[N_nodes-1] = 0.0;
     
     // Set the Matrix //
+    
+    // Diagonal Entries //
+    
+    fk :: vector<prec> A_D(N_nodes);
+    
+    A_D[0] = 1.0;
     
     for ( int i = 1; i < N_nodes-1; i++ )
     {
         A_D[i] = 2.0 / h;
     }
     
-    A_D[0]         = 1.0;
     A_D[N_nodes-1] = 1.0;
+    
+    // Off-Diagonal Entries //
+    
+    fk :: vector<prec> A_DL(N_nodes-1);
+    fk :: vector<prec> A_DU(N_nodes-1);
     
     for ( int i = 0; i < N_nodes-1; i++ )
     {
@@ -80,14 +89,19 @@ void poisson_solver( fk::vector<prec> const &source, fk::vector<prec> &phi,     
         A_DU[i] = - 1.0 / h;
     }
     
-    A_DL[N_nodes-2] = 0.0;
     A_DU[0]         = 0.0;
+    A_DL[N_nodes-2] = 0.0;
     
     // Linear Solve //
+    
+    int INFO = 0;
+    int NRHS = 1;
     
     dgtsv_( &N_nodes, &NRHS, A_DL.data(), A_D.data(), A_DU.data(), b.data(), &N_nodes, &INFO );
     
     // Set Potential and Electric Field in DG Nodes //
+    
+    double dg = ( phi_max - phi_min ) / ( x_max - x_min );
     
     for ( int i = 0; i < N_elements; i++ )
     {
@@ -96,7 +110,6 @@ void poisson_solver( fk::vector<prec> const &source, fk::vector<prec> &phi,     
             
             int k      = i * (degree+1) + q;
             double x_k = ( x_min + i * h ) + 0.5 * h * ( 1.0 + lgwt[0][q] );
-            double dg  = ( phi_max - phi_min ) / ( x_max - x_min );
[…]
