### 2023-01-18 4347181
Switched to symmetric tri-diagonal solver and matrix factorization.
