+++ src/time_advance.cpp
@@ -60,8 +60,9 @@ adaptive_advance(method const step_method, PDE<P> &pde,
       return explicit_advance(pde, adaptive_grid, transformer, program_opts,
                               unscaled_parts, x_orig, workspace_size_MB, time);
     case (method::imp):
-      return implicit_advance(pde, adaptive_grid, transformer, unscaled_parts,
-                              x_orig, time, program_opts.solver, update_system);
+      return implicit_advance(pde, adaptive_grid, transformer, program_opts,
+                              unscaled_parts, x_orig, workspace_size_MB, time,
+                              update_system);
     case (method::imex):
       return imex_advance(pde, adaptive_grid, transformer, program_opts,
                           unscaled_parts, x_orig, workspace_size_MB, time,
@@ -93,8 +94,9 @@ adaptive_advance(method const step_method, PDE<P> &pde,
         return explicit_advance(pde, adaptive_grid, transformer, program_opts,
                                 unscaled_parts, y, workspace_size_MB, time);
       case (method::imp):
-        return implicit_advance(pde, adaptive_grid, transformer, unscaled_parts,
-                                y, time, program_opts.solver, refining);
+        return implicit_advance(pde, adaptive_grid, transformer, program_opts,
+                                unscaled_parts, y, workspace_size_MB, time,
+                                refining);
       case (method::imex):
         return imex_advance(pde, adaptive_grid, transformer, program_opts,
                             unscaled_parts, y, workspace_size_MB, time,
@@ -269,15 +271,17 @@ fk::vector<P>
 implicit_advance(PDE<P> const &pde,
                  adapt::distributed_grid<P> const &adaptive_grid,
                  basis::wavelet_transform<P, resource::host> const &transformer,
+                 options const &program_opts,
                  std::array<boundary_conditions::unscaled_bc_parts<P>, 2> const
                      &unscaled_parts,
-                 fk::vector<P> const &x_orig, P const time,
-                 solve_opts const solver, bool const update_system)
+                 fk::vector<P> const &x_orig, int const workspace_size_MB,
+                 P const time, bool const update_system)
 {
   expect(time >= 0);
 #ifdef ASGARD_USE_SCALAPACK
   auto sgrid = get_grid();
 #endif
+  solve_opts const solver = program_opts.solver;
   static fk::matrix<P, mem_type::owner, resource::host> A;
   static std::vector<int> ipiv;
   static bool first_time = true;
@@ -393,8 +397,10 @@ implicit_advance(PDE<P> const &pde,
     int const restart  = A.ncols();
     int const max_iter = A.ncols();
     fk::vector<P> fx(x.size());
-    solver::simple_gmres(A, fx, x, fk::matrix<P>(), restart, max_iter,
-                         tolerance);
+    //
+
+    solver::simple_gmres(pde, table, program_opts, grid, workspace_size_MB, fx,
+                         x, fk::matrix<P>(), restart, max_iter, tolerance);
     return fx;
   }
   return x;
@@ -419,7 +425,7 @@ imex_advance(PDE<P> &pde, adapt::distributed_grid<P> const &adaptive_grid,
   expect(time >= 0);
   expect(pde.moments.size() > 0);
 
-  static fk::matrix<P, mem_type::owner, resource::host> A;
+  // static fk::matrix<P, mem_type::owner, resource::host> A;
   static std::vector<int> ipiv;
   static bool first_time = true;
 
[…]
