+++ src/time_advance.hpp
@@ -44,10 +44,11 @@ fk::vector<P>
 implicit_advance(PDE<P> const &pde,
                  adapt::distributed_grid<P> const &adaptive_grid,
                  basis::wavelet_transform<P, resource::host> const &transformer,
+                 options const &program_opts,
                  std::array<boundary_conditions::unscaled_bc_parts<P>, 2> const
                      &unscaled_parts,
-                 fk::vector<P> const &x, P const time, solve_opts const solver,
-                 bool const update_system = true);
+                 fk::vector<P> const &x, int const workspace_size_MB,
+                 P const time, bool const update_system = true);
 
 template<typename P>
 fk::vector<P>
