### 2023-01-18 a40a617
Add pttrs and pttrf wrappers to lib_dispatch
