### 2023-01-18 b502bbe
abstract matrix vector product
Signed-off-by: Steven Hahn <hahnse@ornl.gov>
