+++ src/solver.hpp
@@ -7,13 +7,55 @@
 
 namespace asgard::solver
 {
+template<typename precision>
+class MatrixFree
+{
+public:
+  MatrixFree(PDE<precision> const &pde, elements::table const &elem_table,
+             options const &program_options, element_subgrid const &my_subgrid,
+             int const workspace_size_MB)
+      : pde_{pde}, elem_table_{elem_table}, program_options_{program_options},
+        my_subgrid_{my_subgrid}, workspace_size_MB_{workspace_size_MB}
+  {}
+  // template<typename xmem, typename ymem, typename resrc>
+  void operator()(fk::vector<precision> const &x, fk::vector<precision> &y,
+                  precision const alpha = 1.0, precision const beta = 0.0)
+  {
+    auto tmp = kronmult::execute(pde_, elem_table_, program_options_,
+                                 my_subgrid_, workspace_size_MB_, x);
+    y        = tmp * alpha + y * beta;
+  }
+
+private:
+  PDE<precision> const &pde_;
+  elements::table const &elem_table_;
+  options const &program_options_;
+  element_subgrid const &my_subgrid_;
+  int const workspace_size_MB_;
+};
+
+template<typename precision>
+class DenseMatrix
+{
+public:
+  DenseMatrix(fk::matrix<precision> const &A) : A_{A} {}
+  // template<typename xmem, typename ymem, typename resrc>
+  void operator()(fk::vector<precision> const &x, fk::vector<precision> &y,
+                  precision const alpha = 1.0, precision const beta = 0.0)
+  {
+    bool const trans_A = false;
+    fm::gemv(A_, x, y, trans_A, alpha, beta);
+  }
+
+private:
+  fk::matrix<precision> &A_;
+};
+
 // simple, node-local test version
-template<typename P>
-P simple_gmres(PDE<P> const &pde, elements::table const &elem_table,
-               options const &program_options,
-               element_subgrid const &my_subgrid, int const workspace_size_MB,
-               fk::vector<P> &x, fk::vector<P> const &b, fk::matrix<P> const &M,
-               int const restart, int const max_iter, P const tolerance)
+template<typename P, typename MatrixReplacement>
+P simple_gmres(MatrixReplacement asdf, fk::vector<P> &x, fk::vector<P> const &b,
+               fk::matrix<P> const &M, int const restart, int const max_iter,
+               P const tolerance)
 {
   expect(tolerance >= std::numeric_limits<P>::epsilon());
   int const n = b.size();
@@ -45,9 +87,7 @@ P simple_gmres(PDE<P> const &pde, elements::table const &elem_table,
     P const alpha      = -1.0;
     P const beta       = 1.0;
     residual           = b;
-    auto tmp = kronmult::execute(pde, elem_table, program_options, my_subgrid,
-                                 workspace_size_MB, x);
-    x        = tmp * -1.0 + x;
+    asdf(x, residual, alpha, beta);
     if (do_precond)
     {
       precond_factored ? fm::getrs(precond, residual, precond_pivots)
@@ -90,8 +130,8 @@ P simple_gmres(PDE<P> const &pde, elements::table const &elem_table,
     {
       auto tmp = fk::vector<P>(
           fk::vector<P, mem_type::view>(basis, i, 0, basis.nrows() - 1));
-      fk::vector<P> new_basis = kronmult::execute(
-          pde, elem_table, program_options, my_subgrid, workspace_size_MB, tmp);
+      fk::vector<P> new_basis(tmp.size());
+      asdf(tmp, new_basis, P{1.0}, P{0.0});
 
       if (do_precond)
       {
