+++ src/solver.hpp
@@ -1,217 +1,22 @@
 #pragma once
-#include "fast_math.hpp"
-#include "kronmult.hpp"
-#include "solver.hpp"
+
+#include "batch.hpp"
+#include "pde.hpp"
 #include "tensors.hpp"
-#include "tools.hpp"
 
 namespace asgard::solver
 {
-template<typename precision>
-class MatrixFree
-{
-public:
-  MatrixFree(PDE<precision> const &pde, elements::table const &elem_table,
-             options const &program_options, element_subgrid const &my_subgrid,
-             int const workspace_size_MB)
-      : pde_{pde}, elem_table_{elem_table}, program_options_{program_options},
-        my_subgrid_{my_subgrid}, workspace_size_MB_{workspace_size_MB}
-  {}
-  // template<typename xmem, typename ymem, typename resrc>
-  void operator()(fk::vector<precision> const &x, fk::vector<precision> &y,
-                  precision const alpha = 1.0, precision const beta = 0.0)
-  {
-    auto tmp = kronmult::execute(pde_, elem_table_, program_options_,
-                                 my_subgrid_, workspace_size_MB_, x);
-    y        = tmp * alpha + y * beta;
-  }
-
-private:
-  PDE<precision> const &pde_;
-  elements::table const &elem_table_;
-  options const &program_options_;
-  element_subgrid const &my_subgrid_;
-  int const workspace_size_MB_;
-};
-
-template<typename precision>
-class DenseMatrix
-{
-public:
-  DenseMatrix(fk::matrix<precision> const &A) : A_{A} {}
-  // template<typename xmem, typename ymem, typename resrc>
-  void operator()(fk::vector<precision> const &x, fk::vector<precision> &y,
-                  precision const alpha = 1.0, precision const beta = 0.0)
-  {
-    bool const trans_A = false;
-    fm::gemv(A_, x, y, trans_A, alpha, beta);
-  }
-
-private:
-  fk::matrix<precision> &A_;
-};
-
-// simple, node-local test version
-template<typename P, typename MatrixReplacement>
-P simple_gmres(MatrixReplacement asdf, fk::vector<P> &x, fk::vector<P> const &b,
-               fk::matrix<P> const &M, int const restart, int const max_iter,
-               P const tolerance)
-{
-  expect(tolerance >= std::numeric_limits<P>::epsilon());
-  int const n = b.size();
-  expect(b.size() == x.size());
-
-  bool const do_precond = M.size() > 0;
-  std::vector<int> precond_pivots(n);
-  if (do_precond)
-  {
-    expect(M.ncols() == n);
-    expect(M.nrows() == n);
-  }
-  fk::matrix<P> precond(M);
-  bool precond_factored = false;
-
-  expect(restart > 0);
-  expect(restart <= n);
-  expect(max_iter >= restart);
-  expect(max_iter <= n);
-
-  P const norm_b = [&b]() {
-    P const norm = fm::nrm2(b);
-    return (norm == 0.0) ? static_cast<P>(1.0) : norm;
-  }();
-
-  fk::vector<P> residual(b);
-  auto const compute_residual = [&]() {
-    bool const trans_A = false;
-    P const alpha      = -1.0;
-    P const beta       = 1.0;
-    residual           = b;
-    asdf(x, residual, alpha, beta);
-    if (do_precond)
-    {
-      precond_factored ? fm::getrs(precond, residual, precond_pivots)
-                       : fm::gesv(precond, residual, precond_pivots);
-      precond_factored = true;
-    }
-    return fm::nrm2(residual);
-  };
-
-  auto const done = [](P const error, int const outer_iters,
-                       int const inner_iters) {
[…]
