+++ src/program_options.hpp
@@ -48,7 +48,8 @@ enum class PDE_opts
   fokkerplanck_2d_complete_case4,
   diffusion_1,
   diffusion_2,
-  vlasov_lb_full_f
+  vlasov_lb_full_f,
+  vlasov_two_stream
   // FIXME will need to add the user supplied PDE choice
 };
 
@@ -148,7 +149,9 @@ static pde_map_t const pde_mapping = {
          PDE_opts::advection_1)},
     {"vlasov", PDE_descriptor("Vlasov lb full f. df/dt == -v*grad_x f + div_v( "
                               "(v-u)f + theta*grad_v f)",
-                              PDE_opts::vlasov_lb_full_f)}};
+                              PDE_opts::vlasov_lb_full_f)},
+    {"two_stream", PDE_descriptor("Vlasov two-stream. df/dt == -v*grad_x f -E*grad_v f",
+                              PDE_opts::vlasov_two_stream)}};
 
 // class to parse command line input
 class parser
