+++ src/pde/pde_two_stream.hpp
@@ -1,0 +1,257 @@
+#pragma once
+#include "pde_base.hpp"
+
+namespace asgard
+{
+// 2D test case using continuity equation, i.e.,
+//
+//  df/dt == -v*\grad_x f + div_v( (v-u)f + theta\grad_v f)
+//
+//  BC is peridoic in x
+//  BC in v is all inflow in advection for v and Neumann for diffusion in v
+
+template<typename P>
+class PDE_vlasov_two_stream : public PDE<P>
+{
+public:
+  PDE_vlasov_two_stream(parser const &cli_input)
+      : PDE<P>(cli_input, num_dims_, num_sources_, num_terms_, dimensions_,
+               terms_, sources_, exact_vector_funcs_, exact_scalar_func_,
+               get_dt_, do_poisson_solve_, has_analytic_soln_, moments_)
+  {
+    param_manager.add_parameter(parameter<P>{"n", n});
+    param_manager.add_parameter(parameter<P>{"u", u});
+    param_manager.add_parameter(parameter<P>{"theta", theta});
+    param_manager.add_parameter(parameter<P>{"E", E});
+    param_manager.add_parameter(parameter<P>{"S", S});
+  }
+
+private:
+  static int constexpr num_dims_           = 2;
+  static int constexpr num_sources_        = 0;
+  static int constexpr num_terms_          = 5;
+  static bool constexpr do_poisson_solve_  = true;
+  static bool constexpr has_analytic_soln_ = false;
+  static int constexpr default_degree      = 3;
+
+  static fk::vector<P>
+  initial_condition_dim_x_0(fk::vector<P> const &x, P const t = 0)
+  {
+    ignore(t);
+    fk::vector<P> fx(x.size());
+    std::transform(x.begin(), x.end(), fx.begin(), [](P const x_v) -> P {
+      return 1.0-0.5*std::cos(0.5*x_v);
+    });
+    return fx;
+  }
+
+  static fk::vector<P>
+  initial_condition_dim_v_0(fk::vector<P> const &x, P const t = 0)
+  {
+    ignore(t);
+
+    P const coefficient = 1.0 / std::sqrt(PI);
+
+    fk::vector<P> fx(x.size());
+    std::transform(x.begin(), x.end(), fx.begin(),
+                   [coefficient](P const x_v) -> P {
+                     return coefficient * std::pow(x_v,2) * std::exp(-std::pow(x_v, 2));
+                   });
+    return fx;
+  }
+
+  static P dV(P const x, P const time)
+  {
+    ignore(x);
+    ignore(time);
+    return 1.0;
+  }
+
+  /* Define the dimension */
+  inline static dimension<P> const dim_0 = dimension<P>(
+      -2.0*PI, 2.0*PI, 4, default_degree,
+      initial_condition_dim_x_0, dV, "x");
+
+  inline static dimension<P> const dim_1 = dimension<P>(
+      -2.0*PI, 2.0*PI, 3, default_degree,
+      initial_condition_dim_v_0, dV, "v");
+
+  inline static std::vector<dimension<P>> const dimensions_ = {dim_0, dim_1};
+
+  /* Define the moments */
+  static fk::vector<P> moment0_f1(fk::vector<P> const &x, P const t = 0)
+  {
+    ignore(t);
+
+    fk::vector<P> f(x.size());
+    std::fill(f.begin(), f.end(), 1.0);
+    return f;
+  }
+
+  static fk::vector<P> moment1_f1(fk::vector<P> const &x, P const t = 0)
+  {
+    ignore(t);
+    return fk::vector<P>(x);
+  }
+
+  static fk::vector<P> moment2_f1(fk::vector<P> const &x, P const t = 0)
+  {
+    ignore(t);
+
+    fk::vector<P> f(x.size());
[…]
