### 2023-01-19 c8f9c1d
Added E dot grad_v f term.
