+++ src/pde/pde_two_stream.hpp
@@ -24,6 +24,7 @@ public:
     param_manager.add_parameter(parameter<P>{"theta", theta});
     param_manager.add_parameter(parameter<P>{"E", E});
     param_manager.add_parameter(parameter<P>{"S", S});
+    param_manager.add_parameter(parameter<P>{"MaxAbsE",MaxAbsE});
   }
 
 private:
@@ -149,6 +150,13 @@ private:
       ignore(x);
       return 0.0;
   }
+  
+  static P MaxAbsE(P const &x, P const t = 0)
+  {
+      ignore(t);
+      ignore(x);
+      return 0.0;
+  }
 
   /* build the terms */
 
@@ -229,11 +237,96 @@ private:
   inline static std::vector<term<P>> const terms_2 = {term_e2x, term_e2v};
 
   // Term 3
-  // E\cdot\grad_v f
+  // Central Part of E\cdot\grad_v f
   //
+    
+  static P E_func(P const x, P const time = 0)
+  {
+      auto param = param_manager.get_parameter("E");
+      expect(param != nullptr);
+      return param->value(x, time);
+  }
+    
+  static P negOne(P const x, P const time = 0)
+  {
+      ignore(x);
+      ignore(time);
+      return -1.0;
+  }
   
+  inline static const partial_term<P> pterm_E_mass_x = partial_term<P>(
+        coefficient_type::mass, E_func, partial_term<P>::null_gfunc,
+        flux_type::central, boundary_condition::periodic,
+        boundary_condition::periodic);
+  
+    inline static term<P> const E_mass_x = term<P>(true, // time-dependent
+                                                 "",    // name
+                                                 {pterm_E_mass_x});
+    
+  inline static const partial_term<P> pterm_div_v =
+        partial_term<P>(coefficient_type::div, negOne, partial_term<P>::null_gfunc,
+                        flux_type::central,
+                        boundary_condition::dirichlet,
+                        boundary_condition::dirichlet,
+                        homogeneity::homogeneous,
+                        homogeneity::homogeneous);
+    
+    inline static term<P> const div_v = term<P>(false, // time-dependent
+                                                 "",    // name
+                                                 {pterm_div_v});
+    
+    inline static std::vector<term<P>> const terms_3 = {E_mass_x, div_v};
 
-  inline static term_set<P> const terms_ = {terms_1, terms_2};
+    // Term 4 + 5
+    // Penalty Part of E\cdot\grad_v f
+    //
+    
+    static P MaxAbsE_func(P const x, P const time = 0)
+    {
+        auto param = param_manager.get_parameter("MaxAbsE");
+        expect(param != nullptr);
+        return param->value(x, time);
+    }
+    
+    static P posOne(P const x, P const time = 0)
+    {
+        ignore(x);
+        ignore(time);
+        return 1.0;
+    }
+    
+    inline static const partial_term<P> pterm_MaxAbsE_mass_x = partial_term<P>(
+          coefficient_type::mass, MaxAbsE_func, partial_term<P>::null_gfunc,
+          flux_type::central, boundary_condition::periodic,
+          boundary_condition::periodic);
+    
+      inline static term<P> const MaxAbsE_mass_x_1 = term<P>(true, // time-dependent
+                                                   "",    // name
+                                                   {pterm_MaxAbsE_mass_x});
+    
+    inline static term<P> const MaxAbsE_mass_x_2 = term<P>(true, // time-dependent
+                                                 "",    // name
+                                                 {pterm_MaxAbsE_mass_x});
+    
+    inline static const partial_term<P> pterm_div_v_downwind =
+          partial_term<P>(coefficient_type::div, posOne, partial_term<P>::null_gfunc,
+                          flux_type::downwind,
+                          boundary_condition::dirichlet,
[…]
