### 2023-01-19 9a5445f
add tests
Signed-off-by: Steven Hahn <hahnse@ornl.gov>
