### 2023-01-19 d43af7e
rename variables
Signed-off-by: Steven Hahn <hahnse@ornl.gov>
