+++ src/solver_tests.cpp
@@ -39,9 +39,6 @@ void test_kronmult(parser const &parse, int const workspace_size_MB,
     auto const system_size = elem_size * table.size();
     fk::matrix<P> A(system_size, system_size);
     fk::vector<P> x(b);
-    int const restart  = A.ncols();
-    int const max_iter = A.ncols();
-    P const tolerance  = std::is_same_v<float, P> ? 1e-6 : 1e-12;
     build_system_matrix(*pde, table, A, my_subgrid);
     std::vector<int> ipiv(A.nrows());
     fm::gesv(A, x, ipiv);
