+++ src/pde.hpp
@@ -27,8 +27,8 @@
 #include "pde/pde_fokkerplanck1_pitch_C.hpp"
 #include "pde/pde_fokkerplanck1_pitch_E.hpp"
 #include "pde/pde_fokkerplanck2_complete.hpp"
-#include "pde/pde_vlasov_lb_full_f.hpp"
 #include "pde/pde_two_stream.hpp"
+#include "pde/pde_vlasov_lb_full_f.hpp"
 #include "tensors.hpp"
 
 namespace asgard
