+++ src/pde/pde_two_stream.hpp
@@ -24,7 +24,7 @@ public:
     param_manager.add_parameter(parameter<P>{"theta", theta});
     param_manager.add_parameter(parameter<P>{"E", E});
     param_manager.add_parameter(parameter<P>{"S", S});
-    param_manager.add_parameter(parameter<P>{"MaxAbsE",MaxAbsE});
+    param_manager.add_parameter(parameter<P>{"MaxAbsE", MaxAbsE});
   }
 
 private:
@@ -41,7 +41,7 @@ private:
     ignore(t);
     fk::vector<P> fx(x.size());
     std::transform(x.begin(), x.end(), fx.begin(), [](P const x_v) -> P {
-      return 1.0-0.5*std::cos(0.5*x_v);
+      return 1.0 - 0.5 * std::cos(0.5 * x_v);
     });
     return fx;
   }
@@ -54,10 +54,10 @@ private:
     P const coefficient = 1.0 / std::sqrt(PI);
 
     fk::vector<P> fx(x.size());
-    std::transform(x.begin(), x.end(), fx.begin(),
-                   [coefficient](P const x_v) -> P {
-                     return coefficient * std::pow(x_v,2) * std::exp(-std::pow(x_v, 2));
-                   });
+    std::transform(
+        x.begin(), x.end(), fx.begin(), [coefficient](P const x_v) -> P {
+          return coefficient * std::pow(x_v, 2) * std::exp(-std::pow(x_v, 2));
+        });
     return fx;
   }
 
@@ -69,13 +69,13 @@ private:
   }
 
   /* Define the dimension */
-  inline static dimension<P> const dim_0 = dimension<P>(
-      -2.0*PI, 2.0*PI, 4, default_degree,
-      initial_condition_dim_x_0, dV, "x");
+  inline static dimension<P> const dim_0 =
+      dimension<P>(-2.0 * PI, 2.0 * PI, 4, default_degree,
+                   initial_condition_dim_x_0, dV, "x");
 
-  inline static dimension<P> const dim_1 = dimension<P>(
-      -2.0*PI, 2.0*PI, 3, default_degree,
-      initial_condition_dim_v_0, dV, "v");
+  inline static dimension<P> const dim_1 =
+      dimension<P>(-2.0 * PI, 2.0 * PI, 3, default_degree,
+                   initial_condition_dim_v_0, dV, "v");
 
   inline static std::vector<dimension<P>> const dimensions_ = {dim_0, dim_1};
 
@@ -120,7 +120,7 @@ private:
   {
     ignore(t);
 
-    return (1-0.5*std::cos(0.5*x))*0.5;
+    return (1 - 0.5 * std::cos(0.5 * x)) * 0.5;
   }
 
   static P u(P const &x, P const t = 0)
@@ -139,23 +139,23 @@ private:
 
   static P E(P const &x, P const t = 0)
   {
-      ignore(t);
-      ignore(x);
-      return 0.0;
+    ignore(t);
+    ignore(x);
+    return 0.0;
   }
 
-  static P S(P const &x, P const t = 0)
+  static P S(P const &y, P const t = 0)
   {
-      ignore(t);
-      ignore(x);
-      return 0.0;
+    ignore(t);
+    // subtracts quadrature values by one
+    return y - 1.0;
   }
-  
+
   static P MaxAbsE(P const &x, P const t = 0)
   {
-      ignore(t);
-      ignore(x);
-      return 0.0;
+    ignore(t);
+    ignore(x);
+    return 0.0;
   }
 
   /* build the terms */
@@ -239,94 +239,94 @@ private:
   // Term 3
   // Central Part of E\cdot\grad_v f
   //
-    
+
   static P E_func(P const x, P const time = 0)
   {
-      auto param = param_manager.get_parameter("E");
[…]
