### 2023-01-20 c8f726b
use lambda functions
Signed-off-by: Steven Hahn <hahnse@ornl.gov>
