### 2023-01-23 7bceb0c
Add tridiagonal solver wrappers to fast_math
