### 2023-01-23 28ee223
Move workspace_size_MB to options class
Signed-off-by: Steven Hahn <hahnse@ornl.gov>
