+++ src/transformations_tests.cpp
@@ -178,9 +178,6 @@ void test_wavelet_to_realspace(PDE<P> const &pde,
                                std::string const &gold_filename,
                                P const tol_factor)
 {
-  // memory limit for routines
-  static auto constexpr limit_MB = 4000;
-
   // FIXME assume uniform level and degree
   auto const &d     = pde.get_dimensions()[0];
   auto const level  = d.get_level();
@@ -216,8 +213,8 @@ void test_wavelet_to_realspace(PDE<P> const &pde,
       fk::vector<P, mem_type::view, resource::host>(workspace_0),
       fk::vector<P, mem_type::view, resource::host>(workspace_1)};
 
-  wavelet_to_realspace<P>(pde, wave_space, table, transformer, limit_MB,
-                          tmp_workspace, real_space);
+  wavelet_to_realspace<P>(pde, wave_space, table, transformer, tmp_workspace,
+                          real_space);
 
   auto const gold = fk::vector<P>(read_vector_from_txt_file(gold_filename));
 
