+++ src/solver_tests.cpp
@@ -16,9 +16,19 @@ struct distribution_test_init
 static distribution_test_init const distrib_test_info;
 #endif
 
+parser get_parser(PDE_opts const pde_choice,
+                  fk::vector<int> const &starting_levels, int const degree,
+                  int const memory_limit)
+{
+  return parser(pde_choice, starting_levels, degree, parser::DEFAULT_CFL,
+                parser::DEFAULT_USE_FG, parser::DEFAULT_MAX_LEVEL,
+                parser::DEFAULT_TIME_STEPS, parser::DEFAULT_USE_IMPLICIT,
+                parser::DEFAULT_DO_ADAPT, parser::DEFAULT_ADAPT_THRESH,
+                parser::DEFAULT_SOLVER_STR, parser::DEFAULT_USE_IMEX,
+                memory_limit);
+}
 template<typename P>
-void test_kronmult(parser const &parse, int const workspace_size_MB,
-                   P const tol_factor)
+void test_kronmult(parser const &parse, P const tol_factor)
 {
   auto pde = make_PDE<P>(parse);
   options const opts(parse);
@@ -84,15 +94,14 @@ void test_kronmult(parser const &parse, int const workspace_size_MB,
 
   // perform gmres with kron product
   fk::vector<P> const gmres_matrix_free = [&pde, &table, &my_subgrid, &gold, &b,
-                                           elem_size, &opts,
-                                           workspace_size_MB]() {
+                                           elem_size, &opts]() {
     auto const system_size = elem_size * table.size();
     fk::vector<P> x(gold);
     int const restart  = system_size;
     int const max_iter = system_size;
     P const tolerance  = std::is_same_v<float, P> ? 1e-6 : 1e-12;
-    solver::simple_gmres(*pde, table, opts, my_subgrid, workspace_size_MB, x, b,
-                         fk::matrix<P>(), restart, max_iter, tolerance);
+    solver::simple_gmres(*pde, table, opts, my_subgrid, x, b, fk::matrix<P>(),
+                         restart, max_iter, tolerance);
     return x;
   }();
 
@@ -180,51 +189,56 @@ TEMPLATE_TEST_CASE("test kronmult", "[kronmult]", float, double)
 
   SECTION("1d")
   {
-    auto const pde_choice = PDE_opts::continuity_1;
-    auto const degree     = 2;
-    auto const levels     = fk::vector<int>{3};
-    parser const test_parse(pde_choice, levels, degree);
+    auto const pde_choice        = PDE_opts::continuity_1;
+    auto const degree            = 2;
+    auto const levels            = fk::vector<int>{3};
     auto const workspace_size_MB = 1000;
-    test_kronmult(test_parse, workspace_size_MB, tol_factor);
+    parser const test_parse =
+        get_parser(pde_choice, levels, degree, workspace_size_MB);
+    test_kronmult(test_parse, tol_factor);
   }
 
   SECTION("2d - uniform level")
   {
-    auto const pde_choice = PDE_opts::continuity_2;
-    auto const degree     = 3;
-    auto const levels     = fk::vector<int>{2, 2};
-    parser const test_parse(pde_choice, levels, degree);
+    auto const pde_choice        = PDE_opts::continuity_2;
+    auto const degree            = 3;
+    auto const levels            = fk::vector<int>{2, 2};
     auto const workspace_size_MB = 1000;
-    test_kronmult(test_parse, workspace_size_MB, tol_factor);
+    parser const test_parse =
+        get_parser(pde_choice, levels, degree, workspace_size_MB);
+    test_kronmult(test_parse, tol_factor);
   }
   SECTION("2d - non-uniform level")
   {
[…]
