### 2023-01-23 13efd27
Add simple poisson test case
