+++ src/solver.hpp
@@ -20,7 +20,6 @@ P simple_gmres(PDE<P> const &pde, elements::table const &elem_table,
                int const restart, int const max_iter, P const tolerance,
                imex_flag const imex = imex_flag::unspecified);
 
-
 template<typename P>
 void setup_poisson(const int N_nodes, P const x_min, P const x_max,
                    fk::vector<P> &diag, fk::vector<P> &off_diag);
