+++ testing/sandbox.cpp
@@ -27,218 +27,10 @@ using prec = double;
 using prec = float;
 #endif
 
-#include <Accelerate/Accelerate.h>
-
-using namespace asgard;
-
-void poisson_solver( fk::vector<prec> const &source, fk::vector<prec> &phi, fk::vector<prec> &E, int degree, int N_elements, double x_min, double x_max, double phi_min, double phi_max )
-{
-    
-    // Solving: - phi_xx = source Using Linear Finite Elements
-    // Boundary Cconditions: phi(x_min)=phi_min and phi(x_max)=phi_max
-    // Returns phi and E = - Phi_x in Gauss-Legendre Nodes
-    
-    double dx = ( x_max - x_min ) / (double) N_elements;
-    
-    auto const lgwt = legendre_weights( degree+1, -1.0, 1.0, true );
-    
-    int N_nodes = N_elements - 1;
-    
-    // Set the Source Vector //
-    
-    fk :: vector<prec> b(N_nodes);
-    
-    for ( int i = 0; i < N_nodes; i++ )
-    {
-        b[i] = 0.0;
-        for ( int q = 0; q < degree+1; q++ )
-        {
-            b[i] += 0.25 * dx * lgwt[1][q]
-                    * (   source[(i  )*(degree+1)+q] * ( 1.0 + lgwt[0][q] )
-                        + source[(i+1)*(degree+1)+q] * ( 1.0 - lgwt[0][q] ) );
-        }
-    }
-        
-    // Set the Matrix //
-    
-    // Diagonal Entries //
-    
-    fk :: vector<prec> A_D(N_nodes);
-    
-    for ( int i = 0; i < N_nodes; i++ )
-    {
-        A_D[i] = 2.0 / dx;
-    }
-    
-    // Off-Diagonal Entries //
-    
-    fk :: vector<prec> A_E(N_nodes-1);
-    
-    for ( int i = 0; i < N_nodes-1; i++ )
-    {
-        A_E[i] = - 1.0 / dx;
-    }
-    
-    // Matrix Factorization //
-    
-    int INFO = 0;
-    
-    dpttrf_( &N_nodes, A_D.data(), A_E.data(), &INFO );
-    
-    // Linear Solve //
-    
-    int NRHS = 1;
-    
-    dpttrs_( &N_nodes, &NRHS, A_D.data(), A_E.data(), b.data(), &N_nodes, &INFO );
-    
-    // Set Potential and Electric Field in DG Nodes //
-    
-    double dg = ( phi_max - phi_min ) / ( x_max - x_min );
-    
-    // First Element //
-    
-    for ( int k = 0; k < degree+1; k++ )
-    {
-        
-        double x_k = x_min + 0.5 * dx * ( 1.0 + lgwt[0][k] );
-        double g_k = phi_min + dg * ( x_k - x_min );
-        
-        phi[k] = 0.5 * b[0] * ( 1.0 + lgwt[0][k] ) + g_k;
-        
-        E[k] = - b[0] / dx - dg;
-        
-    }
-    
-    // Interior Elements //
-    
-    for ( int i = 1; i < N_elements-1; i++ )
-    {
-        for ( int q = 0; q < degree+1; q++ )
-        {
-            
-            int k      = i * (degree+1) + q;
-            double x_k = ( x_min + i * dx ) + 0.5 * dx * ( 1.0 + lgwt[0][q] );
-            double g_k = phi_min + dg * ( x_k - x_min );
-            
-            phi[k] = 0.5 * (   b[i-1] * ( 1.0 - lgwt[0][q] )
-                             + b[i]   * ( 1.0 + lgwt[0][q] ) ) + g_k;
-            
[…]
