+++ src/program_options.cpp
@@ -65,8 +65,8 @@ parser::parser(int argc, char const *const *argv)
       clara::detail::Opt(use_imex_stepping)["-x"]["--imex"](
           "Use IMEX (implicit-explicit) time advance (vs. explicit or "
           "implicit)") |
-      clara::detail::Opt(workspace_MB, "size > 0")["-m"]["--memory"](
-          "Default workspace size in MB that will be resident on an "
+      clara::detail::Opt(memory_limit, "size > 0")["-m"]["--memory"](
+          "Maximum workspace size in MB that will be resident on an "
           "accelerator");
 
   auto result = cli.parse(clara::detail::Args(argc, argv));
@@ -140,9 +140,10 @@ parser::parser(int argc, char const *const *argv)
         valid = false;
       }
     }
-    if (workspace_MB <= 0)
+    if (memory_limit <= 0)
     {
-      std::cerr << "Kronmult workspace size must be a positive integer\n";
+      std::cerr << "Kronmult max memory size must be a positive integer\n";
+      valid = false;
     }
   }
 
@@ -316,7 +317,7 @@ fk::vector<int> parser::get_active_terms() const { return active_terms; }
 int parser::get_degree() const { return degree; }
 int parser::get_max_level() const { return max_level; }
 int parser::get_time_steps() const { return num_time_steps; }
-int parser::get_workspace_size() const { return workspace_MB; }
+int parser::get_memory_limit() const { return memory_limit; }
 int parser::get_wavelet_output_freq() const { return wavelet_output_freq; }
 int parser::get_realspace_output_freq() const { return realspace_output_freq; }
 
