### 2023-01-23 65bebca
reduce likelihood of integer overflow in fk::matrix::size()
Signed-off-by: Steven Hahn <hahnse@ornl.gov>
