### 2023-01-24 293ffd3
Avoid call to build_system_matrix
Signed-off-by: Steven Hahn <hahnse@ornl.gov>
