### 2023-01-25 c596d83
Add flag to PDE to exclude collision operator
