### 2023-01-25 8803bf0
Strengthen checks on value of inner and outer iterations
Signed-off-by: Steven Hahn <hahnse@ornl.gov>
