### 2023-01-25 589ebf1
Modify matlab interface to plot electric field
