### 2023-01-25 adab282
Set gmres variables at runtime (#487)
* Set gmres variables at runtime
* Strengthen checks on value of inner and outer iterations
