+++ src/program_options.hpp
@@ -175,49 +175,59 @@ public:
   static auto constexpr DEFAULT_SOLVER_STR        = std::string_view("direct");
   static auto constexpr DEFAULT_PDE_SELECTED_CASE = PDE_case_opts::case0;
   static auto constexpr DEFAULT_MEMORY_LIMIT_MB   = 10000;
+  static auto constexpr DEFAULT_GMRES_TOLERANCE   = NO_USER_VALUE_FP;
+  static auto constexpr DEFAULT_GMRES_INNER_ITERATIONS = NO_USER_VALUE;
+  static auto constexpr DEFAULT_GMRES_OUTER_ITERATIONS = NO_USER_VALUE;
 
   // construct from command line
   explicit parser(int argc, char const *const *argv);
 
   // construct from provided values - to simplify testing
-  explicit parser(PDE_opts const pde_choice_in,
-                  fk::vector<int> const &starting_levels_in,
-                  int const degree_in                  = NO_USER_VALUE,
-                  double const cfl_in                  = DEFAULT_CFL,
-                  bool const use_full_grid_in          = DEFAULT_USE_FG,
-                  int const max_level_in               = DEFAULT_MAX_LEVEL,
-                  int const num_steps                  = DEFAULT_TIME_STEPS,
-                  bool const use_implicit              = DEFAULT_USE_IMPLICIT,
-                  bool const do_adapt_levels           = DEFAULT_DO_ADAPT,
-                  double const adapt_threshold_in      = DEFAULT_ADAPT_THRESH,
-                  std::string_view const solver_str_in = DEFAULT_SOLVER_STR,
-                  bool const use_imex                  = DEFAULT_USE_IMEX,
-                  int const memory_limit_in = DEFAULT_MEMORY_LIMIT_MB)
+  explicit parser(
+      PDE_opts const pde_choice_in, fk::vector<int> const &starting_levels_in,
+      int const degree_in = NO_USER_VALUE, double const cfl_in = DEFAULT_CFL,
+      bool const use_full_grid_in          = DEFAULT_USE_FG,
+      int const max_level_in               = DEFAULT_MAX_LEVEL,
+      int const num_steps                  = DEFAULT_TIME_STEPS,
+      bool const use_implicit              = DEFAULT_USE_IMPLICIT,
+      bool const do_adapt_levels           = DEFAULT_DO_ADAPT,
+      double const adapt_threshold_in      = DEFAULT_ADAPT_THRESH,
+      std::string_view const solver_str_in = DEFAULT_SOLVER_STR,
+      bool const use_imex                  = DEFAULT_USE_IMEX,
+      int const memory_limit_in            = DEFAULT_MEMORY_LIMIT_MB,
+      double const gmres_tolerance_in      = DEFAULT_GMRES_TOLERANCE,
+      int const gmres_inner_iterations_in  = DEFAULT_GMRES_INNER_ITERATIONS,
+      int const gmres_outer_iterations_in  = DEFAULT_GMRES_OUTER_ITERATIONS)
       : use_implicit_stepping(use_implicit), use_full_grid(use_full_grid_in),
         do_adapt(do_adapt_levels), starting_levels(starting_levels_in),
         degree(degree_in), max_level(max_level_in), num_time_steps(num_steps),
         cfl(cfl_in), adapt_threshold(adapt_threshold_in),
         pde_choice(pde_choice_in), solver_str(solver_str_in),
         solver(solver_mapping.at(solver_str_in)), use_imex_stepping(use_imex),
-        memory_limit(memory_limit_in){};
+        memory_limit(memory_limit_in), gmres_tolerance(gmres_tolerance_in),
+        gmres_inner_iterations(gmres_inner_iterations_in),
[…]
