### 2023-01-26 87bb2e3
Update poisson with periodic boundary conditions
