### 2023-01-31 bcc97f0
Add multi function support for analytical solutions
