+++ src/pde/pde_continuity2.hpp
@@ -254,8 +254,8 @@ private:
                                                          source2_};
 
   // define exact soln
-  inline static std::vector<vector_func<P>> const exact_vector_funcs_ = {
-      exact_solution_dim0, exact_solution_dim1};
+  inline static std::vector<md_func_type<P>> const exact_vector_funcs_ = {
+      {exact_solution_dim0, exact_solution_dim1}};
 
   inline static scalar_func<P> const exact_scalar_func_ = exact_time;
 };
