+++ src/pde/pde_continuity1.hpp
@@ -183,8 +183,8 @@ private:
   inline static std::vector<source<P>> const sources_ = {source0_, source1_};
 
   // define exact soln functions
-  inline static std::vector<vector_func<P>> const exact_vector_funcs_ = {
-      exact_solution_dim0};
+  inline static std::vector<md_func_type<P>> const exact_vector_funcs_ = {
+      {exact_solution_dim0}};
 
   inline static scalar_func<P> const exact_scalar_func_ = exact_time;
 };
