### 2023-01-31 5d6f934
clear_and_resize overflows with large numbers of rows and cols (#490)
Signed-off-by: Steven Hahn <hahnse@ornl.gov>
