### 2023-02-01 9dabb9a
Add alternate IO with element metadata and script to plot
