### 2023-02-01 a621bb5
Set better default restart value
Signed-off-by: Steven Hahn <hahnse@ornl.gov>
