### 2023-02-02 e8327e3
Set better default restart value (#491)
* Set better default restart value
* Changes discussed in code review

signed-off-by: Steven Hahn <hahnse@ornl.gov>
