### 2023-02-06 fa18952
Added penalty term to PDE operator.  Still needs to be added to BCs
