+++ src/basis.hpp
@@ -34,12 +34,17 @@ template<typename P, resource resrc>
 class wavelet_transform
 {
 public:
-  wavelet_transform(options const &program_opts, int const max_degree,
+  wavelet_transform(int const max_level_in, int const max_degree,
                     bool const quiet = true);
+
+  wavelet_transform(options const &program_opts, int const max_degree,
+                    bool const quiet = true)
+      : wavelet_transform(program_opts.max_level, max_degree,
+                    quiet){}
 
   wavelet_transform(options const &program_opts, PDE<P> const &pde,
                     bool const quiet = true)
-      : wavelet_transform(program_opts, pde.get_dimensions()[0].get_degree(),
+      : wavelet_transform(program_opts.max_level, pde.get_dimensions()[0].get_degree(),
                           quiet)
   {}
 
