+++ src/coefficients.cpp
@@ -106,10 +106,10 @@ fk::matrix<P> generate_coefficients(
   expect(level <= transformer.max_level);
 
   // setup jacobi of variable x and define coeff_mat
-  auto const num_points = fm::two_raised_to(level);
+  auto const num_cells = fm::two_raised_to(level);
 
-  auto const grid_spacing = (dim.domain_max - dim.domain_min) / num_points;
-  auto const degrees_freedom_1d = dim.get_degree() * num_points;
+  auto const grid_spacing = (dim.domain_max - dim.domain_min) / num_cells;
+  auto const degrees_freedom_1d = dim.get_degree() * num_cells;
   fk::matrix<P> coefficients(degrees_freedom_1d, degrees_freedom_1d);
 
   // get quadrature points and quadrature_weights.
@@ -149,7 +149,7 @@ fk::matrix<P> generate_coefficients(
   // get jacobian
   auto const jacobi = grid_spacing / 2;
 
-  for (auto i = 0; i < num_points; ++i)
+  for (auto i = 0; i < num_cells; ++i)
   {
     // get left and right locations for this element
     auto const x_left  = dim.domain_min + i * grid_spacing;
@@ -158,7 +158,7 @@ fk::matrix<P> generate_coefficients(
     // get index for current, first and last element
     auto const current = dim.get_degree() * i;
     auto const first   = 0;
-    auto const last    = dim.get_degree() * (num_points - 1);
+    auto const last    = dim.get_degree() * (num_cells - 1);
 
     // map quadrature points from [-1,1] to physical domain of this i element
     fk::vector<P> const quadrature_points_i = [&, quadrature_points =
@@ -205,6 +205,10 @@ fk::matrix<P> generate_coefficients(
       {
         output = legendre_prime_t * tmp * (-1);
       }
+      else if (pterm.coeff_type == coefficient_type::penalty)
+      {
+        // output is zeros (no volume term)
+      }
       return output;
     }();
 
@@ -217,7 +221,8 @@ fk::matrix<P> generate_coefficients(
     coefficients.set_submatrix(current, current, curr_block);
 
     if (pterm.coeff_type == coefficient_type::grad ||
-        pterm.coeff_type == coefficient_type::div)
+        pterm.coeff_type == coefficient_type::div  ||
+        pterm.coeff_type == coefficient_type::penalty )
     {
       // setup numerical flux choice/boundary conditions
       //
@@ -232,27 +237,64 @@ fk::matrix<P> generate_coefficients(
       // the dat is going to be used in the G function (above it is used as
       // linear multuplication but this is not always true)
 
-      auto const flux_left =
+      // Penalty term is just <|gfunc|/2[[f]],[[v]]> so we need to remove the central
+      // flux <gfunc{f},[[v]]> from the operators
+      P central_coeff = 1.0;
+      if ( pterm.coeff_type == coefficient_type::penalty ){
+        central_coeff = 0.0;
+      }
+
+      P const flux_left =
           pterm.g_func(x_left, time) * pterm.dv_func(x_left, time);
-      auto const flux_right =
+      P const flux_right =
           pterm.g_func(x_right, time) * pterm.dv_func(x_right, time);
 
       // get the "trace" values
-      // (values at the left and right of each element for all k)
+      // (values at the left and right of each element for all k).
+      // -------------------------------------------------------------------------
+      // More detailed explanation 
+      // Each trace_value_ evaluates <FLUX_f,[[v]]> 
+      // where v is a DG functions with support on I_i. The 
+      // difference between the trace_values_ varies with the edge the flux
[…]
