+++ src/pde.hpp
@@ -15,6 +15,7 @@
 #include "pde/pde_base.hpp"
 
 #include "pde/pde_advection1.hpp"
+#include "pde/pde_collisional_landau.hpp"
 #include "pde/pde_continuity1.hpp"
 #include "pde/pde_continuity2.hpp"
 #include "pde/pde_continuity3.hpp"
@@ -96,6 +97,8 @@ std::unique_ptr<PDE<P>> make_PDE(parser const &cli_input)
     return std::make_unique<PDE_vlasov_lb<P>>(cli_input);
   case PDE_opts::vlasov_two_stream:
     return std::make_unique<PDE_vlasov_two_stream<P>>(cli_input);
+  case PDE_opts::collisional_landau:
+    return std::make_unique<PDE_collisional_landau<P>>(cli_input);
   default:
     std::cout << "Invalid pde choice" << std::endl;
     exit(-1);
@@ -184,6 +187,9 @@ make_PDE(PDE_opts const pde_choice, int const level = parser::NO_USER_VALUE,
       return fk::vector<int>(std::vector<int>(2, level));
 
     case PDE_opts::vlasov_two_stream:
+      return fk::vector<int>(std::vector<int>(2, level));
+
+    case PDE_opts::collisional_landau:
       return fk::vector<int>(std::vector<int>(2, level));
 
     default:
