### 2023-02-06 592f755
Combine moment 3d and 4d functions into one
