### 2023-02-06 bb295a5
Use imex flag in euler operator kronmult
