### 2023-02-07 f886321
definitions needed by openblas
Signed-off-by: Steven Hahn <hahnse@ornl.gov>
