### 2023-02-10 f87f2af
Add location of 1D slices to figure
Signed-off-by: Steven Hahn <hahnse@ornl.gov>
