### 2023-02-12 ec45a56
Add wiki examples to README
Signed-off-by: Steven Hahn <hahnse@ornl.gov>
