### 2023-02-13 324355b
Add location of 1D slices to figure (#499)
Signed-off-by: Steven Hahn <hahnse@ornl.gov>
