### 2023-02-14 0ff51f3
Update src/coefficients.cpp
Done

Co-authored-by: Steven Hahn <hahnse@ornl.gov>
