+++ src/coefficients.cpp
@@ -108,7 +108,7 @@ fk::matrix<P> generate_coefficients(
   // setup jacobi of variable x and define coeff_mat
   auto const num_cells = fm::two_raised_to(level);
 
-  auto const grid_spacing = (dim.domain_max - dim.domain_min) / num_cells;
+  auto const grid_spacing       = (dim.domain_max - dim.domain_min) / num_cells;
   auto const degrees_freedom_1d = dim.get_degree() * num_cells;
   fk::matrix<P> coefficients(degrees_freedom_1d, degrees_freedom_1d);
 
@@ -205,7 +205,7 @@ fk::matrix<P> generate_coefficients(
       {
         output = legendre_prime_t * tmp * (-1);
       }
-      // If pterm.coeff_type == coefficient_type::penalty is true, there's 
+      // If pterm.coeff_type == coefficient_type::penalty is true, there's
       // no volume term so the output is zeros.
       return output;
     }();
@@ -219,22 +219,21 @@ fk::matrix<P> generate_coefficients(
     coefficients.set_submatrix(current, current, curr_block);
 
     if (pterm.coeff_type == coefficient_type::grad ||
-        pterm.coeff_type == coefficient_type::div  ||
-        pterm.coeff_type == coefficient_type::penalty )
+        pterm.coeff_type == coefficient_type::div ||
+        pterm.coeff_type == coefficient_type::penalty)
     {
       // setup numerical flux choice/boundary conditions
       //
       // - <funcCoef*{q},p>
       //----------------------------------------------
       // Numerical Flux is defined as
-      // Flux = {{{f}}} + C/2*[[u]]
+      // Flux = {{f}} + C/2*[[u]]
       //      = ( f_L + f_R )/2 + FunCoef*( u_R - u_L )/2
       // [[v]] = v_R - v_L
 
       // FIXME G functions should accept G(x,p,t,dat), since we don't know how
       // the dat is going to be used in the G function (above it is used as
       // linear multuplication but this is not always true)
-
 
       // Penalty term is just <|gfunc|/2[[f]],[[v]]> so we need to remove the central
       // flux <gfunc{{f}},[[v]]> from the operators
@@ -248,49 +247,53 @@ fk::matrix<P> generate_coefficients(
       // get the "trace" values
       // (values at the left and right of each element for all k)
       // -------------------------------------------------------------------------
-      // More detailed explanation 
-      // Each trace_value_ evaluates <FLUX_f,[[v]]> 
-      // where v is a DG functions with support on I_i. The 
+      // More detailed explanation
+      // Each trace_value_ evaluates <FLUX_f,[[v]]>
+      // where v is a DG functions with support on I_i. The
       // difference between the trace_values_ varies with the edge the flux
       // is evaluated on and the support of the DG function f.
       // The legendre_poly_X is the trace of f and legende_poly_X_t is for v
       // We will use f=p_X for the polynomials where X=L (left boundary of cell)
-      // or X=R (right boundary of cell).  Similar for v but depends on the support
+      // or X=R (right boundary of cell).  Similar for v but depends on the
+      // support
 
-      // trace_value_1 is the interaction on x_{i-1/2} --  
+      // trace_value_1 is the interaction on x_{i-1/2} --
       // the edge between cell I_{i-1} and I_i or the left boundary of I_i.
       // f is a DG function with support on I_{i-1}
       // In this case:  {{f}} = p_R/2, [[f]] = p_R, [[v]] = -p_L
       auto trace_value_1 =
[…]
