+++ src/basis.cpp
@@ -437,8 +437,7 @@ template<typename P, resource resrc>
 wavelet_transform<P, resrc>::wavelet_transform(int const max_level_in,
                                                int const max_degree,
                                                bool const quiet)
-    : max_level(max_level_in), degree(max_degree),
-      dense_blocks_(max_level * 2)
+    : max_level(max_level_in), degree(max_degree), dense_blocks_(max_level * 2)
 {
   // this is to get around unused warnings
   // because can't unpack only some args w structured binding (until
