### 2023-02-17 7e30824
Add Openblas definition checks
