### 2023-02-17 05773af
Use static cast for indices with std::pow
