### 2023-02-17 17102f4
Add overload for single vector analytical solutions
