+++ src/pde/pde_advection1.hpp
@@ -72,8 +72,8 @@ private:
     return 1.0;
   }
 
-  inline static std::vector<md_func_type<P>> const exact_vector_funcs_ = {
-      {exact_solution_dim0}};
+  inline static std::vector<vector_func<P>> const exact_vector_funcs_ = {
+      exact_solution_dim0};
 
   inline static scalar_func<P> const exact_scalar_func_ = exact_time;
   // =========================================================================
