+++ src/solver.cpp
@@ -30,14 +30,14 @@ P simple_gmres(PDE<P> const &pde, elements::table const &elem_table,
                element_subgrid const &my_subgrid, fk::vector<P> &x,
                fk::vector<P> const &b, fk::matrix<P> const &M,
                int const restart, int const max_iter, P const tolerance,
-               imex_flag const imex)
+               imex_flag const imex, P const alpha_in)
 {
-  auto euler_operator = [&pde, &elem_table, &program_options, &my_subgrid,
-                         imex](fk::vector<P> const &x_in, fk::vector<P> &y,
-                               P const alpha = 1.0, P const beta = 0.0) {
+  auto euler_operator = [&pde, &elem_table, &program_options, &my_subgrid, imex,
+                         alpha_in](fk::vector<P> const &x_in, fk::vector<P> &y,
+                                   P const alpha = 1.0, P const beta = 0.0) {
     auto tmp = kronmult::execute(pde, elem_table, program_options, my_subgrid,
                                  x_in, imex);
-    tmp      = x_in - tmp * pde.get_dt();
+    tmp      = x_in - tmp * pde.get_dt() * alpha_in;
     y        = tmp * alpha + y * beta;
   };
   return simple_gmres(euler_operator, x, b, M, restart, max_iter, tolerance);
@@ -270,6 +270,7 @@ void poisson_solver(fk::vector<P> const &source, fk::vector<P> const &A_D,
                     P const x_min, P const x_max, P const phi_min,
                     P const phi_max, poisson_bc const bc)
 {
+  tools::timer.start("poisson_solver");
   // Solving: - phi_xx = source Using Linear Finite Elements
   // Boundary Conditions: phi(x_min)=phi_min and phi(x_max)=phi_max
   // Returns phi and E = - Phi_x in Gauss-Legendre Nodes
@@ -354,6 +355,7 @@ void poisson_solver(fk::vector<P> const &source, fk::vector<P> const &A_D,
 
     E[k] = b[i - 1] / dx - dg;
   }
+  tools::timer.stop("poisson_solver");
 }
 
 template float simple_gmres(fk::matrix<float> const &A, fk::vector<float> &x,
@@ -371,14 +373,15 @@ simple_gmres(PDE<float> const &pde, elements::table const &elem_table,
              options const &program_options, element_subgrid const &my_subgrid,
              fk::vector<float> &x, fk::vector<float> const &b,
              fk::matrix<float> const &M, int const restart, int const max_iter,
-             float const tolerance, imex_flag const imex);
+             float const tolerance, imex_flag const imex, const float alpha_in);
 
 template double
 simple_gmres(PDE<double> const &pde, elements::table const &elem_table,
              options const &program_options, element_subgrid const &my_subgrid,
              fk::vector<double> &x, fk::vector<double> const &b,
              fk::matrix<double> const &M, int const restart, int const max_iter,
-             double const tolerance, imex_flag const imex);
+             double const tolerance, imex_flag const imex,
+             const double alpha_in);
 
 template void setup_poisson(const int N_elements, float const x_min,
                             float const x_max, fk::vector<float> &diag,
