+++ src/moment.hpp
@@ -6,6 +6,7 @@ template<typename P>
 class PDE;
 }
 
+#include "basis.hpp"
 #include "elements.hpp"
 #include "pde/pde_base.hpp"
 #include "program_options.hpp"
@@ -17,6 +18,12 @@ namespace asgard
 namespace elements
 {
 class table;
+}
+
+namespace basis
+{
+template<typename P, resource resrc>
+class wavelet_transform;
 }
 
 template<typename P>
@@ -39,6 +46,13 @@ public:
   void createMomentReducedMatrix(PDE<P> const &pde,
                                  elements::table const &hash_table);
 
+  fk::vector<P> const &get_realspace_moment() const { return realspace; }
+
+  fk::vector<P> const &create_realspace_moment(
+      PDE<P> const &pde_1d, fk::vector<P> &wave, elements::table const &table,
+      asgard::basis::wavelet_transform<P, resource::host> const &transformer,
+      std::array<fk::vector<P, mem_type::view, resource::host>, 2> &workspace);
+
 private:
   template<int nvdim>
   void createMomentReducedMatrix_nd(PDE<P> const &pde,
@@ -48,6 +62,7 @@ private:
   std::vector<std::vector<fk::vector<P>>> fList;
   fk::vector<P> vector;
   fk::matrix<P> moment_matrix;
+  fk::vector<P> realspace;
   // moment_fval_integral;
   // moment_analytic_integral;
 };

[…]
