### 2023-02-17 2df4d08
Save nodes for each dimension
