### 2023-02-17 3ed1697
Return iteration counts from gmres
