### 2023-02-17 5b2046d
Save gmres error and iters to file
